Transform to multivariate Pareto samples via rank transformation

ParetoTransRank transforms the original data to multivariate Pareto samples with suitable margins (the default: Pareto margins) via rank transformation.

Usage

ParetoTransRank(data, u_quan, scaleType = "U")

Arguments

data

An \(n\times d\) data matrix

u_quan

A numeric quantile, set a lower enough quantile \(u_{quan}\in(0,1)\)

scaleType

A character specifying the type of scales for generated samples (the default is uniform scale: scaleType="U"). Other possible scales:

scaleType="P": Pareto scale

scaleType="E": Exponential scale

Value

An \(N\times d\) pseudo-observations for the inverted Pareto distribution (default)

Details

Let \(\boldsymbol(X)_{i}, i=1,\ldots,n\) be independent samples from the unkown distribution \(F\). Transform to the uniform margin via rank transformation \(\widehat{U}_{i,j}=1-(rank_{i,j}-0.5)/n\), where \(rank_{i,j}=\sum_{s=1}^{n}\mathbb{I}(X_{s,j}\le X_{i,j})\) is the maximal rank of \(X_{i,j}\) among \(X_{1,j},\ldots,X_{n,j}\). The scaled points \((n/k)\widehat{U}_i\) for \(i=1,\ldots,n\) such that \(\min\widehat{U}_i<k/n\) are pseudo-observations from the inverted multivariate Pareto distribution. We set \(k\in\{1,\ldots,n\}\) such that \(k\) is large and \(k/n\) is small. If we take the reciprocal of the rescaled samples, then we have Pareto scales. If we take the negative log of the rescaled samples, then we have exponential scales.

Examples

StrMtx <- matrix(c(1, 1, 2, 2, 4,
             0, 2, 1, 3, 2,
             0, 0, 3, 1, 3,
             0, 0, 0, 4, 1,
             0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                  0, 0, 2, 2.5, 0.7,
                  0, 0, 0, 0.4, -0.3,
                  0, 0, 0, 0, 0.1,
                  0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                0, 0, 3, 4, 1,
                0, 0, 0, 3, 1,
                0, 0, 0, 0, 1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
## X-Vine specification (Vine structure, Bivariate parametric families, Parameters)
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Multivariate Pareto samples
Dat_P=ParetoSim(n = 5000, XVS = XVS) # Pareto scale
##  Transform to pseudo-observations for inverted Pareto distribution
Dat_U=ParetoTransRank(data = Dat_P, u_quan = 0.2, scaleType = "U")
Dat_U
#>           [,1]   [,2]   [,3]   [,4]   [,5]
#>    [1,] 0.4895 0.2225 0.3675 0.2935 0.0985
#>    [2,] 0.5155 1.8355 0.9135 0.9135 2.0145
#>    [3,] 3.8575 0.6205 3.5315 2.4755 2.0085
#>    [4,] 2.9565 1.7305 2.2785 1.1235 0.5725
#>    [5,] 3.9915 1.9955 3.3325 2.0715 0.5625
#>    [6,] 1.6495 1.0925 0.9445 1.6025 3.5435
#>    [7,] 0.0335 0.0185 0.0465 0.0235 0.0125
#>    [8,] 2.0275 2.4955 2.4455 1.5975 0.8895
#>    [9,] 0.6625 3.1265 1.1955 1.8845 3.4205
#>   [10,] 1.0145 0.4575 0.6405 0.5185 1.0695
#>   [11,] 1.1015 0.5445 1.8645 2.3525 1.2405
#>   [12,] 0.5565 1.0695 0.7425 0.3965 0.3015
#>   [13,] 2.3255 4.0865 3.7985 3.9285 0.9015
#>   [14,] 2.1925 2.5455 3.1705 0.8905 1.0305
#>   [15,] 0.1405 4.2235 0.2615 0.7685 4.7895
#>   [16,] 0.7275 4.7465 2.3295 3.6545 4.7635
#>   [17,] 2.8835 0.3445 0.5305 0.4755 0.4385
#>   [18,] 2.1795 1.6855 1.7315 1.7145 0.8195
#>   [19,] 1.3215 0.4965 1.2775 1.2705 0.4005
#>   [20,] 3.3115 0.5495 0.8195 1.4955 1.3885
#>   [21,] 1.7065 0.9145 3.2585 2.3095 0.2275
#>   [22,] 0.5865 0.1525 0.7805 0.3865 0.0935
#>   [23,] 1.0635 4.3015 0.9805 3.9125 4.8735
#>   [24,] 0.1265 0.1345 0.1555 0.1095 0.1645
#>   [25,] 0.1005 0.4595 0.2405 0.2665 0.7455
#>   [26,] 4.5895 1.9335 1.1005 0.8385 1.5295
#>   [27,] 0.7305 2.4985 0.8675 1.7215 4.2095
#>   [28,] 0.5625 0.8475 0.9425 0.5465 0.5605
#>   [29,] 2.2365 1.4175 1.2415 1.9765 0.8855
#>   [30,] 0.7095 1.3515 0.6315 0.6645 4.0165
#>   [31,] 2.1935 0.8855 1.5535 1.5635 0.9285
#>   [32,] 0.4015 0.7875 0.3755 0.5965 0.7205
#>   [33,] 3.0325 1.8365 1.4165 0.9385 3.6305
#>   [34,] 0.0065 0.0025 0.0075 0.0025 0.0025
#>   [35,] 0.7985 2.8505 1.5505 0.6545 0.5885
#>   [36,] 0.1705 0.3905 0.2445 0.2375 1.2905
#>   [37,] 0.5895 3.0405 1.0945 0.5545 3.0895
#>   [38,] 1.6475 0.3845 0.4165 0.3265 0.5225
#>   [39,] 0.1015 0.2115 0.1065 0.1295 0.4045
#>   [40,] 3.9825 1.1035 2.9295 3.9245 0.8485
#>   [41,] 2.4625 0.9355 2.0925 1.4215 2.9165
#>   [42,] 1.0835 0.7365 1.2905 2.0255 0.6645
#>   [43,] 4.0765 0.8875 2.0335 3.0795 0.9755
#>   [44,] 4.2495 0.8835 2.8365 1.6725 0.5495
#>   [45,] 3.6145 1.8655 1.3405 0.9695 3.4495
#>   [46,] 0.1985 0.0585 0.0945 0.0825 0.0825
#>   [47,] 0.1545 0.0095 0.0455 0.0735 0.0065
#>   [48,] 2.1805 3.0465 1.5495 0.9735 3.0695
#>   [49,] 3.0855 1.3345 1.1945 1.1405 0.9185
#>   [50,] 0.2585 3.5395 0.3945 3.1185 4.8035
#>   [51,] 0.5745 3.5115 1.2985 2.5525 3.3785
#>   [52,] 3.6105 2.8755 4.3945 4.1855 0.7315
#>   [53,] 2.1705 3.9515 2.2655 0.6675 1.3065
#>   [54,] 0.1645 1.0175 0.2975 0.2565 1.3225
#>   [55,] 0.6805 3.2145 2.4335 2.2015 1.5335
#>   [56,] 1.1425 0.7235 0.6815 0.4805 0.6575
#>   [57,] 0.8805 4.2905 0.7945 1.4805 4.2685
#>   [58,] 0.3695 0.1925 0.4285 0.4945 0.3365
#>   [59,] 2.4255 2.4445 2.6565 0.5005 2.4085
#>   [60,] 0.7935 2.8565 1.8285 0.6575 2.8935
#>   [61,] 0.8065 4.1965 2.1675 0.8915 4.6275
#>   [62,] 3.4645 0.7245 0.9125 1.1835 0.6895
#>   [63,] 0.4585 4.5735 0.8865 4.5735 4.8955
#>   [64,] 1.9285 0.7225 2.0135 2.4545 1.5865
#>   [65,] 3.4845 0.4735 0.3765 0.5075 0.5915
#>   [66,] 0.1215 0.1285 0.0615 0.0765 0.4025
#>   [67,] 1.1225 0.9415 0.9385 1.0535 2.7765
#>   [68,] 1.5675 0.5925 1.6125 1.3045 0.9705
#>   [69,] 0.0855 0.8765 0.1785 0.3215 0.9305
#>   [70,] 0.5765 3.0555 0.6155 1.9515 2.8485
#>   [71,] 0.2835 0.0475 0.0905 0.1005 0.0505
#>   [72,] 4.8165 0.7505 0.8385 0.7625 1.0975
#>   [73,] 3.7225 2.7335 4.7065 4.0735 0.8015
#>   [74,] 3.2425 0.9785 1.9545 2.6905 1.4875
#>   [75,] 0.4205 0.4565 0.7245 0.9885 0.9625
#>   [76,] 2.7565 0.5025 3.0535 3.6275 0.8915
#>   [77,] 0.3175 4.8225 0.7785 4.7315 4.9665
#>   [78,] 1.9665 2.0085 1.3945 0.6885 3.0615
#>   [79,] 0.0935 0.1995 0.1935 0.1335 0.0955
#>   [80,] 3.1795 0.7685 2.4235 2.9305 0.6775
#>   [81,] 0.3115 0.5695 1.2265 0.8215 2.1115
#>   [82,] 3.0655 2.6805 1.3035 1.2375 0.7705
#>   [83,] 0.7355 4.3145 2.7145 0.9225 2.7315
#>   [84,] 1.1595 0.2725 0.6705 1.8675 0.4595
#>   [85,] 0.6195 1.3825 1.7025 1.0835 0.4515
#>   [86,] 0.4065 0.3135 0.4695 0.3975 0.5585
#>   [87,] 0.0775 0.6865 0.1305 2.1495 3.2955
#>   [88,] 4.1605 2.2445 4.8675 4.4905 0.8475
#>   [89,] 1.0905 2.4125 1.3125 0.9785 2.9505
#>   [90,] 1.8205 1.6475 1.0705 0.9045 2.9855
#>   [91,] 0.1385 0.1225 0.1175 0.1585 0.2655
#>   [92,] 2.1825 0.2205 0.3345 0.3055 0.4125
#>   [93,] 2.0555 1.6145 1.7835 0.9985 1.8915
#>   [94,] 3.3365 0.5655 0.9495 1.1215 0.5655
#>   [95,] 4.4965 0.2585 1.2305 0.5375 0.0535
#>   [96,] 0.0035 0.0065 0.0035 0.0045 0.0395
#>   [97,] 0.8725 3.5245 2.4845 1.8005 2.3445
#>   [98,] 0.4645 4.4085 0.8825 1.3345 3.9545
#>   [99,] 0.4085 0.3835 0.4305 0.2115 0.1365
#>  [100,] 0.6955 0.8505 1.4265 0.3925 0.3285
#>  [101,] 0.0805 0.1665 0.1345 0.1145 0.1955
#>  [102,] 3.9945 0.8385 1.6335 1.6515 1.0085
#>  [103,] 0.2105 0.2865 0.1755 0.1555 0.2045
#>  [104,] 0.0835 0.1075 0.1705 0.1445 0.0245
#>  [105,] 0.9775 3.0665 1.7605 1.9175 3.2055
#>  [106,] 0.2285 0.0895 0.1335 0.1265 0.0555
#>  [107,] 0.3335 1.8385 0.5165 0.7335 1.3045
#>  [108,] 1.0355 0.8715 0.8425 0.5955 0.3275
#>  [109,] 2.3345 1.5655 3.2465 2.8925 0.5665
#>  [110,] 0.3315 2.0135 0.5115 0.7925 3.6545
#>  [111,] 3.2325 0.6645 0.5125 1.3625 1.3365
#>  [112,] 0.4125 0.9655 0.4505 0.5535 1.1225
#>  [113,] 0.8495 1.3475 0.7015 0.5305 1.1205
#>  [114,] 0.7455 0.6405 4.5355 2.6935 0.5115
#>  [115,] 0.6135 3.0765 0.8895 0.8435 3.7615
#>  [116,] 0.4345 2.2235 0.8335 0.2425 2.5045
#>  [117,] 0.1245 0.0195 0.0515 0.0615 0.0425
#>  [118,] 0.2325 0.2835 0.3035 0.2435 0.1295
#>  [119,] 1.5645 1.6955 1.7255 1.5905 0.9845
#>  [120,] 0.5085 3.8205 0.7595 3.7315 3.4025
#>  [121,] 3.2725 1.4145 2.2115 0.8235 1.7765
#>  [122,] 0.3585 1.4085 0.4815 1.0245 1.3985
#>  [123,] 2.4535 0.7965 3.0445 3.7885 1.1305
#>  [124,] 0.5705 0.7835 1.1415 0.7655 1.4355
#>  [125,] 0.4905 0.7275 0.6165 0.8335 1.8295
#>  [126,] 0.3425 3.0945 0.5415 1.6715 3.7125
#>  [127,] 0.1325 0.4305 0.1535 0.5635 0.5155
#>  [128,] 1.2745 1.0365 1.1745 0.6385 1.7785
#>  [129,] 0.3145 0.1275 0.0985 0.2055 0.2705
#>  [130,] 0.9915 3.1115 1.6405 1.2065 2.5805
#>  [131,] 0.0535 0.0395 0.0435 0.1305 0.0325
#>  [132,] 0.2725 0.7255 0.6385 1.0425 1.1315
#>  [133,] 0.0355 0.0225 0.0535 0.0785 0.0695
#>  [134,] 4.4625 1.1655 3.3165 3.0405 0.9575
#>  [135,] 0.6415 0.3685 1.9805 4.2965 1.5045
#>  [136,] 3.9035 0.6335 0.9755 1.8165 0.3265
#>  [137,] 0.0285 0.0755 0.0445 0.0555 0.0795
#>  [138,] 0.8275 0.1165 0.6945 1.5205 0.3715
#>  [139,] 0.6015 2.1865 1.1205 0.7165 0.8925
#>  [140,] 0.8405 4.7085 1.8945 3.3755 4.7755
#>  [141,] 1.8325 0.1845 0.2545 0.2225 0.2985
#>  [142,] 1.9975 0.6415 0.8885 1.0045 0.5435
#>  [143,] 2.0955 0.6945 2.4285 2.1185 0.1155
#>  [144,] 1.0105 0.1405 0.3405 0.4045 0.1525
#>  [145,] 3.0295 1.1285 3.1735 1.5805 0.7105
#>  [146,] 1.1275 0.2415 0.3425 0.3515 0.8515
#>  [147,] 0.5305 1.3685 1.0295 2.3325 2.2015
#>  [148,] 3.4955 0.8235 0.7975 0.5485 0.9325
#>  [149,] 4.0755 1.2585 3.6655 3.2385 0.9225
#>  [150,] 0.3845 0.2085 0.3055 0.2395 0.1445
#>  [151,] 0.2685 0.1235 0.1445 0.1725 0.1385
#>  [152,] 2.2495 0.4365 0.7795 0.6455 0.6835
#>  [153,] 1.1195 0.4585 0.9045 0.8995 0.4755
#>  [154,] 1.0365 1.1265 1.2735 0.8195 0.6815
#>  [155,] 0.3325 3.7425 0.4575 1.1825 3.4265
#>  [156,] 4.1445 3.1335 2.8265 0.7595 3.3705
#>  [157,] 1.8315 0.3005 0.2785 0.2545 1.2495
#>  [158,] 0.3635 0.8085 0.9175 0.2795 0.5815
#>  [159,] 0.1785 2.7845 0.5295 0.4265 3.5795
#>  [160,] 0.9175 0.8125 1.0935 0.5655 1.5645
#>  [161,] 4.9605 0.8935 2.2635 2.3885 1.1645
#>  [162,] 0.3525 0.4065 0.8415 0.8755 0.3145
#>  [163,] 0.3925 0.1055 0.2185 0.2525 0.4355
#>  [164,] 0.8975 2.5955 1.2615 3.0315 4.2375
#>  [165,] 0.6705 0.5635 0.5465 0.4745 0.6635
#>  [166,] 0.2965 0.5835 0.3855 0.3185 0.2865
#>  [167,] 3.7815 2.4275 4.4755 3.5575 0.8665
#>  [168,] 4.2315 2.2155 3.5835 1.2325 0.6205
#>  [169,] 2.3415 0.6285 1.0885 0.9515 0.9715
#>  [170,] 3.5765 1.5405 1.8745 0.8875 0.8465
#>  [171,] 0.2245 2.0225 0.2485 0.2155 1.2215
#>  [172,] 1.2605 0.8675 2.0655 2.8695 1.0285
#>  [173,] 0.6795 4.3165 1.9645 1.3575 4.1375
#>  [174,] 1.2635 0.1745 0.2085 0.2725 0.3865
#>  [175,] 0.1755 0.1625 0.1355 0.2035 0.3685
#>  [176,] 1.3685 1.1985 0.9575 1.5375 2.4445
#>  [177,] 0.0825 0.3515 0.0735 0.2675 2.1765
#>  [178,] 0.2935 0.2545 0.1115 0.1975 0.3545
#>  [179,] 1.1455 2.2115 0.8105 0.8605 4.7415
#>  [180,] 0.4095 3.6255 0.6295 0.6965 4.6695
#>  [181,] 4.4445 1.6525 2.8405 1.8365 0.8585
#>  [182,] 1.1055 1.8205 0.9285 1.6205 2.2775
#>  [183,] 4.0305 0.9185 4.3295 4.2015 0.2465
#>  [184,] 2.0235 0.8725 1.7585 1.7325 2.7965
#>  [185,] 0.1825 0.0715 0.1405 0.1775 0.0455
#>  [186,] 0.9045 0.2645 1.5475 2.2225 0.2205
#>  [187,] 4.5345 0.6425 1.1155 0.7635 0.4085
#>  [188,] 0.3975 0.8845 0.4675 0.4875 0.7005
#>  [189,] 0.4295 0.0425 0.1215 0.1755 0.1035
#>  [190,] 0.4355 3.2245 1.2185 0.4165 1.6195
#>  [191,] 2.8465 3.6525 3.4065 0.3945 1.0665
#>  [192,] 0.5385 0.9555 0.5585 0.3985 0.4485
#>  [193,] 1.2335 0.5805 1.0805 0.7945 0.3875
#>  [194,] 0.0635 0.2455 0.0895 0.2105 0.1305
#>  [195,] 0.2495 1.0415 0.2815 0.5805 0.7555
#>  [196,] 3.8235 0.4035 0.4545 1.2135 2.3305
#>  [197,] 1.5965 1.0375 0.9215 1.8885 2.2075
#>  [198,] 3.8865 1.0425 2.9605 3.5405 0.7745
#>  [199,] 0.3825 1.4575 0.4095 0.8415 3.0785
#>  [200,] 2.3145 1.1185 0.8905 1.3265 2.4805
#>  [ reached getOption("max.print") -- omitted 1619 rows ]