Box-plots of ML estimates for a specific edge across different quantiles

Given the vine tree sequence and the families of bivariate parametric models, XVineBoxPlotEachPar() creates box-plots of maximum likelihood (ML) estimates for a specific edge across three different quantiles. The functions takes as input a list of parameter estimate matrices from the function XVineBoxPlot(). The purpose of this function is reproduce Figure 8 in Kiriliouk, A., Lee, J., & Segers, J. (2023).

Usage

XVineBoxPlotEachPar(BoxOut1, BoxOut2, BoxOut3, RowInd = 1, ColInd = 2, XVS)

Arguments

BoxOut1

A list of parameter estimate matrices for repeated simulations with a specific \((k,n)\) from the output of XVineBoxPlot()

BoxOut2

A list of parameter estimate matrices for repeated simulations with a specific \((k,n)\) from the output of XVineBoxPlot()

BoxOut3

A list of parameter estimate matrices for repeated simulations with a specific \((k,n)\) from the output of XVineBoxPlot()

RowInd

Numeric; A location of a parameter matrix in row

ColInd

Numeric; A location of a parameter matrix in column

XVS

A list consisting of three components: reconstructed structure matrices, family matrices, parameter matrices, see:XVineSpec().

Value

A BoxOnePar object returns the boxplot of ML estimates for a specific edge across three different quantiles that the user specifies.

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

# Specify the number of iterations
ite=200
# Specify the structure matrix
StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
# Specify the parameter matrix
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
# Specify the family matrix                 
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
# n = 1000, k/n = 0.02, 0.05, 0.1
# Not to run (less than 1 min for a 5-dim X-vine model)
# BoxOut1K02=XVineBoxplot(N = 1000, qt = 0.02, ite = ite, XVS = XVS, RankT = T)
# BoxOut1K05=XVineBoxplot(N = 1000, qt = 0.05, ite = ite, XVS = XVS, RankT = T)
# BoxOut1K1=XVineBoxplot(N = 1000, qt = 0.1, ite = ite, XVS = XVS, RankT = T)
# BoxPar1K=XVineBoxPlotEachPar(BoxOut1 = BoxOut1K02$MLE_WithoutFamSel,
# BoxOut2 = BoxOut1K05$MLE_WithoutFamSel,
# BoxOut3 = BoxOut1K1$MLE_WithoutFamSel,
# RowInd = 1, ColInd = 2, XVS = XVS)