Model selection for bivariate parametric families along a vine structure

Assuming that the vine structure is given, XVineFamSel chooses bivariate parametric families for each edge in each tree. Specifically, in \(T_1\), we consider the list of four candidate bivariate tail copula densities: the H\(\"{u}\)sler-Reiss exponent measure, negative logistic model, logistic model, and Dirichlet model. Similarly, we consider any type of bivariate pair-copula families with a single parameter in \(T_i\), \(i\ge 2\). (e.g. 0=Independence, 1=Gaussian, 3=Clayton, 4=Gumbel, 5=Frank, 6=Joe, 13=Survival Clayton, 14=Survival Gumbel, 16=Survival Joe) We use the AIC for model selection and select the bivariate (tail) copula density families with the lowest AIC value for each edge in each tree sequentially. Besides the AIC, we use two additional criteria to set bivariate parametric copulas to independence copulas: the effective sample size falls below 10 (\(n_{D_e} < 10\)) or the empirical Kendall's tau is less than 0.05 (\(\hat{\tau} < 0.05\)) for each edge \(e\in E_i\), \(i=2,\ldots,d-1\). After that, the algorithm estimates the associated parameter(s) for the selected copula family. For more details on the selection of bivariate parametric families along the vine tree sequence, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023).

Usage

XVineFamSel(
  data,
  Rank = TRUE,
  qt = 0.2,
  XVS,
  famset_tc,
  famset_cop,
  selectioncrit = "AIC",
  method = "mle",
  effsampsize = 10,
  tau_threshold = 0.05,
  weights = NA
)

Arguments

data

An \(n\times d\) data matrix from the inverted Pareto distribution. In this case, Rank=T is not recommended as the rank transformation does not properly work for the data already on uniform scale. If either the original data or samples from multivariate Pareto distribution on Pareto scale are used, then Rank=T is required.

Rank

Logical; whether rank transformation is performed or not (Rank=T; default).

qt

Numeric; a lower enough threshold used in the rank transformation (e.g. switches from Pareto scale to uniform scale)

XVS

A list consisting of three components: reconstructed structure matrices, family matrices, parameter matrices, see:XVineSpec().

famset_tc

Numeric vector; the model type of bivariate tail copulas. Possible tail copula models are:

1=Husler-Reiss model

2=Negative logistic model

3=Logistic model

4=Dirichlet model

famset_cop

Numeric vector; the class of bivariate copula families with a single parameter

selectioncrit

Character; the information criteria for model selections (default: AIC)

method

Character; parameter estimation method (Default; mle)

effsampsize

Integer; indicates the effective sample size (Default: 10) for the independence copula

tau_threshold

Numeric; indicates the value of the Kendall's tau (Default: 0.05) for the independence copula

weights

Logical; weights for missing values (Default; NA)

Value

A list of matrix components:

Params: A \(d\times d\) strict upper triangular matrix including ML estimates

famsel: A \(d\times d\) strict upper triangular matrix including the selected families based on the selection criteria

count: A \(d\times d\) strict upper triangular matrix indicating 1 if correctly selected 0 otherwise

DepMeasure: A \(d\times d\) strict upper triangular matrix containing estimates for dependence measures

ll_selected: An upper triangular matrix that contains log-likelihood values for selected families

AIC_selected: An upper triangular matrix that contains (averaged) AIC values for selected families

BIC_selected: An upper triangular matrix that contains (averaged) BIC values for selected families

EffectSamp: A \(d\times d\) strict upper triangular matrix containing effective sample sizes \(n_{D_e}\) for selected families

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
# Pareto random samples
Dat_P=ParetoSim(n = 2000, XVS = XVS) # Pareto scale
# Specify the class of families
familyset_tc=c(1:4)
familyset_cop=c(0,1,3,4,5,6,13,14,16)
# Sequential model selection of bivariate parametric families
FamSelOut=XVineFamSel(data = Dat_P, Rank = TRUE, qt = 0.05, XVS = XVS
, famset_tc = familyset_tc, famset_cop = familyset_cop)