Fit X-vine models

Fits X-Vine models to a d-dimensional inverted-Pareto data set. The function selects vine tree structures, using maximum spanning tree. and select bivariate exponent measure classes and pair-copula classes and estimate parameters for the corresponding classes Users can specify the type of edge weights for the first tree and subsequent trees separately and the class of bivariate exponent measures and pair-copulas

Usage

XVineModelFit(
  data,
  Rank = TRUE,
  Rank_chiU = TRUE,
  Rank_chiL = FALSE,
  Rank_chi3 = TRUE,
  MST1_HR = FALSE,
  qt = 0.2,
  N = 8000,
  XVS,
  tcfamset = c(1, 2, 3, 4),
  pcfamset = c(0, 1, 3, 4, 5, 6, 13, 14, 16),
  selectioncrit = "AIC",
  BIC_graph = TRUE,
  Chi3_graph = FALSE,
  trunclevel = FALSE,
  progress = TRUE,
  treecritT1 = "chi",
  treecritT2 = "tau",
  si = 0.9,
  effsampsize = 10,
  tau_threshold = 0.05,
  se = FALSE,
  weights = NA,
  cores = 1
)

Arguments

data

A \(n\times d\) data matrix of multivariate Pareto samples.

Rank

Logical; whether rank transformation is performed or not (Rank=T; default).

Rank_chiU

Logical; whether rank-based samples are used or not for creating the chi-plot of the upper pairwise tail dependence measure. If Rank_chiU=FALSE, then ChiUPlot compares model-based chi's with fitted chi's via Monte Carlo simulation.

Rank_chiL

Logical; whether rank-based samples are used or not for creating the chi-plot of the lower pairwise tail dependence measure. If Rank_chiL=FALSE, then ChiLPlot compares model-based chi's with fitted chi's via Monte Carlo simulation.

Rank_chi3

Logical; Logical; whether rank-based samples are used or not for creating the chi-plot of the trivariate tail dependence measure. If Rank_chi3=FALSE, then ChiLPlot compares model-based chi's with fitted chi's via Monte Carlo simulation.

MST1_HR

Logical; the minimum spanning tree for the Husler-Reiss model is plotted or not.

qt

Numeric; a lower threshold for the rank transformation. It switches from Pareto scale to uniform scale.

N

Numeric; sample size

XVS

A list consisting of three components: reconstructed structure matrices, family matrices, parameter matrices, see:XVineSpec().

tcfamset

Numeric vector; the class of bivariate exponent measures.

pcfamset

Numeric vector; the class of bivariate pair-copula models.

selectioncrit

Character string; indicates the selection criteria for the class of bivariate exponent measures or pair-copulas.

BIC_graph

Logical; whether the BIC graph over the tree level is plotted or not.

Chi3_graph

Logical; whether the chi-plot of the trivariate tail dependence measure is plotted or not.

trunclevel

Numeric; indicates the specified truncation level.

progress

Logical; whether the progress of selecting vine tree structures is printed.

treecritT1

A character string, indicating the tree criterion for the first tree.

treecritT2

A character string, indicating the tree criterion for subsequent trees.

si

Numeric; a tuning parameter for mBIC (\(\si_0=0.9\); default).

effsampsize

Numeric; the specified effective sample size for the independence copula (\(n_{D_e}\)<10; default).

tau_threshold

Numeric; the specified Kendall's tau value for the independence copula (\(\hat{\tau}_e\)<0.05; default)

se

Logical; whether standard errors for ML estimators are reported.

weights

Logical; whether weights should be assigned to observations when missing values exist.

cores

Numeric; indicates the number of cores for parallel computing (optional).

Value

A nested list object containing:

MST: Maximum spanning trees

VineTree: Fitted X-vine models

mBIC_g: the plot of mBIC values across tree levels

BIC_g: the plot of BIC values across tree levels

XVS_spec: the X-vine specification from fitted X-vine models

emp_chimat: the matrix of empirical pairwise chi's

XVine_chimat: the matrix of fitted pairwise chi's

ChiLPlot: the chi-plot of the lower pairwise tail dependence measure

ChiUPlot: the chi-plot of the upper pairwise tail dependence measure

Chi3Plot: the chi-plot of the trivariate tail dependence measure

TruncLevelStar: the optimal truncation level determined as the minimum between the tree level with the lowest mBIC value and the tree level such that all pair-copulas are set to independence copula

TruncLevel_mBICmin: the truncation level with the lowest mBIC value

Examples

StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
Dat_P=ParetoSim(n = 2000, XVS = XVS) # Pareto scale
Out=XVineModelFit(data = 1/Dat_P, N = 2000, XVS = XVS, Rank = FALSE
, Rank_chiU = FALSE, Rank_chiL = FALSE, Rank_chi3 = FALSE
, Chi3_graph = FALSE, tcfamset = c(1:4), pcfamset = c(0,1,3,4,5,6,13,14,16)
, selectioncrit = "AIC", trunclevel = FALSE, progress = TRUE
, treecritT1 = "chi", treecritT2 = "tau", effsampsize = 10, tau_threshold = 0.05
, weights = NA, cores=1)
#> V1 + V2 --> V1,V2 ; V3
#> V2 + V4 --> V2,V4 ; V3
#> V3 + V5 --> V3,V5 ; V2
#> V1 + V4 --> V1,V4 ; V2,V3
#> V4 + V5 --> V4,V5 ; V2,V3
#> V1 + V5 --> V1,V5 ; V4,V2,V3