Sequential parameter estimation for X-Vine models

XVineSeqEst estimates parameters associated with bivariate (tail) copula densities tree by tree sequentially from \(T_1\) to \(T_{d-1}\). Information about the X-Vine specification is contained in the object from XVineSpec().

Usage

XVineSeqEst(data, Rank = TRUE, qt = 0.2, XVS, method = "mle")

Arguments

data

An \(n\times d\) data matrix from the inverted Pareto distribution. In this case, Rank=T is not recommended as the rank transformation does not properly work for the data already on uniform scale. If either the original data or samples from multivariate Pareto distribution on Pareto scale are used, then Rank=T is required.

Rank

Logical; whether rank transformation is performed or not (Rank=T; default). It switches from Pareto scale to uniform scale.

qt

Numeric; a lower threshold for the rank transformation.

XVS

A list consisting of three components: reconstructed structure matrices, family matrices, parameter matrices, see:XVineSpec().

method

Character; indicates the parameter estimation method (method="mle"; default).

Value

A list of four components: the matrix of sequentially estimated parameters are stored in Params, the estimated dependence measures are stored in the matrix DepMeasure where the first row has empirical Chi values and subsequent trees have empirical Kendall's Tau, , the effective sample size for each edge in the vine tree is stored in the matrix EffectSamp, and the log-likelihood value for each edge in each tree is stored in the matrix logLik.

Examples

StrMtx <- matrix(c(1, 1, 2, 2, 4,
0, 2, 1, 3, 2,
0, 0, 3, 1, 3,
0, 0, 0, 4, 1,
0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                0, 0, 2, 2.5, 0.7,
                0, 0, 0, 0.4, -0.3,
                0, 0, 0, 0, 0.1,
                0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
# X-vine speicification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
# Pareto random samples
Dat_P=ParetoSim(n = 5000, XVS = XVS) # Pareto scale
# Sequential parameter estimation
SeqEstOut=XVineSeqEst(data = Dat_P, Rank = TRUE, qt = 0.05, XVS=XVS, method = 'mle')