Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>               [,1]        [,2]        [,3]        [,4]         [,5]
#>   [1,] 0.354389624 0.830443438 0.393189148 0.615407582  1.328025920
#>   [2,] 0.167996601 0.186047277 0.342276099 0.165385603  0.125320990
#>   [3,] 0.390837740 0.232515249 0.223203086 0.178303578  0.460405160
#>   [4,] 0.537984893 0.554072497 0.412412844 0.306826663  0.866312876
#>   [5,] 0.285644134 2.020841794 0.602327363 1.263543476  2.531718863
#>   [6,] 0.779145234 1.654030649 1.421138450 1.334895343  1.644656317
#>   [7,] 0.057953481 0.120649438 0.039742083 0.052449641  0.905614186
#>   [8,] 0.308834920 1.288462316 0.475597659 0.669843507  2.545534041
#>   [9,] 0.757872674 0.207196159 0.240395238 0.175472026  0.169888286
#>  [10,] 0.273666105 0.911253574 0.407492303 0.745219907  0.894617850
#>  [11,] 0.237646488 1.331839889 0.830234015 0.526624380  0.735316403
#>  [12,] 0.257990305 0.163896150 0.139827897 0.174789309  0.113087657
#>  [13,] 0.111985289 0.363294480 0.156090719 0.106966334  0.302205977
#>  [14,] 0.679152996 4.067292134 1.092789539 1.179350640  3.799917654
#>  [15,] 0.827385069 4.635186036 1.237032997 3.521219921  3.688841778
#>  [16,] 0.824043555 0.835647781 0.743976293 2.011089253  3.077285334
#>  [17,] 0.105898095 0.527410601 0.184659975 0.109351562  0.282105797
#>  [18,] 0.246566098 0.541612384 0.243468227 0.507759972  0.091694672
#>  [19,] 0.780706571 9.122083341 1.382494104 2.617036919  3.128564526
#>  [20,] 0.846760724 1.018954215 2.013881035 1.717133453  0.496896444
#>  [21,] 0.852260908 9.116949815 2.003392699 2.619216622 39.422608776
#>  [22,] 0.991900122 1.210234719 0.828986798 1.105363065  1.103915618
#>  [23,] 0.876477760 1.973380264 1.422309020 1.704784708  0.871878529
#>  [24,] 0.229551360 5.540247016 0.453242506 0.828073803 10.329211792
#>  [25,] 0.419750036 1.255900767 0.649185530 2.326692985  0.725950332
#>  [26,] 0.312050488 0.117563263 0.117507105 0.151189228  0.124501202
#>  [27,] 0.273973343 0.932473810 0.369207213 0.798163956  1.603769941
#>  [28,] 0.160491458 2.975419431 0.362451083 0.356088967  1.639382389
#>  [29,] 0.626472648 3.457308323 1.054448811 0.462020427  3.913066031
#>  [30,] 0.837921125 0.849880796 1.336283459 2.068660346  1.758493420
#>  [31,] 0.558336226 1.039156082 1.244729095 1.632359816  1.404263843
#>  [32,] 0.457384514 0.381920930 0.226581050 0.445879130  0.871294348
#>  [33,] 0.269861608 1.546478182 0.401043261 0.322278605  4.004657414
#>  [34,] 0.008545102 0.004770188 0.004577683 0.004364536  0.021240975
#>  [35,] 0.893435950 3.944989686 1.327810868 2.391931181  2.652022736
#>  [36,] 0.668577439 0.663626447 0.566573552 0.786106532  2.506467647
#>  [37,] 0.968413192 0.258727563 0.746473715 0.322478919  0.068148505
#>  [38,] 0.546956120 1.026269587 0.731449226 1.013975718  1.358401052
#>  [39,] 0.160344653 0.159005559 0.180714929 0.177937879  0.141312396
#>  [40,] 0.742766666 4.137790418 1.990623596 1.777096811  1.260236377
#>  [41,] 0.282624009 1.760301582 0.652609270 0.390230536  3.224212904
#>  [42,] 0.118704702 0.308681230 0.185190985 0.074545780  0.135412988
#>  [43,] 0.059756865 0.122715806 0.051123207 0.101381557  0.663417507
#>  [44,] 0.264780977 0.016319248 0.034075762 0.029227033  0.007944774
#>  [45,] 0.120737268 1.211742242 0.339533095 0.363825369  0.599443667
#>  [46,] 0.451402592 4.243688685 0.983638007 1.278438329  1.788136109
#>  [47,] 0.248732409 0.399014077 0.201673399 0.431007912  0.534948763
#>  [48,] 0.996246258 0.889881828 0.899410681 1.048031723  3.951525991
#>  [49,] 0.600876339 2.438805207 0.796558018 1.100490428  5.833866839
#>  [50,] 0.319983165 1.069773859 0.507330175 0.923881230  0.755465157
#>  [51,] 0.205653232 4.366231066 0.450552557 1.534073282  8.886556151
#>  [52,] 0.262841954 0.047128673 0.100706590 0.125484637  0.120994257
#>  [53,] 0.963416351 7.991036204 1.554017130 5.095573873 34.432632791
#>  [54,] 0.432791110 0.163996250 0.299704344 0.545793107  0.486366585
#>  [55,] 0.422521611 5.130884438 0.868473965 3.002387878 55.783214428
#>  [56,] 0.751228282 6.681328529 1.893064134 5.274760676  5.432000007
#>  [57,] 0.813249875 1.005697574 3.171446648 2.961593540  2.070229939
#>  [58,] 0.139318482 0.167724257 0.117130784 0.146867849  0.316064906
#>  [59,] 0.960560260 1.938610239 0.692203915 0.862191037 14.933706884
#>  [60,] 0.125119337 0.055346210 0.038312963 0.093238141  0.045479994
#>  [61,] 0.101194211 0.622547195 0.299245561 0.363218390  1.597519789
#>  [62,] 0.730968826 6.930985220 2.396759231 0.514846982  5.411620978
#>  [63,] 0.987397784 0.653320188 2.044011347 1.248328886  0.290346082
#>  [64,] 0.988979317 3.552703649 1.918198981 2.303929897  3.418724795
#>  [65,] 0.865018954 2.461918541 0.896942753 0.994725998 16.760458725
#>  [66,] 0.614401825 1.535454183 0.622780991 0.789544240  0.775565076
#>  [67,] 0.839843981 0.276517521 0.588905653 0.567647902  0.067998263
#>  [68,] 0.651634972 0.249690637 0.184724839 0.157849616  0.067128824
#>  [69,] 0.243467953 0.259843120 0.360255763 0.445241621  0.141324205
#>  [70,] 0.799497975 7.326367760 1.018888581 1.687594798  5.507265274
#>  [71,] 0.723466353 3.307909967 1.131439624 1.092634823  2.671337353
#>  [72,] 0.448180989 0.194111672 0.256412743 0.205089803  0.262486258
#>  [73,] 0.937405116 0.970196688 1.000400783 1.312038076  0.683652901
#>  [74,] 0.157614961 0.308974519 0.241319839 0.528942698  0.181264763
#>  [75,] 0.528343706 0.422372466 0.225203837 0.181962024  0.481557066
#>  [76,] 0.567510255 1.351184174 0.675301646 1.686423466  3.883241617
#>  [77,] 0.021133909 0.071057837 0.037242488 0.070172917  0.291433888
#>  [78,] 0.613329560 3.927884815 1.800330418 1.134161164  1.857967549
#>  [79,] 0.145093988 0.648163656 0.189571673 0.273925644  0.446800966
#>  [80,] 0.183205071 3.433709920 0.374597937 0.482227272  6.371905497
#>  [81,] 0.095425842 0.027935397 1.852801905 3.719061135  0.312840531
#>  [82,] 0.587353955 1.058264132 1.204971545 0.774128656  2.786042663
#>  [83,] 0.077145803 0.683940621 0.132775161 0.090434910  0.688073949
#>  [84,] 0.519620677 1.269480601 0.499088157 0.693695161  1.463489380
#>  [85,] 0.504422645 0.315360446 0.508882206 0.629206357  0.128506425
#>  [86,] 0.322294203 0.138547755 0.422196909 0.281551334  0.142334999
#>  [87,] 0.221482481 1.074141148 0.591181594 1.199350072  1.736304379
#>  [88,] 0.414425447 1.427317499 0.885632873 0.792750676  0.930588930
#>  [89,] 0.432759657 0.336097606 0.241903081 0.292833452  0.681271060
#>  [90,] 0.979976673 6.691695904 2.861535938 1.843751383  6.606225178
#>  [91,] 0.373994218 0.040374323 0.024141279 0.016500747  0.140003955
#>  [92,] 0.545369614 1.006924923 0.554513142 0.426377330  0.697262084
#>  [93,] 0.054426453 0.357526675 0.080465128 0.181143491  0.638446406
#>  [94,] 0.188994883 0.083825051 0.121554942 0.080575022  0.246502266
#>  [95,] 0.637058085 3.039306583 1.118641243 0.965159116  4.266810033
#>  [96,] 0.964766207 3.517963190 3.259866528 2.254192487  4.105343275
#>  [97,] 0.133546668 0.015620199 0.024023468 0.013574599  0.019988817
#>  [98,] 0.284131281 0.416793937 0.388877434 0.257781829  0.447503623
#>  [99,] 0.248851796 0.959369979 0.404700852 1.779493276  2.054770515
#> [100,] 0.121560641 0.167746214 0.085373204 0.042443050  0.096070302