Reproduce the given X-vine matrix specification depending on each conditioning variable

Given the initial X-vine matrix specification of (structure matrix, family matrix, parameter matrix), XVineSpec() generates a list of redefined X-vine specifications for each conditioning variable. Preserving the vine structure, the function reconstructs structure matrices where the first diagonal element , denoted as \(m_{11}\in [d]\), corresponds to the conditioning variable. Additionally, XVineSpec() includes the associated parameter and family matrices based on the given structure matrices.

Usage

XVineSpec(M, Mmod, Mpar)

Arguments

M

A \(d \times d\) upper triangular structure matrix.

Mmod

A \(d \times d\) strict upper triangular family matrix for identifying bivariate (tail) copulas.

Mpar

A \(d \times d\) strict upper triangular parameter matrix associated with the family matrix.

Value

A list of arrays includes:

xmat: array of dimension \(d \times d \times d\) containing rearranged X-vine structure matrices.

fmat: array of dimension \(d \times d \times d\) containing family matrices associated with the structure matrices.

pmat: array of dimension \(d \times d \times d\) containing parameter matrices associated with the structure matrices.

xedge: A list of edges corresponding to the initial structure matrix.

Details

To use the exact simulation algorithm for multivariate Pareto distribution in Engelke and Hitz (2020) and Kiriliouk et al. (2023), we condition on each variable to calculate extremal functions. For each conditioning variable, the index of the conditioning variable is placed on the first diagonal element in the structure matrix. Specifically, in a \(d \times d\) structure matrix, the element \(m_{jj}\in [d]\) corresponds to the index of the conditioning variable \(X_j\) for \(j=1,\ldots,d\). All reproduced structure matrices maintain the same vine tree structure. The corresponding parameter matrices contain the parameter values for the bivariate (tail) copulas. In family matrices, the first row specifies bivariate tail copulas and the remaining rows specify bivariate copulas.

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Engelke, S., & Hitz, A. S. (2020). Graphical models for extremes. Journal of the Royal Statistical Society Series B: Statistical Methodology, 82(4), 871-932.

See also

Examples

##  A 6 x 6 structure matrix
M <- cbind(c(1,0,0,0,0,0), c(1,4,0,0,0,0), c(4,1,6,0,0,0),
          c(4,1,6,3,0,0), c(6,4,1,3,5,0), c(3,4,1,6,5,2))
##  A 6 x 6 family matrix
Mmod <- cbind(rep(0,6),c("neglog", rep(0,5)),
              c("hr", "clayton",  rep(0,4)),
              c("log", "gumbel", "joe", rep(0,3)),
              c("dir", "frank", "clayton", "gumbel", 0, 0),
              c("neglog", "joe", "frank", "clayton", "gumbel", 0))
##  A 6 x 6 parameter matrix
Mpar <- cbind(rep(0,6),c(0.25, rep(0,5)), c(0.3, 0.7,  rep(0,4)),
              c(0.5,0.7,0.1, rep(0,3)), c(0.2,0.8,0.3,0.55, 0, 0),
              c(0.1,0.9,0.75,0.3,0.45, 0))
##  Permute the given X-vine specification
PermXVineSpec <- XVineSpec(M,Mmod,Mpar)
PermXVineSpec[[1]]
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    1    4    4    6    3
#> [2,]    0    4    1    1    4    4
#> [3,]    0    0    6    6    1    1
#> [4,]    0    0    0    3    3    6
#> [5,]    0    0    0    0    5    5
#> [6,]    0    0    0    0    0    2
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    2    2    3    4    4    6
#> [2,]    0    3    2    3    1    4
#> [3,]    0    0    4    2    3    1
#> [4,]    0    0    0    1    2    3
#> [5,]    0    0    0    0    6    2
#> [6,]    0    0    0    0    0    5
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    3    3    4    3    4    6
#> [2,]    0    4    3    4    1    4
#> [3,]    0    0    1    1    3    1
#> [4,]    0    0    0    2    2    3
#> [5,]    0    0    0    0    6    2
#> [6,]    0    0    0    0    0    5
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    4    4    4    3    4    6
#> [2,]    0    1    1    4    1    4
#> [3,]    0    0    3    1    3    1
#> [4,]    0    0    0    2    2    3
#> [5,]    0    0    0    0    6    2
#> [6,]    0    0    0    0    0    5
#> 
#> , , 5
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    5    5    6    4    4    3
#> [2,]    0    6    5    6    1    4
#> [3,]    0    0    4    5    6    1
#> [4,]    0    0    0    1    5    6
#> [5,]    0    0    0    0    3    5
#> [6,]    0    0    0    0    0    2
#> 
#> , , 6
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    6    6    4    4    3    6
#> [2,]    0    4    6    1    4    4
#> [3,]    0    0    1    6    1    1
#> [4,]    0    0    0    3    6    3
#> [5,]    0    0    0    0    2    2
#> [6,]    0    0    0    0    0    5
#> 
cbind(PermXVineSpec[[1]][,,5],PermXVineSpec[[2]][,,5])
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]  [,9]    [,10]     [,11]   
#> [1,] "5"  "5"  "6"  "4"  "4"  "3"  "0"  "dir" "hr"    "neglog"  "log"   
#> [2,] "0"  "6"  "5"  "6"  "1"  "4"  "0"  "0"   "frank" "clayton" "gumbel"
#> [3,] "0"  "0"  "4"  "5"  "6"  "1"  "0"  "0"   "0"     "clayton" "joe"   
#> [4,] "0"  "0"  "0"  "1"  "5"  "6"  "0"  "0"   "0"     "0"       "gumbel"
#> [5,] "0"  "0"  "0"  "0"  "3"  "5"  "0"  "0"   "0"     "0"       "0"     
#> [6,] "0"  "0"  "0"  "0"  "0"  "2"  "0"  "0"   "0"     "0"       "0"     
#>      [,12]    
#> [1,] "neglog" 
#> [2,] "joe"    
#> [3,] "frank"  
#> [4,] "clayton"
#> [5,] "gumbel" 
#> [6,] "0"      
##  Compare with cbind(M,Mmod): looks ok!
##  Truncated X-vine specification at tree level 3
M <- cbind(c(1,0,0,0,0,0), c(1,4,0,0,0,0), c(4,1,6,0,0,0),
           c(4,1,6,3,0,0), c(6,4,1,0,5,0), c(3,4,1,0,0,2))
Mmod <- cbind(rep(0,6),c("neglog", rep(0,5)), c("hr", "clayton",  rep(0,4)), 
           c("log", "gumbel", "joe", rep(0,3)), c("dir", "frank", "clayton", 0, 0, 0),
           c("neglog", "joe", "frank", 0,0, 0))
Mpar <- cbind(rep(0,6),c(0.25, rep(0,5)), c(0.3, 0.7,  rep(0,4)), 
              c(0.5,0.7,0.1, rep(0,3)), c(0.2,0.8,0.3,0, 0, 0),
              c(0.1,0.9,0.75,0,0, 0))
PermTrucXVine <- XVineSpec(M,Mmod,Mpar)
PermTrucXVine[[1]]
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    1    4    4    6    3
#> [2,]    0    4    1    1    4    4
#> [3,]    0    0    6    6    1    1
#> [4,]    0    0    0    3    0    0
#> [5,]    0    0    0    0    5    0
#> [6,]    0    0    0    0    0    2
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    2    2    3    4    4    6
#> [2,]    0    3    2    3    1    4
#> [3,]    0    0    4    2    3    1
#> [4,]    0    0    0    1    0    0
#> [5,]    0    0    0    0    6    0
#> [6,]    0    0    0    0    0    5
#> 
#> , , 3
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    3    3    4    3    4    6
#> [2,]    0    4    3    4    1    4
#> [3,]    0    0    1    1    3    1
#> [4,]    0    0    0    2    0    0
#> [5,]    0    0    0    0    6    0
#> [6,]    0    0    0    0    0    5
#> 
#> , , 4
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    4    4    4    3    4    6
#> [2,]    0    1    1    4    1    4
#> [3,]    0    0    3    1    3    1
#> [4,]    0    0    0    2    0    0
#> [5,]    0    0    0    0    6    0
#> [6,]    0    0    0    0    0    5
#> 
#> , , 5
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    5    5    6    4    4    3
#> [2,]    0    6    5    6    1    4
#> [3,]    0    0    4    5    6    1
#> [4,]    0    0    0    1    0    0
#> [5,]    0    0    0    0    3    0
#> [6,]    0    0    0    0    0    2
#> 
#> , , 6
#> 
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    6    6    4    4    3    6
#> [2,]    0    4    6    1    4    4
#> [3,]    0    0    1    6    1    1
#> [4,]    0    0    0    3    0    0
#> [5,]    0    0    0    0    2    0
#> [6,]    0    0    0    0    0    5
#> 
cbind(PermTrucXVine[[1]][,,5],PermTrucXVine[[2]][,,5]) #compare with cbind(M,Mmod): looks ok!
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]  [,9]    [,10]     [,11]   
#> [1,] "5"  "5"  "6"  "4"  "4"  "3"  "0"  "dir" "hr"    "neglog"  "log"   
#> [2,] "0"  "6"  "5"  "6"  "1"  "4"  "0"  "0"   "frank" "clayton" "gumbel"
#> [3,] "0"  "0"  "4"  "5"  "6"  "1"  "0"  "0"   "0"     "clayton" "joe"   
#> [4,] "0"  "0"  "0"  "1"  "0"  "0"  "0"  "0"   "0"     "0"       "0"     
#> [5,] "0"  "0"  "0"  "0"  "3"  "0"  "0"  "0"   "0"     "0"       "0"     
#> [6,] "0"  "0"  "0"  "0"  "0"  "2"  "0"  "0"   "0"     "0"       "0"     
#>      [,12]   
#> [1,] "neglog"
#> [2,] "joe"   
#> [3,] "frank" 
#> [4,] "0"     
#> [5,] "0"     
#> [6,] "0"