Fit the first maximum spanning tree to multivariate inverted-Pareto data

fit.FirstTree() fits the first maximum spanning tree to multivariate inverted-Pareto data set. The function takes a pair of variables for each edge and variables are passed to the function 'tcSelect'. After selecting the family of bivariate tail copula models, store the output of the fitted model in the list of MST , including parameter estimates, model classes, fitted models, and pseudo-observations.

Usage

fit.FirstTree(MST, data.invPa, tcfamset, si, selectioncrit, weights = NA)

Arguments

MST

A list of maximum spanning trees.

data.invPa

A \(n\times d\) data matrix from multivariate inverse-Pareto distribution.

tcfamset

Numeric vector; specifies the class of bivariate tail copula models that users consider.

si

Numeric; a tuning parameter for mBIC (\(\si_0=0.9\); default).

selectioncrit

Character string, indicating the selection criteria for the class of bivariate tail copula models.

weights

Logical; whether weights should be assigned to observations when missing values exist.

Value

a nested list object:

the maximum spanning tree with a node set and edge set.

the fitted tail copula model with selected classes, parameter estimates, log-likelihood values, AIC, and pseudo-observations.