Fitting subsequent maximum spanning trees to a multivariate inverted-Pareto data set

Fits all previous maximum spanning trees to a multivariate inverted-Pareto data set. The function uses the recursive relationship for conditional distributions. The properly selected arguments for conditional pair-copulas are passed to the function 'copSelect'. After selecting the class of pair-copula models with the lowest AIC, stores the output of fitted models in the list of MST, including parameter estimates, model classesl (numeric) in the list of MST. After that, the function stores properly defined pseudo-copula observations through functions 'repseudo.SecTree' for the second tree and 'pseudo.Subtree' for subsequent trees in the list of MST.

Usage

fit.SubTree(
  data,
  MST,
  VineTree,
  copfamset,
  tree,
  si,
  selectioncrit,
  progress,
  effsampsize,
  tau_threshold,
  weights = NA,
  se = FALSE,
  cores = 1
)

Arguments

data

A \(n\times d\) inverted-Pareto sample data set.

MST

A list of maximum spanning trees.

VineTree

A list of fitted vine tree structures and fitted models.

copfamset

Numeric; indicates the class of pair-copula models users consider.

tree

Numeric value of tree levels.

si

Numeric; a tuning parameter for mBIC (\(\si_0=0.9\); default).

selectioncrit

Character string indicating the model selection criteria, e.g. 'AIC'.

progress

Logical; whether the vine tree selection is reported.

effsampsize

Numeric; the specified effective sample size for the independence copula (\(n_{D_e}\)<10; default).

tau_threshold

Numeric; the specified Kendall's tau value for the independence copula (\(\hat{\tau}_e\)<0.05; default).

weights

Logical; whether the weights are applied to missing values.

se

Logical; whether standard errors for ML estimators are reported.

cores

Numeric; indicates the number of cores for parallel computing (optional).

Value

A nested list of MST including the maximum spanning tree and fitted vine models.