Reconstruct a structure matrix with diagonal elements put in ascending order

standardStrMtx() reconstructs a given structure matrix to ensure that the diagonal elements of a structure matrix are arranged in ascending order. The advantage of this reconstruction is that we can create a so-called max-matrix (see:createMaxMtx()) , making it easier for the algorithm to find the correct argument for conditional pair-copulas when recursive relationships are used.

Usage

standardStrMtx(Matrix, oldOrder = NULL)

Arguments

Matrix

A numeric \(d \times d\) structure matrix

oldOrder

Logical; If NULL (the default), the function uses the order of diagonal elements in the structure matrix.

Value

A numeric \(d \times d\) structure matrix with diagonal elements arranged in ascending order.

Examples

##  A 5 x 5 structure matrix where diagonal elements are not arranged in ascending order.
StrMtx <- cbind(c(2,0,0,0,0),c(2,4,0,0,0),c(4,2,5,0,0),c(2,4,5,3,0),c(2,3,4,5,1))

##  A 5 x 5 structure matrix with diagonal elements arranged in increasing order.
standardStrMtx(StrMtx)
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    1    2    1    1
#> [2,]    0    2    1    2    4
#> [3,]    0    0    3    3    2
#> [4,]    0    0    0    4    3
#> [5,]    0    0    0    0    5