Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>                [,1]         [,2]         [,3]         [,4]         [,5]
#>   [1,] 0.0899445564  0.192382365 0.0861051720 0.1017160108 2.896223e-01
#>   [2,] 0.8615440910  4.514350402 1.8670079676 2.2538673667 1.337345e+01
#>   [3,] 0.8772439281  5.953630682 1.4182663952 1.5997274480 3.796651e+00
#>   [4,] 0.2992546868  0.912429210 0.6910967334 0.1953576848 2.762277e+00
#>   [5,] 0.4363371809  4.161327140 0.8287771372 1.8966533423 3.012894e+00
#>   [6,] 0.0654729158  0.143128307 0.1334544341 0.0573276835 1.511374e-01
#>   [7,] 0.0062730839  0.006597459 0.0038784446 0.0044254046 8.729966e-02
#>   [8,] 0.7507964445  0.787603579 1.3643795214 0.9654766500 6.446428e-01
#>   [9,] 0.9778413307  0.824571697 1.1403516411 0.9670589162 7.212866e-01
#>  [10,] 0.7820857090  2.321530494 1.0998592070 2.1098750977 9.318148e-01
#>  [11,] 0.9872680991  4.290783683 1.7246762173 2.2167481229 4.035997e+00
#>  [12,] 0.5132228169  1.039254970 0.7729895705 1.2835855331 1.282335e+00
#>  [13,] 0.4307093953  0.504058148 0.5892423815 1.1090216008 1.888664e-01
#>  [14,] 0.6896437330  0.979621005 0.7878408512 0.3263778297 1.193698e+00
#>  [15,] 0.1898270010  0.612952557 0.2220439901 0.3123423426 2.308182e-01
#>  [16,] 0.3968370920  1.254063488 0.8084146608 1.0624397376 8.980879e-01
#>  [17,] 0.2367393770  0.605353719 0.2222050689 0.3330648322 2.157401e+00
#>  [18,] 0.0060115431  0.001755029 0.0028998032 0.0035522992 1.082579e-03
#>  [19,] 0.8794566309  1.554903768 0.6800470704 0.7027590650 1.210440e+00
#>  [20,] 0.6621736770  5.170000283 1.2700314257 1.2666493730 3.004943e+00
#>  [21,] 0.5344608768  4.156073254 1.0224843887 1.6159236108 1.161275e+01
#>  [22,] 0.0777318543  0.257467758 0.1636371244 0.0814755009 1.130380e-01
#>  [23,] 0.8912260269  5.645064915 1.8587549278 3.0191002348 9.665153e+00
#>  [24,] 0.0829123831  1.600348184 0.1534377602 0.9150661573 3.361518e+00
#>  [25,] 0.9803906998  2.316698222 1.1392977377 1.1751237314 1.699288e+00
#>  [26,] 0.6808708010  1.042058592 0.5925062275 0.2092521144 1.398503e+00
#>  [27,] 0.5995006065  0.929915152 1.0210523913 0.7283716672 2.260802e+00
#>  [28,] 0.2946460394  3.199514375 0.8238015028 0.4807332835 4.565300e+00
#>  [29,] 0.6092156123  0.391858393 0.5262778203 0.5297077711 4.439458e-01
#>  [30,] 0.5158673986  4.727907498 0.6594053283 2.7081133305 3.098801e+01
#>  [31,] 0.9355649569  0.215682608 0.2327857148 0.2993933730 1.588495e-01
#>  [32,] 0.0833399806  0.621346201 0.1996269508 0.1871879162 6.493334e-01
#>  [33,] 0.5689760603  0.113614372 1.2659807895 1.5584973318 3.932250e-01
#>  [34,] 0.9897076767  0.442522051 0.3319521113 0.6040168498 2.958291e-01
#>  [35,] 0.0857321746  0.629043391 0.2226540250 0.4979168034 8.845438e-01
#>  [36,] 0.4940969471  1.286040480 0.5672848500 0.9196428148 1.826582e+00
#>  [37,] 0.6542937867 13.482717905 1.2132161728 4.6397269409 1.383673e+01
#>  [38,] 0.0794546667  0.129235180 0.0738702792 0.0897445430 2.313369e-01
#>  [39,] 0.0861666594  0.075921053 0.0424669745 0.1817637964 1.640286e-01
#>  [40,] 0.4774168923  3.573707763 0.9270319778 0.1444761614 9.980694e-01
#>  [41,] 0.5976984196  0.421214368 0.2709148350 0.2161206411 6.431867e-01
#>  [42,] 0.8588834179 10.868754969 2.4020422866 1.6336060463 8.817521e+00
#>  [43,] 0.0075949535  0.003414591 0.0157014126 0.0090795605 4.906240e-03
#>  [44,] 0.6146416853  1.087946711 0.4807296340 0.7092911408 5.410391e+00
#>  [45,] 0.8026544678  5.421904787 2.3301007079 3.1189599240 2.526142e+00
#>  [46,] 0.7260885641  0.404095336 0.6810411005 0.6836426159 3.369279e-01
#>  [47,] 0.5215446672  3.490148920 0.9367719013 2.4480353500 7.148051e+00
#>  [48,] 0.2868377017  8.889122838 0.5494137103 2.4196605010 4.370874e+01
#>  [49,] 0.2082752173  0.062814470 0.0862753376 0.0830901467 5.492212e-02
#>  [50,] 0.6092004222  2.982662347 1.0892220535 0.5520781585 1.229889e+01
#>  [51,] 0.9136783318  0.556388093 2.4381042096 1.0405250677 3.855506e-01
#>  [52,] 0.7889046986  0.510686948 0.9930586269 0.8819993828 4.214921e-01
#>  [53,] 0.5843731903  4.154060254 0.8727990621 0.7629413644 1.017098e+01
#>  [54,] 0.2686057470  0.440391212 0.4221505943 0.4223445044 8.130143e-01
#>  [55,] 0.2465180287  0.121300145 0.0999397025 0.2609382509 2.022494e-01
#>  [56,] 0.4518820813  7.480100044 1.1441978490 1.8953314863 6.892534e+00
#>  [57,] 0.3706789508  2.418580060 0.6290063066 0.2334507737 2.896503e+00
#>  [58,] 0.8250714049  0.041246990 0.0683805957 0.0649416348 3.006092e-02
#>  [59,] 0.7901589428  0.198789104 1.3219279086 2.0160132289 4.946930e-01
#>  [60,] 0.6318953370  2.227798571 1.7165658094 2.4680039622 6.359549e+00
#>  [61,] 0.9020997498  1.556046134 1.0270081391 0.9496875720 1.131831e+00
#>  [62,] 0.0813680643  0.366281161 0.1404449721 0.2285484218 6.144594e-01
#>  [63,] 0.2571724209  0.607508052 0.2977655865 0.7641065438 1.866136e+00
#>  [64,] 0.3316755327  4.374776143 0.6045308060 0.8581817283 5.595467e+00
#>  [65,] 0.7803340482  1.109368522 0.7698292587 0.3739068794 2.073534e+00
#>  [66,] 0.3091396424  0.166605075 0.1014708042 0.1955458918 2.476934e+00
#>  [67,] 0.8412469986  2.089676270 1.5013171531 1.0536156088 1.384785e+00
#>  [68,] 0.4072819944  3.053107654 0.8070988904 0.6754163329 6.313706e+00
#>  [69,] 0.1603889922  0.847634866 0.2515730236 0.3897399427 1.724330e+00
#>  [70,] 0.5941139341  0.255815713 0.4028536533 0.2321723469 9.362071e-02
#>  [71,] 0.9933171922  4.499147123 1.4772282108 1.5012332895 6.560555e+00
#>  [72,] 0.8697444601  7.968083524 1.4457766652 7.1915608731 3.125154e+01
#>  [73,] 0.4135997163  7.744356284 0.8947753679 1.5722559659 3.034515e+00
#>  [74,] 0.0566894454  0.510790824 0.1037216806 0.1315420323 5.257088e-01
#>  [75,] 0.3821904466  0.488729760 0.3128509176 0.1996154210 1.413051e+00
#>  [76,] 0.1183257571  0.215815319 0.1242108868 0.0928160515 3.659169e-01
#>  [77,] 0.0929667496  4.915209612 0.2295880562 0.2718483183 2.259440e+00
#>  [78,] 0.4718557994  2.928351938 0.8147959495 0.3393787848 5.251509e+00
#>  [79,] 0.4696221054  4.577448894 0.7724349899 4.7469763011 1.031900e+01
#>  [80,] 0.8523420645  0.470956411 0.3008253418 0.5425921296 3.643228e-01
#>  [81,] 0.5124242788 12.318540785 1.3721306257 8.6087562060 1.881130e+01
#>  [82,] 0.0421466529  0.042079243 0.0657083870 0.0281678838 3.004254e-02
#>  [83,] 0.5496233408  0.826605581 0.3951433646 0.5673304071 1.237620e+00
#>  [84,] 0.6306029155  1.406884412 0.9832568124 1.4765989179 1.517474e+00
#>  [85,] 0.0837524631  0.100209141 0.1273436105 0.0802265461 9.013410e-02
#>  [86,] 0.1703178126  0.812911894 0.3895129923 0.1874095635 8.553312e-01
#>  [87,] 0.0008635079  0.000572249 0.0005156553 0.0005460729 6.823999e-04
#>  [88,] 0.9915712932  0.977060555 0.7942559227 1.1473007517 2.858493e+00
#>  [89,] 0.3163787788  1.034518252 1.1410031301 0.4355591350 4.440755e-01
#>  [90,] 0.5030739724  1.907009386 0.8954145312 1.5903574887 3.135945e+00
#>  [91,] 0.5333111843  0.164349398 0.1964984972 0.3092054743 1.715166e-01
#>  [92,] 0.0798574747  0.214039207 0.0958757895 0.1225874244 1.598562e-01
#>  [93,] 0.8348050043  6.571043097 2.1653614362 1.4033010071 2.663382e+00
#>  [94,] 0.3857410378  2.099322821 0.7530999851 0.9134077851 2.299703e+00
#>  [95,] 0.3748829896  0.176545699 0.9922051154 0.7034735141 2.943657e-01
#>  [96,] 0.7006545290  0.149494897 0.1946402103 0.0977485091 3.437010e-02
#>  [97,] 0.5609298311  3.099674191 1.0936222574 0.6451487041 2.375297e+00
#>  [98,] 0.2858706140  2.780020650 0.4824166994 0.6271159077 2.988726e+00
#>  [99,] 0.4569650898  0.395491566 0.3465188926 0.4839522333 4.988759e-01
#> [100,] 0.4648025485  1.473420629 0.7106493021 0.5849145000 3.251908e-01