Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>               [,1]        [,2]        [,3]        [,4]         [,5]
#>   [1,] 0.089137830  0.15683630 0.119945802  0.23226764 5.177398e-01
#>   [2,] 0.394397167 11.89613669 0.967204832  1.11048372 1.150438e+01
#>   [3,] 0.311474850  0.74262421 0.358755097  0.38125092 1.877366e+00
#>   [4,] 0.481984075  1.19906216 0.726641993  1.28352138 1.506947e+00
#>   [5,] 0.645204298  0.72915126 0.610572034  0.43210072 6.677690e-01
#>   [6,] 0.934088349  1.85609732 0.998932808  1.22994709 1.299155e+01
#>   [7,] 0.833100016 11.25762319 1.856979871 16.30742577 1.610586e+01
#>   [8,] 0.801387997  3.05443396 0.716174503  0.65664644 2.018420e+00
#>   [9,] 0.843446730  7.94877569 1.120608514  1.25376897 7.050062e+00
#>  [10,] 0.729875930  2.40881557 1.216160254  0.44758426 3.143241e-01
#>  [11,] 0.555614206  1.09871908 0.727913497  1.06190034 1.106142e+00
#>  [12,] 0.032070407  0.13429855 0.039450461  0.02874070 1.017026e-01
#>  [13,] 0.267158104 11.96785629 0.629283398  2.51539057 1.107382e+01
#>  [14,] 0.433129603  0.80533958 0.405173958  0.21425302 6.528780e-01
#>  [15,] 0.653538427  1.03276494 0.641325935  1.49289808 1.844886e+00
#>  [16,] 0.936752016  2.00600351 0.942158291  0.67690107 8.632994e-01
#>  [17,] 0.073622297  0.19053928 0.089195142  0.10814207 1.051512e-01
#>  [18,] 0.818686139  2.24940655 0.944205464  1.19959754 1.977644e+00
#>  [19,] 0.364080191  0.36828115 0.374300898  0.66249145 6.837915e-01
#>  [20,] 0.642481420  2.81707296 1.221463534  0.80016122 1.472936e+00
#>  [21,] 0.620792296  0.59933318 0.557607048  0.69679187 1.358438e+00
#>  [22,] 0.321451175  1.98597792 0.728276365  0.40881341 3.780755e+00
#>  [23,] 0.643901476  1.73776188 0.561631126  1.18230846 2.263616e+00
#>  [24,] 0.849095046  2.19227392 4.340949261  1.85869143 3.802560e-01
#>  [25,] 0.896394482  5.72632999 2.433115088  5.40127301 1.707319e+01
#>  [26,] 0.577362312  3.75691540 1.426755184  1.79545232 5.274366e+00
#>  [27,] 0.773462688  0.54195868 0.660786950  0.39038779 3.877321e-01
#>  [28,] 0.482516394  0.46201185 0.461633980  0.37667786 6.125516e-01
#>  [29,] 0.135332775  0.11554725 0.068017717  0.04156079 1.697753e-01
#>  [30,] 0.381046207  0.28501259 3.808566297  1.82001188 9.107033e-01
#>  [31,] 0.636692078  1.16346080 0.756142044  1.11454493 2.241060e+00
#>  [32,] 0.998792471 29.79510197 1.732584092 13.26383618 7.026661e+01
#>  [33,] 0.385367757  1.20119104 0.652741390  0.43688533 6.321519e-01
#>  [34,] 0.747164376  0.50862142 0.428931861  0.47393389 4.479781e-01
#>  [35,] 0.855311406  1.99495867 1.503595026  0.96396866 1.180543e+00
#>  [36,] 0.893519730  1.49233163 0.866885626  1.01738560 4.296832e+00
#>  [37,] 0.668457245  0.12458453 0.226452256  0.30617036 2.035123e-01
#>  [38,] 0.627710868 15.84961314 1.670515981  2.71637263 2.299563e+01
#>  [39,] 0.915695674  0.21637154 0.524814917  0.55203126 1.191219e-01
#>  [40,] 0.286099883  0.04436663 0.201546508  0.32908604 6.917830e-02
#>  [41,] 0.937562776 27.00908453 2.264237787 16.07793369 1.003817e+02
#>  [42,] 0.235377647  0.14511069 0.407467197  0.14439901 1.592253e-01
#>  [43,] 0.535221487  1.23765954 1.388115779  0.50611622 4.926362e-01
#>  [44,] 0.230020387  1.20601355 0.477043827  1.25112173 7.402598e+00
#>  [45,] 0.604024822  0.34354495 0.605958598  0.36304742 1.878715e+00
#>  [46,] 0.494003278  1.40794730 0.702014616  0.48496032 5.905077e-01
#>  [47,] 0.772981566  1.21005921 0.858576135  1.55392047 2.590500e+00
#>  [48,] 0.193213317  0.26252008 0.207663521  0.13397360 9.666638e-01
#>  [49,] 0.419903333  0.19769513 0.258799693  0.24133198 1.178238e-01
#>  [50,] 0.836957217  1.65330636 0.617903984  2.99762801 9.986376e-01
#>  [51,] 0.936789499  1.00496110 0.951870752  1.12532043 2.299798e+00
#>  [52,] 0.447877801  0.43545425 0.331891839  0.21965160 6.952530e-01
#>  [53,] 0.611168843  1.14337401 0.590245727  1.00825832 2.817027e+00
#>  [54,] 0.648807314  7.28669323 0.869722554  3.76420882 2.192038e+01
#>  [55,] 0.602914088  0.67452646 0.483348911  0.36996949 2.847182e-01
#>  [56,] 0.073586791  0.05340566 0.058935213  0.01754975 2.123302e-01
#>  [57,] 0.806435513  6.88979831 1.374911756  2.32115093 1.153677e+01
#>  [58,] 0.838411844  0.92700950 0.443775101  0.42203822 1.082203e+00
#>  [59,] 0.512226726  0.49222226 0.550119159  0.36476976 2.877513e-01
#>  [60,] 0.749197566  0.99980351 1.356096146  2.57021608 2.131092e+00
#>  [61,] 0.403730098  0.36301236 0.256330730  0.15100343 3.652244e-01
#>  [62,] 0.264435706  0.31248452 0.602395575  0.50509535 1.502095e-01
#>  [63,] 0.516071759  1.48790766 0.624873666  0.56064855 1.134074e+00
#>  [64,] 0.171153328  0.27000948 0.136507220  0.15936618 2.746469e-01
#>  [65,] 0.694373080  3.26654186 0.900162613  1.16694738 4.674082e+00
#>  [66,] 0.554410363  4.74166268 1.015192825  0.84704128 9.434604e+00
#>  [67,] 0.677677358  0.37881967 0.517046330  0.44820499 2.286748e-01
#>  [68,] 0.967300741  1.91791271 1.455517916  0.88855942 1.665809e+00
#>  [69,] 0.140700415  0.04715290 0.071018941  0.16241986 2.387297e-02
#>  [70,] 0.732152653 14.69273935 1.306860392 11.63551622 3.277808e+01
#>  [71,] 0.633688275  6.59591769 1.281552202  0.75752634 2.632306e+00
#>  [72,] 0.167263787  0.18085270 0.136173903  0.12567550 3.849004e-01
#>  [73,] 0.351618619  0.21446958 0.264784032  0.29803773 2.960079e-01
#>  [74,] 0.707042386  1.28336743 0.777658462  1.18672858 9.543586e-01
#>  [75,] 0.906927384  2.57078363 1.711656757  0.86363558 2.007126e+00
#>  [76,] 0.608636076  0.50095553 0.512912879  0.20752092 4.382603e-01
#>  [77,] 0.830105738 33.20837524 1.958350776  5.23440023 8.114941e+01
#>  [78,] 0.983910874  1.37979393 0.823767534  1.77168323 2.516068e+00
#>  [79,] 0.302743709  1.80597086 0.338800834  0.97156200 2.695327e+00
#>  [80,] 0.083755080  0.52781973 0.207499209  0.11039458 2.756847e-01
#>  [81,] 0.524788734  0.25211139 0.595033711  0.68641815 3.468578e-01
#>  [82,] 0.910798300  4.81796978 1.259398069  4.81669236 2.742987e+01
#>  [83,] 0.371542904  0.71060196 0.697929238  0.67283092 2.896110e-01
#>  [84,] 0.161699997  0.16293468 0.161145371  0.16221827 1.189128e-01
#>  [85,] 0.611565351  0.44794190 0.567877525  0.39244850 4.334704e-01
#>  [86,] 0.624324274  5.31299623 0.919631745  1.49802758 1.046753e+01
#>  [87,] 0.789553221  0.92774320 0.800024127  0.52376453 4.553551e-01
#>  [88,] 0.507938741  0.96826882 0.494725204  0.84521506 5.363424e-01
#>  [89,] 0.767198596  1.22626471 1.397662919  0.81894138 7.019638e-01
#>  [90,] 0.003334409  0.00659603 0.006553927  0.01152376 7.312155e-03
#>  [91,] 0.067995107  0.88831817 0.150270616  0.25429173 1.581513e+00
#>  [92,] 0.861234014  3.24977124 1.712136720  3.14771700 1.644406e+00
#>  [93,] 0.577540499  0.15399824 0.374242691  0.39443624 1.901765e-01
#>  [94,] 0.026436984  0.02018365 0.043180587  0.03714420 1.011794e-02
#>  [95,] 0.309938789  0.52924810 0.435434476  0.36487084 1.068393e+00
#>  [96,] 0.830663551  0.36589359 0.424238026  0.29274083 1.321038e+00
#>  [97,] 0.916362989  0.97590435 0.817961476  0.35793030 8.650806e-01
#>  [98,] 0.155182897 22.36089959 0.234532972  1.56907845 1.862842e+01
#>  [99,] 0.104637833  0.76846076 0.259314009  0.14918679 9.370451e-01
#> [100,] 0.422024768  0.07627676 0.116515007  0.16650128 8.956265e-02