Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>              [,1]         [,2]        [,3]         [,4]         [,5]
#>   [1,] 0.83131808 2.722260e+00 1.114718143  1.025123063 3.264527e+00
#>   [2,] 0.08208064 7.594938e-03 0.030016596  0.032101124 5.782173e-03
#>   [3,] 0.01623749 6.067312e-02 0.024160365  0.010297962 9.175930e-02
#>   [4,] 0.14628243 4.792839e-01 0.206141192  0.445838640 8.042194e-01
#>   [5,] 0.50993056 3.992672e+00 1.037956267  1.350714807 4.056440e+00
#>   [6,] 0.56206772 7.830468e-01 0.463934845  0.828365036 1.995642e+00
#>   [7,] 0.46304533 9.722803e-01 0.727051548  0.513454424 9.766142e-01
#>   [8,] 0.01235123 1.535188e-03 0.002376922  0.002825144 5.030683e-03
#>   [9,] 0.14963057 3.714898e-02 0.381551103  0.727875840 4.419979e-02
#>  [10,] 0.66073484 3.703317e+00 1.107475599  1.093106253 3.053480e+00
#>  [11,] 0.64288719 6.718690e-01 0.393152383  0.341348896 5.376890e-01
#>  [12,] 0.41543380 1.233170e+01 0.935227993  2.307133536 2.663797e+01
#>  [13,] 0.82451570 7.199584e+00 1.365385612  1.836469140 1.548177e+01
#>  [14,] 0.81115189 4.070655e+00 1.015995534  1.600762653 1.157486e+01
#>  [15,] 0.65791597 1.375128e+01 1.154378785  2.744720951 9.192352e+00
#>  [16,] 0.78112955 5.291699e+00 1.685235158  2.773632583 1.304153e+01
#>  [17,] 0.48453416 9.470772e-01 0.775579337  0.508741625 7.461724e-01
#>  [18,] 0.23460063 2.502445e+00 0.585072013  0.207959369 2.146922e+00
#>  [19,] 0.88398496 3.657157e+00 1.478603580  1.580646078 1.381762e+00
#>  [20,] 0.81565881 2.527386e+00 1.277613267  0.682988078 2.434519e+01
#>  [21,] 0.98848817 2.042208e+00 1.313391952  0.582055482 8.320910e+00
#>  [22,] 0.17760001 1.103255e+00 0.392156528  0.493210496 1.340290e+00
#>  [23,] 0.19917407 1.988740e-01 0.268753466  0.128569053 1.345332e-01
#>  [24,] 0.86138065 4.760445e-01 0.918305707  1.525302415 1.086541e+00
#>  [25,] 0.29517734 1.193773e+00 0.424192028  0.713157716 1.079393e+00
#>  [26,] 0.88868625 6.767238e+00 1.259422130  1.481043881 7.063267e+00
#>  [27,] 0.35251729 5.443166e-01 0.600401690  0.471620068 1.178403e+00
#>  [28,] 0.99792331 1.338553e+00 1.886794050  1.313818863 7.128347e-01
#>  [29,] 0.59033733 2.684530e+00 1.116499086  1.366754058 6.283163e+00
#>  [30,] 0.83165282 2.109662e-01 0.121615700  0.097842363 1.773622e-01
#>  [31,] 0.18172989 5.305708e-01 0.266344727  1.838132705 1.331069e+00
#>  [32,] 0.52217257 4.967794e-01 0.540554391  0.888025291 1.990590e+00
#>  [33,] 0.37023123 1.094047e+00 0.567478406  0.607881675 2.305124e+00
#>  [34,] 0.95289713 1.567495e+00 1.333767619  0.610458765 6.487456e-01
#>  [35,] 0.68235525 1.441637e+00 0.834466061  0.678059394 1.082090e+00
#>  [36,] 0.49844335 7.220344e-01 0.963103985  1.117258591 5.974104e-01
#>  [37,] 0.38478095 6.159127e-02 0.236483813  0.074953863 6.366767e-02
#>  [38,] 0.34758680 3.914523e-01 0.290575829  0.419224941 1.175969e+00
#>  [39,] 0.60899683 3.805405e-01 0.404602736  1.422650408 3.575747e-01
#>  [40,] 0.79439886 2.169674e+00 1.412814581 11.024773387 2.645083e+00
#>  [41,] 0.26809370 2.218408e+00 0.427491115  0.736019122 1.039115e+01
#>  [42,] 0.20488706 2.588470e-01 0.234604598  0.234854902 8.422954e-01
#>  [43,] 0.28507463 7.426026e-01 0.397827277  0.484395433 7.193773e-01
#>  [44,] 0.45086051 1.666157e+00 0.643526771  0.476392457 2.091594e+00
#>  [45,] 0.86755779 6.278522e-01 2.580176393  1.568403091 8.729780e-01
#>  [46,] 0.58497301 1.896709e+02 0.966083686 15.054395587 4.368313e+02
#>  [47,] 0.17772342 1.171516e+00 0.641542924  0.561223188 8.902416e-01
#>  [48,] 0.59166193 1.568536e+01 1.258781576  7.274080619 1.577110e+01
#>  [49,] 0.85047935 8.864464e-01 1.091603061  0.858118773 4.802436e-01
#>  [50,] 0.05627196 1.429038e-01 0.115941663  0.251106507 2.207046e-01
#>  [51,] 0.47069260 3.588825e+00 0.769180040  1.799832053 2.357775e+00
#>  [52,] 0.81616241 2.072019e+01 3.031512934  0.862417758 9.648363e+00
#>  [53,] 0.73016077 7.584449e-01 2.387520518  2.335156386 1.777266e+00
#>  [54,] 0.89614625 3.747439e+00 1.569097867  1.319395727 5.648928e+00
#>  [55,] 0.52286510 4.042095e-01 0.440763793  1.028678446 7.787379e-01
#>  [56,] 0.82794380 7.104438e-01 0.560299031  0.905334712 4.298994e-01
#>  [57,] 0.71995592 3.547465e-01 0.921031227  0.532409949 6.224507e-01
#>  [58,] 0.83422340 2.636799e+00 1.004200433  1.786122858 2.005157e+00
#>  [59,] 0.04093395 2.726120e-01 0.088542763  0.175612978 9.024496e-02
#>  [60,] 0.12980550 1.123071e-01 0.172636482  0.172353241 6.472635e-02
#>  [61,] 0.26585715 2.328361e+00 0.512079628  2.497506284 2.724413e+00
#>  [62,] 0.85447638 3.740998e-01 0.455051712  0.977713366 1.663618e+00
#>  [63,] 0.83868233 2.410121e+00 1.410283790  0.316959372 8.810639e-01
#>  [64,] 0.75944742 1.241935e+00 0.904696188  1.128299693 1.085864e+00
#>  [65,] 0.17806679 1.326174e+00 0.333757491  0.745763268 4.570342e+00
#>  [66,] 0.65388927 1.442405e+00 0.871564196  1.306803581 1.542591e+00
#>  [67,] 0.63085808 1.721205e+00 1.173275764  2.444383390 4.413010e+00
#>  [68,] 0.51969757 1.723574e+00 0.801394802  1.582358258 3.767792e+00
#>  [69,] 0.84471881 2.198294e+00 1.533747317  0.890020744 1.862717e+00
#>  [70,] 0.29934396 3.832281e-01 0.265574682  0.325905614 5.470633e-01
#>  [71,] 0.35276305 3.615214e-01 0.374820899  0.887564708 4.630920e-01
#>  [72,] 0.82991675 2.221879e-01 0.251199018  0.210190589 1.370368e-01
#>  [73,] 0.01970023 1.635192e-01 0.027327968  0.053523302 1.819712e-01
#>  [74,] 0.88088173 5.535952e-01 0.583139317  0.741739721 6.005458e-01
#>  [75,] 0.80390062 7.124015e-01 0.504105271  0.541833980 1.184716e+00
#>  [76,] 0.90381802 9.585846e+00 1.773153422 12.202112723 5.078510e+01
#>  [77,] 0.79082871 3.399850e+00 1.770931076  0.800502905 3.280955e+00
#>  [78,] 0.01134968 3.056029e-01 0.022581717  0.068318516 2.809298e-01
#>  [79,] 0.05860121 2.333018e-01 0.139766851  0.027900595 2.021699e-01
#>  [80,] 0.84816766 1.396008e+00 1.256990998  1.053576723 1.368899e+00
#>  [81,] 0.48503592 1.291429e+00 0.407032704  0.760533002 8.734607e-01
#>  [82,] 0.53378679 2.360321e+01 1.684229752  0.887394157 7.217750e+00
#>  [83,] 0.70566953 9.112894e-01 2.340722139  0.936243361 7.213986e-01
#>  [84,] 0.26549138 1.742006e-01 0.330761018  0.320707777 9.792730e-02
#>  [85,] 0.13135157 3.966316e-01 0.201328764  0.121386514 3.314979e-01
#>  [86,] 0.02991794 1.964802e-01 0.038211645  0.137532197 3.797939e-01
#>  [87,] 0.63802422 4.512236e-01 0.448863827  0.437962210 1.060778e+00
#>  [88,] 0.35772494 1.546741e-01 0.241185219  0.227200737 1.089048e-01
#>  [89,] 0.70974981 9.974767e-01 0.680099949  0.501726622 8.729027e-01
#>  [90,] 0.17867797 2.499573e+00 0.449899580  0.858914681 4.811928e+00
#>  [91,] 0.87217709 4.846576e+00 1.005506230  7.232595274 1.940667e+01
#>  [92,] 0.98690895 4.559821e+01 3.596079827  2.567976381 1.757027e+01
#>  [93,] 0.19151342 5.600543e-01 0.306808371  0.577249039 5.828383e-01
#>  [94,] 0.62600362 1.050984e+00 0.471121767  1.075208775 1.883138e+00
#>  [95,] 0.10469368 6.659875e-02 0.082420083  0.116077046 2.930245e-02
#>  [96,] 0.81668887 3.016481e+00 1.400332205  0.848257515 3.881319e+01
#>  [97,] 0.25290233 7.359204e-01 0.306251121  0.251834186 1.963169e-01
#>  [98,] 0.92683731 3.019314e-01 2.765963508  9.334010792 1.121690e+00
#>  [99,] 0.90312179 1.802214e+00 1.890175055  0.468995401 7.959358e-01
#> [100,] 0.83035350 2.313953e+00 1.012518700  1.452109363 1.536910e+00