Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>              [,1]         [,2]       [,3]        [,4]         [,5]
#>   [1,] 0.16120068   0.62006958 0.36408472  0.34951721 1.339641e+00
#>   [2,] 0.61014416   3.86495518 0.90849817  1.34948058 1.968346e+00
#>   [3,] 0.05936032   0.09182419 0.02976390  0.05218601 1.190690e-01
#>   [4,] 0.87974226   5.92052397 1.70905082  2.19066767 2.568491e+01
#>   [5,] 0.83308839   1.83608181 1.06758189  1.07537467 9.447238e-01
#>   [6,] 0.45915684   0.32505966 0.15970149  0.22453140 4.606418e-01
#>   [7,] 0.30203412   0.16324965 0.29666554  0.35547508 1.801927e-01
#>   [8,] 0.19131700   0.80056186 0.27717403  0.55539649 1.465900e+00
#>   [9,] 0.12308232   1.73199574 0.38009665  0.31368234 1.493842e+00
#>  [10,] 0.57158887   6.61773662 0.98962532  1.20317988 4.493999e+00
#>  [11,] 0.54526555   0.89683527 0.64922903  0.30336822 3.824332e-01
#>  [12,] 0.52862266   0.58871426 1.29975448  1.84500140 9.578951e-01
#>  [13,] 0.57631852   1.67741418 0.92316491  0.22633344 9.026069e-01
#>  [14,] 0.51407896   0.91936936 0.67168965  0.78478241 5.103349e-01
#>  [15,] 0.80644815   5.33529614 1.72853183  3.19917544 5.785708e+00
#>  [16,] 0.18896769   0.22355128 0.12182108  0.16820458 5.436598e-01
#>  [17,] 0.62503300   1.81919966 1.98938138  0.39478704 1.180099e+00
#>  [18,] 0.60252961   5.66983436 1.49248454  0.86242069 2.175970e+00
#>  [19,] 0.06599866   0.16709652 0.13208590  0.07448767 5.260491e-02
#>  [20,] 0.52971535   0.23212368 0.31524891  0.11794637 2.399838e-01
#>  [21,] 0.32221507   3.86413948 0.77621532  1.05548760 3.884044e+00
#>  [22,] 0.46560833  10.98010275 0.73773371  9.59779574 4.262863e+01
#>  [23,] 0.15976417   0.06645942 0.39309870  0.68200808 1.360944e-01
#>  [24,] 0.82887114  20.43271694 1.53183893  2.53778846 4.849710e+01
#>  [25,] 0.73393309  12.82423037 2.27977077  1.50183130 1.678697e+01
#>  [26,] 0.57494853   0.32537087 0.41042757  0.36699379 2.199787e-01
#>  [27,] 0.54210717   0.47923135 0.33248339  0.35746679 1.149666e+00
#>  [28,] 0.80580686   1.83457852 1.21845162  1.49399173 2.674757e+00
#>  [29,] 0.81743487   3.84092542 0.98894170  1.08684150 2.857482e+00
#>  [30,] 0.97880940   1.56418367 1.44635841  1.18493765 1.037170e+00
#>  [31,] 0.76303849   9.98849786 1.64642779  1.94646248 1.164008e+01
#>  [32,] 0.77019181   0.73358745 0.53767434  0.54051999 7.752946e-01
#>  [33,] 0.81382054   0.26503511 0.73596572  0.72952410 1.313430e-01
#>  [34,] 0.39527285   0.64603750 0.38314558  0.52038744 5.368311e+00
#>  [35,] 0.30774698   0.19124413 0.25004952  0.24267679 3.163338e-01
#>  [36,] 0.16202040   0.02500402 0.03851349  0.02894432 1.976882e-02
#>  [37,] 0.98466138   8.52479790 1.96076842 12.22768470 8.480611e+00
#>  [38,] 0.43626194   1.62923578 0.68378924  3.00811544 2.501301e+00
#>  [39,] 0.20147122   0.42697858 0.25397753  0.11960044 2.647499e-01
#>  [40,] 0.95249170   1.30318704 3.51816770  1.52741292 1.713493e+00
#>  [41,] 0.42765817   0.30235710 0.43698548  0.27880891 4.655102e-01
#>  [42,] 0.18813326   0.29671517 0.19024757  0.10369178 2.921223e-01
#>  [43,] 0.11559543   1.42631626 0.38416381  0.23284010 3.561889e-01
#>  [44,] 0.17662792   0.04550468 0.14103320  0.27658731 1.008424e-01
#>  [45,] 0.29797904   1.35292175 0.36260276  0.74963703 5.585996e+00
#>  [46,] 0.01896636   0.17779565 0.02671508  0.03832764 2.930575e-01
#>  [47,] 0.41684745   2.93530698 0.91385871  0.57077499 3.454571e+00
#>  [48,] 0.95797489  17.41331932 1.77127780  4.31235991 2.480974e+01
#>  [49,] 0.84746934  11.36947208 2.31817507 11.68602617 1.085748e+02
#>  [50,] 0.72094912   2.48343152 1.17283082  0.92465817 6.168214e+00
#>  [51,] 0.20190366   0.18227873 0.14312536  0.05550497 2.892383e-01
#>  [52,] 0.20902240   2.26884057 0.37520326  2.28133276 2.689667e+00
#>  [53,] 0.34489963   0.56417550 0.35859035  0.26860718 2.920739e-01
#>  [54,] 0.33786364   0.59033998 0.23496369  0.34373520 1.137709e+00
#>  [55,] 0.78062451   0.41748788 0.49026062  0.67566441 1.614423e+00
#>  [56,] 0.15760413   1.18732631 0.47221692  0.15836186 7.878926e-01
#>  [57,] 0.13440603   0.03594691 0.03757375  0.07257769 2.455610e-02
#>  [58,] 0.72420404   2.04380224 1.03668885  1.32897328 2.069595e+00
#>  [59,] 0.21381000   0.13726525 0.18039345  0.18525873 8.075091e-02
#>  [60,] 0.47716411   0.32007499 0.57253386  0.99879716 6.577355e-01
#>  [61,] 0.80364899   4.72580003 1.88664302  4.23365327 5.512545e+00
#>  [62,] 0.69280784   0.13757607 0.30281030  0.25615830 1.463435e-01
#>  [63,] 0.37184191   0.87327081 0.38926237  0.94100421 1.170098e+00
#>  [64,] 0.39991745   0.35353526 0.39256359  0.23176591 1.341658e-01
#>  [65,] 0.93152486   1.17345469 0.96962897  1.03770747 7.739657e-01
#>  [66,] 0.88058944   0.21692583 0.22107818  0.14429771 4.010723e-01
#>  [67,] 0.42521140   4.88167016 0.82464989  1.54144747 7.974510e+00
#>  [68,] 0.81438512   0.08105398 0.24116956  0.30864254 4.726913e-02
#>  [69,] 0.16413686   0.07275608 0.30390324  0.23635510 8.209855e-02
#>  [70,] 0.57708032   1.51683721 0.65918989  0.47625515 1.315375e+00
#>  [71,] 0.23661532   0.63283143 0.26258304  0.49459735 6.617328e-01
#>  [72,] 0.77335891   0.56449813 0.79203183  1.16796342 8.225974e-01
#>  [73,] 0.75341461   1.09074098 0.92906087  0.85614666 1.466240e+00
#>  [74,] 0.98388735   0.86942108 0.93756674  1.55225426 9.401316e-01
#>  [75,] 0.87506533   1.99850963 1.26509832  1.68971041 4.840663e+00
#>  [76,] 0.64084773   1.51945422 1.05651058  0.99210648 1.784148e+00
#>  [77,] 0.70181485   0.89830167 0.97409683  0.98391683 3.828175e+00
#>  [78,] 0.81665984   1.02217752 1.19536892  1.10502866 1.584903e+00
#>  [79,] 0.88035215   3.34382807 1.60242854  1.08744352 3.143060e+00
#>  [80,] 0.15452374   0.06974164 0.10880154  0.08514245 7.255407e-02
#>  [81,] 0.52295808   1.04602860 1.90108377  0.44246062 9.826502e-01
#>  [82,] 0.02408722   0.03065405 0.02969018  0.02373221 2.273064e-02
#>  [83,] 0.77850307   2.28541649 1.31701578  0.90017971 4.069574e+00
#>  [84,] 0.98601699   0.34565373 0.32645442  0.30351473 3.205167e-01
#>  [85,] 0.08417630   0.01208244 0.02364847  0.01115811 1.464816e-02
#>  [86,] 0.95383838  19.88513861 2.35117725  2.18422945 2.054426e+01
#>  [87,] 0.17109368   0.88294931 0.31648099  0.91045679 1.001482e+00
#>  [88,] 0.59915814   0.82957059 0.62595115  0.42344325 5.074767e-01
#>  [89,] 0.19804871   0.33030700 0.33093345  0.33289639 1.602049e-01
#>  [90,] 0.11509228   1.55559709 0.36959654  0.53871034 1.977199e+00
#>  [91,] 0.04726508   0.01794624 0.04778181  0.11281914 4.810518e-02
#>  [92,] 0.68438300   0.83993418 0.55242424  0.80260803 1.143525e+00
#>  [93,] 0.85966791   1.60395014 1.16873493  1.09476067 3.501155e+00
#>  [94,] 0.63133730   1.20742518 0.71066185  0.40058157 6.955171e-01
#>  [95,] 0.67747574   0.06624967 0.18498592  0.30643117 1.290203e-01
#>  [96,] 0.52747074   4.15197488 1.17121683  0.77071616 4.671600e+00
#>  [97,] 0.12116934   0.01868337 0.02516871  0.03276741 7.076728e-03
#>  [98,] 0.61254325   0.77958165 0.87600547  0.56351736 2.803949e-01
#>  [99,] 0.84639169   7.77614434 0.88153646  6.15696623 1.962265e+01
#> [100,] 0.90412826 163.16530979 2.19018135  0.53267010 1.673255e+01