Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>              [,1]         [,2]        [,3]       [,4]         [,5]
#>   [1,] 0.50418787  0.043077276 0.105151901 0.09642570  0.026388613
#>   [2,] 0.41121221  1.326481023 0.749214774 0.47074279  0.615682804
#>   [3,] 0.22894104  0.087875178 0.151822478 0.46084085  0.220676761
#>   [4,] 0.79383159  5.430084215 3.068725369 1.85929119 10.914158366
#>   [5,] 0.35936716  0.702272541 0.331072968 0.62240625  1.037727232
#>   [6,] 0.48914744  1.241700526 1.297040515 0.94865420  0.323934762
#>   [7,] 0.35026049  0.820025181 0.405112126 0.48815775  1.278700791
#>   [8,] 0.01665783  0.521520162 0.036023272 0.30804809  4.176495407
#>   [9,] 0.18064292  0.370294118 0.188991144 0.18681896  0.305922719
#>  [10,] 0.69965522  4.225884585 1.602853705 3.28670415  4.889492779
#>  [11,] 0.61075029  0.757443255 0.513305078 0.73311209  5.385282791
#>  [12,] 0.25670077  0.296234090 0.349318585 0.20823955  0.944414362
#>  [13,] 0.04419513  0.881561047 0.173394345 0.30343414  0.845637041
#>  [14,] 0.88977115  0.497530118 0.501877078 0.27583021  0.299892223
#>  [15,] 0.15500247  0.039131944 0.103042880 0.13581977  0.071834673
#>  [16,] 0.23375719  0.221802719 0.222267082 0.27445294  0.244037554
#>  [17,] 0.77556960  0.299530345 0.487633485 0.65674990  0.191226957
#>  [18,] 0.86324641  2.214387657 1.347528541 2.90251488  1.167463700
#>  [19,] 0.70630886  1.010823505 0.977767876 0.98468871  0.737611675
#>  [20,] 0.58172807  0.921647262 0.576930257 0.58931569  0.985010207
#>  [21,] 0.13981829  0.042112417 0.092806992 0.14439891  0.041199300
#>  [22,] 0.05458474  0.417022118 0.121193831 0.28444448  0.886437742
#>  [23,] 0.98975932  2.796505153 1.935166330 1.92191523  4.800538507
#>  [24,] 0.62520541  1.226548544 0.740682825 0.92446631  1.481508857
#>  [25,] 0.36735789  0.517384885 0.426225377 0.44441282  2.461605197
#>  [26,] 0.02204176  0.006975209 0.023069513 0.02049393  0.025247328
#>  [27,] 0.65352997  2.240427612 1.568771672 1.21611158  2.760618717
#>  [28,] 0.82233295  1.144395701 1.203397907 1.84145066  1.381770242
#>  [29,] 0.36337839  0.767627861 0.546104703 0.49299009  1.501697404
#>  [30,] 0.39237935  8.036288654 1.047749308 7.97670924  4.170567520
#>  [31,] 0.47810054  4.374683671 1.011360819 1.99568406 15.020895651
#>  [32,] 0.85130624  0.062360346 0.143759947 0.18269983  0.051913386
#>  [33,] 0.57977092  1.211418849 0.698257947 2.43629302  1.463121543
#>  [34,] 0.22052981  0.125674634 0.208275319 0.30863410  0.164730175
#>  [35,] 0.17292226  0.691694098 0.303074750 0.46516265  0.542223240
#>  [36,] 0.96819204  0.766021773 1.157598879 0.89532164  0.490050975
#>  [37,] 0.42259422  0.705535697 0.466926090 0.85731626  1.079543728
#>  [38,] 0.70016982  0.393795686 0.461320027 0.32949234  0.520508784
#>  [39,] 0.86265333  3.536589775 1.619991283 4.01654320  7.040220981
#>  [40,] 0.58305712  1.371761064 0.531790000 0.68531078 17.385877369
#>  [41,] 0.45282855  0.651804303 0.496938160 0.55995066  0.246674025
#>  [42,] 0.84419084  3.992590787 1.435951379 3.00860600  2.887246711
#>  [43,] 0.92611177  2.164730513 1.353872892 1.91398371  1.745666416
#>  [44,] 0.32516426  0.092388975 0.218046077 0.29747180  0.202755130
#>  [45,] 0.88662355  4.860216333 1.524510783 1.89820843  4.665192125
#>  [46,] 0.92156387  1.073594323 0.654637311 1.02329115  1.829194751
#>  [47,] 0.65999934  1.585207711 0.762957099 0.82009599  6.488267904
#>  [48,] 0.91194793  6.355333242 3.026057296 1.99359354 23.125405086
#>  [49,] 0.71924842  1.740565597 0.754287159 0.66725551  1.626296454
#>  [50,] 0.29377058  0.178719793 0.360461597 0.27494624  0.216780293
#>  [51,] 0.79494682  1.267241261 0.579391040 0.91704684  2.147227213
#>  [52,] 0.63191701  7.299049982 1.823790707 5.86202095  4.176035801
#>  [53,] 0.99852662  4.167235330 1.252849111 2.95673248  8.978535757
#>  [54,] 0.98184831 15.256797155 3.366120319 1.23608177  3.867157995
#>  [55,] 0.16932879  0.639962162 0.291856906 0.21434234  0.396489891
#>  [56,] 0.23541682  0.201995833 0.391650692 0.09375369  0.197926287
#>  [57,] 0.99810255  1.081098821 0.959021862 0.31550176  1.568500252
#>  [58,] 0.40551343  1.370368148 0.631276519 1.17240781  4.561778254
#>  [59,] 0.68709199  2.019970404 0.721150073 0.72732890  1.113065794
#>  [60,] 0.53627150  2.776949381 0.751570528 1.31039601  2.673042114
#>  [61,] 0.54539283  1.535287826 0.704362226 1.56215427  4.151048167
#>  [62,] 0.68669950  1.210137174 0.723075563 0.88208053  0.858956703
#>  [63,] 0.14889664  0.253054058 0.105337606 0.08608637  0.741594087
#>  [64,] 0.44533622  1.218386935 0.598003934 0.38024555  1.060183697
#>  [65,] 0.01656475  0.007401306 0.008652629 0.00803951  0.002259741
#>  [66,] 0.74299055  2.630948511 1.563118507 1.49780793  2.019941141
#>  [67,] 0.65311521  0.363880854 0.487999668 0.60647504  0.235831402
#>  [68,] 0.25029869  0.458263300 0.232739293 0.63390680  0.608608078
#>  [69,] 0.15924590  0.098301828 0.077338315 0.06365060  0.083384201
#>  [70,] 0.23776094  0.302938663 0.183164503 0.29086620  0.538836147
#>  [71,] 0.65850321 12.109798978 2.142667423 3.63714915 16.141469045
#>  [72,] 0.70015097  0.538361746 0.853246749 0.49914860  1.035271633
#>  [73,] 0.04797433  0.023373850 0.034941406 0.05580513  0.074621924
#>  [74,] 0.77228552  4.149237687 1.695096714 1.19027440  2.341974856
#>  [75,] 0.24104623  0.240103220 0.140646262 0.07042628  0.115611686
#>  [76,] 0.47551023  0.531301086 0.319252163 0.22932273  0.478384381
#>  [77,] 0.61427693  3.163403759 2.127939116 0.68402687  1.918416446
#>  [78,] 0.03555717  0.123695501 0.073110938 0.06300400  0.174494763
#>  [79,] 0.63975739  1.124092080 0.585479219 0.47212869  2.021628392
#>  [80,] 0.31754030  1.052105435 0.849078097 0.25917482  1.534602291
#>  [81,] 0.11807607  0.105463250 0.122127305 0.11601265  0.384809369
#>  [82,] 0.22284443  0.936795213 0.485469469 0.31771818  1.347189974
#>  [83,] 0.91358906  0.393869157 0.405575430 0.52278475  1.598197955
#>  [84,] 0.05650556  0.111241686 0.100752468 0.03547465  0.327919781
#>  [85,] 0.97053242  2.758989732 1.866511969 0.87276249  4.648030787
#>  [86,] 0.07892690  0.618704337 0.175951305 0.26693012  2.026024979
#>  [87,] 0.39466175  0.296958575 0.635941143 0.42529509  0.808445341
#>  [88,] 0.82573830  3.742799961 1.064520785 2.04981592 11.400498985
#>  [89,] 0.04998506  0.026116418 0.020757772 0.01488358  0.025003456
#>  [90,] 0.94201580  5.447778223 1.988542803 1.79224481  3.427055865
#>  [91,] 0.82680439  0.774242541 0.771452872 1.30046992  1.706143737
#>  [92,] 0.35050048  0.565878723 0.302556562 0.47176522  0.482970151
#>  [93,] 0.56939975  0.466029238 0.326117845 0.24428805  0.451775194
#>  [94,] 0.94852535  9.166365584 2.802537928 1.36833703  4.226499174
#>  [95,] 0.24232395  0.114056623 0.168738260 0.19696270  0.109659363
#>  [96,] 0.76253650  1.427285327 0.719534848 1.17551666  1.973082477
#>  [97,] 0.99590906  0.278640778 0.508108698 1.64227046  0.659501750
#>  [98,] 0.38370586  1.312678917 0.785707025 1.22859205  2.203326437
#>  [99,] 0.19860146  0.028795331 0.192925532 0.21548203  0.038881660
#> [100,] 0.20184908  0.074737170 0.134618338 0.13052465  0.125137109