Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>              [,1]        [,2]       [,3]        [,4]        [,5]
#>   [1,] 0.94359508  0.19378777 0.56726018 0.451077696  0.20109913
#>   [2,] 0.93638721  0.49444025 0.61239912 0.523084072  0.33861248
#>   [3,] 0.21844805  0.97565066 0.38044166 1.605733622  2.65101425
#>   [4,] 0.05368853  0.13663763 0.08797745 0.087375471  0.09954170
#>   [5,] 0.32663908  0.84842574 0.89995189 0.217531307  0.46877670
#>   [6,] 0.95459460  7.51978069 1.83327380 3.488761187  8.87015502
#>   [7,] 0.91657437  3.33652743 1.19811355 1.113173552  3.53730234
#>   [8,] 0.46353888  0.70418211 0.62695675 0.675055135  1.13504729
#>   [9,] 0.80349515  0.33055169 0.43281653 0.588557742  0.35366639
#>  [10,] 0.67511771  0.70425535 0.59288862 1.039299194  2.00309969
#>  [11,] 0.06853207  1.52327075 0.16198474 0.416216668  1.30673234
#>  [12,] 0.22076606  0.70657841 0.54315691 0.305926139  0.40042435
#>  [13,] 0.48438076  0.97792976 0.68433465 0.648362380  3.71063420
#>  [14,] 0.01918574  0.02614045 0.03689798 0.016170144  0.01886721
#>  [15,] 0.08901212  0.92603621 0.18607491 0.376439118  1.59108645
#>  [16,] 0.39718965  6.91376384 1.19724327 1.534553433  7.82207253
#>  [17,] 0.31755740  2.87110112 0.62222231 0.534058837  1.11236391
#>  [18,] 0.59218985  0.35972677 1.19899818 0.696386488  0.44279621
#>  [19,] 0.82155761  1.29135236 0.81562538 1.091601767  1.70915720
#>  [20,] 0.56604691  1.43046031 0.43131527 0.309124911 21.90599913
#>  [21,] 0.49664179  2.00350844 0.58711229 0.842594921  4.80825971
#>  [22,] 0.45450943  0.41846562 0.91148028 0.896605736  0.44813443
#>  [23,] 0.06560320  0.05066143 0.06036602 0.056076493  0.04691553
#>  [24,] 0.22828587  0.07185706 0.05817002 0.103090707  0.27026661
#>  [25,] 0.12637862  0.12041528 0.09017577 0.247451613  0.63290339
#>  [26,] 0.17531806  5.26472373 0.28335866 2.579153809 11.03791809
#>  [27,] 0.44335716  7.98014976 0.86657454 1.470728794  6.61359203
#>  [28,] 0.14770475  0.15455315 0.10751829 0.076712781  0.35465872
#>  [29,] 0.77194553 12.69811743 1.95162423 3.140026849 17.42899889
#>  [30,] 0.97627114  1.36681259 1.16620758 0.445644866  0.47861372
#>  [31,] 0.01948851  0.13201842 0.03483667 0.089493593  0.12420273
#>  [32,] 0.82686675  0.78033245 0.71451296 0.888074923  1.05527204
#>  [33,] 0.30854663  0.18650696 0.12334587 0.204443187  0.14055968
#>  [34,] 0.92047767  6.72973048 1.39774380 3.959900997  2.16780182
#>  [35,] 0.83801109 13.94724212 1.40496253 2.122092912  8.65452787
#>  [36,] 0.89946256  0.28106815 0.43310073 0.376610850  0.09783677
#>  [37,] 0.46735549  0.56076656 0.36439976 0.246422060  0.41569715
#>  [38,] 0.35867069  0.31325040 0.26580879 0.125205988  0.14102973
#>  [39,] 0.97535989  2.22543803 1.54404761 1.639109111  2.85014890
#>  [40,] 0.32386149  0.10804939 0.28107055 0.309348716  0.03165254
#>  [41,] 0.82973222  2.10936276 0.83275597 2.325466591  6.38545163
#>  [42,] 0.09250707  0.35457845 0.09989084 0.208926402  0.80922463
#>  [43,] 0.89580625  2.49579216 1.41145431 0.888883496  8.80288791
#>  [44,] 0.34508040  5.53384822 1.38135302 1.094883638  1.75170131
#>  [45,] 0.03250410  0.03146141 0.05531419 0.074976578  0.03433826
#>  [46,] 0.28390129  2.70975860 0.64272283 1.709835263  3.98574539
#>  [47,] 0.37921162  2.26169058 0.90076228 0.369369416  2.28239994
#>  [48,] 0.63717479  0.94186914 0.67578004 0.875331896  0.82338367
#>  [49,] 0.96576511  0.14282584 0.15973045 0.156423397  0.27121805
#>  [50,] 0.27261251  0.79448585 0.31102531 0.168006969  0.77429666
#>  [51,] 0.26327403  0.48157621 0.34458495 0.238912736  0.49245669
#>  [52,] 0.44761152  0.02637906 0.04187537 0.032039493  0.03413857
#>  [53,] 0.47880609  0.21891599 0.25210675 0.322297630  0.22407768
#>  [54,] 0.37714938  1.52923708 1.21910387 0.631489596  0.49426277
#>  [55,] 0.15729032  0.41023011 0.81573513 0.698095351  0.21801372
#>  [56,] 0.75025088  2.13526788 1.86351152 2.612787467  5.32861358
#>  [57,] 0.99585571  0.49653812 1.30783844 0.741525527  0.29782849
#>  [58,] 0.72372124  8.49328026 1.68501329 3.999470914  8.27380332
#>  [59,] 0.05725038  0.06395270 0.08363648 0.090428779  0.03291395
#>  [60,] 0.41317905  0.67007890 0.39679260 0.291724003  3.05262772
#>  [61,] 0.88967899 18.56822887 3.06363828 1.562588903  3.12788040
#>  [62,] 0.08040233  0.37827698 0.16977733 0.092126898  0.11205601
#>  [63,] 0.35098854  1.30206446 0.58049367 0.456653875  0.37416909
#>  [64,] 0.63818679  0.77414824 0.65838539 1.085551088  0.98180661
#>  [65,] 0.25562064  2.00273646 0.39507418 0.848632458  3.45302552
#>  [66,] 0.34144767  0.14688009 0.23109915 0.164869395  0.03255749
#>  [67,] 0.96884554  0.96600040 1.04581009 2.350239156  1.69358501
#>  [68,] 0.38912277  2.40757292 0.79987108 1.120446406  1.34680154
#>  [69,] 0.78306328  4.49946809 0.98211715 0.591452167 13.32938704
#>  [70,] 0.10512067  0.38091515 0.71330456 0.106854264  0.34311171
#>  [71,] 0.69915620  6.93957534 1.17961516 6.098529351 17.40153691
#>  [72,] 0.06577456  0.33469501 0.07795909 0.198486461  0.57178981
#>  [73,] 0.83123771  2.36768657 1.15078854 0.671056545  0.93280935
#>  [74,] 0.76127537  7.83910941 1.56876175 1.630485308 10.08013595
#>  [75,] 0.17649434  0.35728414 0.17209360 0.400206061  1.18291043
#>  [76,] 0.22921037  0.46321467 0.44656686 0.634509594  0.57490020
#>  [77,] 0.82700545  2.27538512 1.31172433 1.112642237  0.75503262
#>  [78,] 0.09630719  0.04810554 0.06892008 0.103272487  0.06534421
#>  [79,] 0.82542911  2.00323181 1.03658544 2.115847222  2.60223448
#>  [80,] 0.61452615  0.39214451 0.65467295 0.599979591  0.57977369
#>  [81,] 0.12998647  0.15758823 0.07468281 0.096625015  0.50543514
#>  [82,] 0.63334946  2.63640353 2.05660721 0.801810356  1.93556297
#>  [83,] 0.05717050  0.20778442 0.11273047 0.145849320  0.03653126
#>  [84,] 0.37483364  1.53969045 0.55633050 0.343263424  3.36107800
#>  [85,] 0.15066043  0.18514600 0.09312492 0.103097568  0.15000810
#>  [86,] 0.01106311  0.04257188 0.02245124 0.009374821  0.01553867
#>  [87,] 0.43522359  5.62830147 0.84216065 2.921411007 19.66798500
#>  [88,] 0.78047878  0.41786900 0.46810200 0.727053590  0.57215332
#>  [89,] 0.44179671  0.22804871 0.29014044 0.533759625  0.26435681
#>  [90,] 0.02376452  0.04303698 0.02963884 0.062918896  0.08390163
#>  [91,] 0.83135503  0.41960885 2.80449800 4.442695895  0.36905883
#>  [92,] 0.11769345  0.23981252 0.70088588 1.355129850  1.20516656
#>  [93,] 0.12990632  0.09277828 0.23887826 0.240395951  0.05380863
#>  [94,] 0.85063423  1.27924958 0.67393800 2.320171797  5.04811044
#>  [95,] 0.39740721  1.72891082 0.72835964 0.433827529  1.66178198
#>  [96,] 0.89844266  0.49022865 0.66620638 0.407722893  0.76924931
#>  [97,] 0.71950254  4.61927831 1.62666324 2.822039867  2.37809427
#>  [98,] 0.33511766  0.75053206 0.27771940 0.230566455  2.13281727
#>  [99,] 0.78130521  0.09084084 0.16552354 0.213976777  0.13342506
#> [100,] 0.83097045  0.63722551 0.71249722 0.920990032  1.14434878