Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>               [,1]         [,2]       [,3]       [,4]         [,5]
#>   [1,] 0.683879916  0.851528193 0.26263677 0.19819755   0.32717530
#>   [2,] 0.688499785  1.470543442 0.42690441 0.57483276   0.54031662
#>   [3,] 0.122926442  0.094896605 0.12516385 0.10848563   0.19601632
#>   [4,] 0.252618012  0.176951831 1.32374583 0.67963149   0.23112566
#>   [5,] 0.319804740  0.217631160 0.18293071 0.16011216   0.31789828
#>   [6,] 0.615591525  7.257750048 2.02864191 0.33926414   1.56529655
#>   [7,] 0.476509053  0.717246783 0.83034568 0.80214101   0.40550321
#>   [8,] 0.716099792  5.030662342 1.06011728 0.95172553   4.35313227
#>   [9,] 0.499135924  2.295198319 0.72950476 0.70277169   1.44377010
#>  [10,] 0.444490548  3.646603616 2.44830632 0.45278642   3.02381726
#>  [11,] 0.577608940  5.039342974 1.09296723 0.73436850   2.28787301
#>  [12,] 0.669221171  4.184532202 1.15594977 1.24748606   2.83593664
#>  [13,] 0.401894399  0.553500016 1.87645153 0.52066926   0.63538454
#>  [14,] 0.718478477  4.139122327 1.68820711 0.83278799   7.70254970
#>  [15,] 0.529662495 18.168323802 1.94431837 2.45247882   9.67852856
#>  [16,] 0.369945043  0.124788247 0.10104256 0.13508240   0.55435681
#>  [17,] 0.145419932  0.782821170 0.34410454 0.37054690   0.48979120
#>  [18,] 0.596590879 32.716910758 2.44518751 0.56162653   6.46108563
#>  [19,] 0.545072485  0.891144973 0.62256247 0.49114364   2.45993097
#>  [20,] 0.103721157  0.145157555 0.08037034 0.19815064   0.18730290
#>  [21,] 0.377571179  1.470206389 1.13241968 0.53308934   3.13628360
#>  [22,] 0.009757345  0.022046886 0.01066709 0.01323637   0.03199119
#>  [23,] 0.725639940  0.390765427 0.92242932 0.53080576   0.14685431
#>  [24,] 0.549375883  3.102833359 0.87304650 3.31541858   2.38804901
#>  [25,] 0.199012785  1.997089567 0.66061927 0.27188098   0.28146931
#>  [26,] 0.632554559  1.185090128 0.70322352 0.85503232   2.37351876
#>  [27,] 0.187886154  0.787366779 0.26864348 0.93546519   2.61006160
#>  [28,] 0.188895515  0.716922537 0.39979811 0.31686930   1.78983430
#>  [29,] 0.365346593  1.614429925 0.78578727 0.40320197   0.88801367
#>  [30,] 0.545521530  0.376993186 0.23078381 0.13269516   0.30185029
#>  [31,] 0.699531875  0.742812775 0.64774335 0.64498840   3.24749052
#>  [32,] 0.744044313  0.844144602 0.63994257 0.80130505   0.59212871
#>  [33,] 0.098504595  0.415807711 0.15930598 0.47913870   1.21055837
#>  [34,] 0.949711844  2.439733339 0.54402776 0.30494716   1.62669993
#>  [35,] 0.736525137  0.065933858 0.14042216 0.13521891   0.04975196
#>  [36,] 0.682107698  5.493063972 0.98243023 2.52073203   7.19119600
#>  [37,] 0.954500227  1.790152259 1.52040142 1.76168539   1.51912819
#>  [38,] 0.442053512  2.415805987 0.60921484 0.85736645   1.59301581
#>  [39,] 0.846247721  9.479590732 1.30918703 8.84915008  16.33136281
#>  [40,] 0.797238171  3.273264301 1.46351997 0.83008208   0.80607833
#>  [41,] 0.197055341  2.485186030 0.37104074 1.97796441   4.08639370
#>  [42,] 0.501183293 31.532081036 1.28449048 2.04031660  13.92461733
#>  [43,] 0.844379784  0.453781646 0.86990609 2.84936747   1.22138682
#>  [44,] 0.530437584  1.821244275 1.93823712 0.49598636   2.91872781
#>  [45,] 0.529057568  2.035852390 0.63133961 0.25406961   2.22856558
#>  [46,] 0.582057948 21.799168121 1.65113125 1.35312664  32.64117009
#>  [47,] 0.168432053  0.363755176 0.20611354 0.99836291   2.67332978
#>  [48,] 0.185354911  0.103337197 0.06581966 0.15180429   0.14881140
#>  [49,] 0.348008022  0.938175110 0.80368579 0.30848880   0.73191906
#>  [50,] 0.172003049  0.055630702 0.04879595 0.05945446   0.04391343
#>  [51,] 0.205352153  1.224226846 0.29266592 0.23243631   0.42077248
#>  [52,] 0.360079312  0.367821627 0.37708017 0.68271199   0.37825957
#>  [53,] 0.910246127  4.222904893 1.42976571 5.32172733   7.46257645
#>  [54,] 0.693114429  2.964896267 1.13832714 1.43570028   1.63730330
#>  [55,] 0.646907151  0.263896166 0.44738950 0.46895426   0.17929663
#>  [56,] 0.670571876  0.281142645 1.05367032 1.13539755   0.11867435
#>  [57,] 0.064855737  0.201299064 0.11028005 0.15753582   0.22549086
#>  [58,] 0.177789221  0.462116321 0.23227307 0.35088844   0.47196691
#>  [59,] 0.351683224  0.476098928 0.48229572 0.26014501   0.22350403
#>  [60,] 0.195291498  0.050588902 0.22922038 0.27997410   0.27534858
#>  [61,] 0.904527771  1.641554868 3.34363013 2.50254984   4.27344926
#>  [62,] 0.171555251  4.803347455 0.48806434 0.70019308   5.78718480
#>  [63,] 0.213944558  0.467662608 0.32242474 0.14086420   0.43031935
#>  [64,] 0.138583643  0.041927948 0.16245472 0.07831125   0.03485623
#>  [65,] 0.589277975  2.423316927 3.41129060 0.53857184   1.97450911
#>  [66,] 0.917906411  0.454826646 0.50418281 0.43310803   1.12197796
#>  [67,] 0.639787875  0.330058418 0.35532298 0.45062041   1.36067876
#>  [68,] 0.298549312  0.178624481 0.27848363 0.12640819   0.11768290
#>  [69,] 0.568181339  0.511913335 0.39381389 0.46288479   0.39976907
#>  [70,] 0.169221850  0.351289918 0.28195869 0.35801824   0.20140533
#>  [71,] 0.649135224  3.524821781 1.86346304 0.52824368   2.81892079
#>  [72,] 0.269470473  0.641931688 1.04017689 0.86091314   0.55270210
#>  [73,] 0.150702036  0.712442712 0.33689297 0.64249546   1.74057087
#>  [74,] 0.631460623  7.064788240 1.30715004 3.30678015  18.15748675
#>  [75,] 0.517497512  2.125772068 0.77720288 1.60396142   4.19926349
#>  [76,] 0.174093201  0.114063659 0.11796022 0.07417540   0.05015649
#>  [77,] 0.327233541  3.739573805 0.73488711 2.53948841  37.56249643
#>  [78,] 0.391466120  4.520193097 0.94611792 3.07728958   5.22579194
#>  [79,] 0.125364537  0.618344092 0.22447706 0.20738785   0.56208251
#>  [80,] 0.696591670  7.084913700 1.13777616 4.08659681  38.90294271
#>  [81,] 0.499529793  0.430281259 0.56776126 0.24704225   0.19179532
#>  [82,] 0.819979363 11.516816272 1.82730949 4.07729259  40.02816786
#>  [83,] 0.793968245  2.029783349 0.95997683 0.73722991   3.29990382
#>  [84,] 0.153006594  0.243031821 0.15189266 0.16899270   0.30930299
#>  [85,] 0.660108573  1.353477018 1.29157447 0.48850974   1.24607201
#>  [86,] 0.485761831  0.366184189 0.41098994 0.34468068   0.28371502
#>  [87,] 0.456536655  2.008363273 0.68535318 0.35527135   2.64374263
#>  [88,] 0.298657154  3.847881285 0.87724565 0.31797387   2.60542654
#>  [89,] 0.307811825  0.399105077 0.53186138 0.98141504   0.75872099
#>  [90,] 0.660460262  2.047059208 1.79493327 1.96579620   1.43250195
#>  [91,] 0.347834481  0.126027329 0.17837662 0.10796250   0.03951240
#>  [92,] 0.785843760 49.826564626 2.19985685 3.58906978 103.35935491
#>  [93,] 0.171654808  0.549333088 0.21005718 0.37620307   0.49521737
#>  [94,] 0.035545886  0.008096066 0.03934770 0.05933931   0.03222667
#>  [95,] 0.518241285  0.262830569 0.30555790 0.30296474   0.17675679
#>  [96,] 0.169240076  0.166787988 0.16397129 0.38303980   0.18566785
#>  [97,] 0.904175753  1.345557819 2.75121524 1.86160125   5.17988786
#>  [98,] 0.635113767  2.226796790 0.75900070 0.89291217   3.51992319
#>  [99,] 0.211071156  0.106845830 0.26765515 0.21743762   0.13941326
#> [100,] 0.039631331  0.078304289 0.06807842 0.08922255   0.03989234