Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>               [,1]         [,2]        [,3]        [,4]         [,5]
#>   [1,] 0.819466807 22.693495846 2.705671194 7.136300835 6.732027e+00
#>   [2,] 0.012515525  0.178009647 0.026011398 0.022505639 7.442673e-02
#>   [3,] 0.471645116  1.122451625 0.534265518 0.562502433 2.301333e+00
#>   [4,] 0.402464766  0.596893634 0.679883663 0.316490495 7.426935e-01
#>   [5,] 0.063665505  0.202528845 0.126649733 0.070649387 8.262764e-01
#>   [6,] 0.337042396  1.582382151 0.511646646 0.778291993 2.029723e+00
#>   [7,] 0.630420525  8.408675747 1.523695551 1.702553156 5.412247e+00
#>   [8,] 0.743281163  0.585833271 0.605642283 0.595474484 2.378956e-01
#>   [9,] 0.921522405  4.240031640 2.172597199 2.898180924 5.138789e+00
#>  [10,] 0.008826341  0.006913145 0.006140025 0.005394606 1.179562e-02
#>  [11,] 0.771217261  3.054933195 1.432111424 1.982942065 1.085063e+01
#>  [12,] 0.720236169  2.138709797 1.231537983 0.155129764 4.003586e-01
#>  [13,] 0.344854887  0.152140380 0.121017338 0.128338184 1.430765e-01
#>  [14,] 0.399927981  0.345462200 0.506866457 0.531255450 3.614721e-01
#>  [15,] 0.539984379  0.137537804 0.109462661 0.139995687 3.888451e-01
#>  [16,] 0.360326978  0.193424058 1.289575876 0.246283192 1.588677e-01
#>  [17,] 0.479145916  0.737365089 0.516969333 0.754742064 4.633521e-01
#>  [18,] 0.832949081  4.016548077 1.476520194 1.262457883 4.736994e+00
#>  [19,] 0.328804188  1.512860415 0.433705365 0.475721567 2.226466e+00
#>  [20,] 0.767912899  1.191272339 1.481185537 1.703294880 2.750702e+00
#>  [21,] 0.679647796  0.930122297 0.481824922 1.168785512 1.772731e+00
#>  [22,] 0.943626442  2.112255111 1.491242344 3.084350313 7.641295e+00
#>  [23,] 0.771044939  1.156535187 1.186634262 0.794793816 8.585073e-01
#>  [24,] 0.073819510  0.028631975 0.051982659 0.090582941 1.452737e-02
#>  [25,] 0.307345615  0.317823338 0.256529933 0.250583286 1.887925e-01
#>  [26,] 0.580436137  0.627383092 0.296349191 0.437670951 1.650751e+00
#>  [27,] 0.219671458  0.867511283 0.399634143 0.304315699 1.377433e+00
#>  [28,] 0.709031457  5.037024083 1.238928714 4.450088799 6.615581e+00
#>  [29,] 0.339294022  0.113664959 0.791275304 1.855740405 1.599498e+00
#>  [30,] 0.723453911  2.977560662 1.046977432 1.549738047 4.327180e+00
#>  [31,] 0.480640231  0.288293355 0.345681725 0.163293987 7.463299e-01
#>  [32,] 0.893182368  2.538677711 0.939515714 0.880685981 1.295677e+00
#>  [33,] 0.659928685  7.682151725 1.381582112 1.163018486 3.250924e+00
#>  [34,] 0.528031781  0.350539390 0.474484444 1.344389046 1.626792e+00
#>  [35,] 0.618083814  5.037646456 1.670805962 3.413862535 1.973944e+01
#>  [36,] 0.305960610  1.186417142 0.357440233 0.654119539 1.928337e+00
#>  [37,] 0.009574595  0.007449398 0.007050519 0.011046851 9.953455e-03
#>  [38,] 0.123159303  0.016034151 0.031178810 0.090690174 9.092907e-02
#>  [39,] 0.157222615  0.006689909 0.039693516 0.061833499 1.288093e-02
#>  [40,] 0.615254011  8.771923726 1.440930456 5.986393043 8.941412e+00
#>  [41,] 0.291457439  0.175120806 0.132497337 0.111380048 4.919402e-01
#>  [42,] 0.898417059  3.094296772 1.267151519 1.036531459 2.146323e+00
#>  [43,] 0.445008150  0.831157674 0.264740055 0.691098657 2.371159e+00
#>  [44,] 0.287854437  2.914917738 0.449132548 3.384564949 9.141195e+00
#>  [45,] 0.542381491  0.793487638 0.365185106 1.467887555 1.273016e+00
#>  [46,] 0.964182665  6.761112696 1.342702395 3.794708934 1.768753e+01
#>  [47,] 0.567885123  2.156851032 0.573101364 1.302303113 8.161481e+00
#>  [48,] 0.298817051  0.479593983 0.275855732 0.557821100 1.132369e+00
#>  [49,] 0.353501244  0.988672086 0.429356509 0.559496458 1.484407e+00
#>  [50,] 0.665329115  6.346414838 1.502336983 0.823283010 2.571507e+00
#>  [51,] 0.591020421  0.814731639 0.572513861 0.573474366 1.965458e-01
#>  [52,] 0.337130040  0.142738896 0.220895817 0.263541577 8.951587e-02
#>  [53,] 0.257921776  0.670145495 0.311507056 0.237241096 1.625450e+00
#>  [54,] 0.256380437  0.271989363 0.197275138 0.376993259 4.058476e-01
#>  [55,] 0.394501965  1.199949462 0.512363862 0.356998890 3.638596e+00
#>  [56,] 0.707095121  0.278680553 0.710707904 0.732379544 2.323092e-01
#>  [57,] 0.725768862  0.290381572 0.567948584 0.932372796 5.673136e-01
#>  [58,] 0.567634551  4.274319152 1.474781349 0.485145713 5.227685e+00
#>  [59,] 0.775850477  0.159312736 0.523697382 0.427182068 3.470014e-01
#>  [60,] 0.334062750  2.571156710 0.673080976 2.341089581 4.625978e+00
#>  [61,] 0.644603366  2.335320667 1.554037342 1.608693086 1.486632e+00
#>  [62,] 0.632040046  0.727945307 0.836646258 0.364213014 1.768046e+00
#>  [63,] 0.436726516  0.794111610 0.461962912 0.340447315 8.442503e-01
#>  [64,] 0.968532786  2.194459413 0.917733875 1.369960699 2.956877e+00
#>  [65,] 0.841693802  0.773072101 0.701411010 1.084436157 1.200410e+00
#>  [66,] 0.975353249  0.999509046 0.845807653 0.456523953 1.251766e+00
#>  [67,] 0.896633190  9.268136561 1.389211735 3.445138685 1.145903e+02
#>  [68,] 0.184051449  0.306740407 0.311688573 0.359237651 2.711235e-01
#>  [69,] 0.780282026  3.263258634 1.209087506 3.219185950 5.153808e+00
#>  [70,] 0.512464319  3.859918800 0.951143411 0.846486978 2.928704e+00
#>  [71,] 0.641875057  0.301557392 0.797494034 1.221126024 2.802197e-01
#>  [72,] 0.103995776  0.320954571 0.168182425 0.258496283 5.402768e-01
#>  [73,] 0.598733948  0.741402108 0.408467744 0.370466870 7.304416e-01
#>  [74,] 0.746421892  1.315991150 0.466938259 0.730116485 6.928596e-01
#>  [75,] 0.160329291  0.163441461 0.174272238 0.200192013 5.126084e-02
#>  [76,] 0.942961524 13.194124546 2.443309744 4.374803087 4.261513e+00
#>  [77,] 0.666712905  0.788256895 0.786636289 0.656861380 8.287803e-01
#>  [78,] 0.369176797  0.305559743 0.250573272 0.227636804 4.312565e-01
#>  [79,] 0.340466708  3.866201181 0.521297762 2.629894784 2.196953e+01
#>  [80,] 0.479984313  0.188529287 0.295327859 0.254447113 3.233472e-01
#>  [81,] 0.270724965  1.878355787 0.474477591 1.197236381 3.665304e+00
#>  [82,] 0.589242815  0.550370408 0.488634034 0.613919767 8.298482e-01
#>  [83,] 0.712040994  3.470793780 1.748347121 1.338336561 1.395261e+00
#>  [84,] 0.924417532  3.155192568 1.632849109 3.368230264 9.510238e+00
#>  [85,] 0.251323221  0.376121777 0.270221913 0.390830786 6.549786e-01
#>  [86,] 0.963224562  1.133704316 1.175881296 0.720633543 3.127542e+00
#>  [87,] 0.931984884  1.099657159 0.884470370 0.650970120 1.313341e+00
#>  [88,] 0.989732213  6.027432102 1.286664344 1.423908897 3.790662e+00
#>  [89,] 0.113386538  0.120840068 0.120508628 0.052151823 6.924174e-02
#>  [90,] 0.106867622  0.253027306 0.093820564 0.131624121 4.103824e-01
#>  [91,] 0.746662087  3.679024869 1.624759999 1.016830983 4.566840e+00
#>  [92,] 0.281111545  2.413393607 0.434773902 3.103384013 3.561759e+00
#>  [93,] 0.775631895  0.768501046 0.633736148 0.753821580 8.864795e-01
#>  [94,] 0.249649544  0.262520319 0.243157703 0.131326919 5.676900e-01
#>  [95,] 0.112198380  0.060835064 0.138122104 0.115192245 4.371226e-02
#>  [96,] 0.564334280  3.878433512 1.368545073 0.667997156 8.644530e-01
#>  [97,] 0.848631339  1.617145591 1.643209404 1.266397160 2.412912e+00
#>  [98,] 0.454496408  0.814403410 0.630381878 0.332713667 6.093934e-01
#>  [99,] 0.463784639  2.145415319 0.832636524 0.954252814 1.497850e+00
#> [100,] 0.024528362  0.049644998 0.135546077 0.056010278 5.861277e-02