Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>              [,1]         [,2]       [,3]       [,4]         [,5]
#>   [1,] 0.43901962  1.951962058 0.58141420 1.27763604 3.545917e+00
#>   [2,] 0.84124070  3.485502922 1.48700688 6.85424232 2.550659e+00
#>   [3,] 0.65528023  7.154773087 1.03766520 6.20994386 3.686728e+01
#>   [4,] 0.93800904  1.804189317 1.08583986 1.27931339 3.906348e+00
#>   [5,] 0.61856947  0.403703799 0.39727058 0.58807726 3.698035e-01
#>   [6,] 0.28326951  2.715939963 1.11281613 0.22511649 1.319237e+00
#>   [7,] 0.87452837  0.693429750 0.50401391 0.50319913 1.268390e+00
#>   [8,] 0.86311851  0.184369315 0.23096431 0.33068826 2.911883e-01
#>   [9,] 0.80044080  1.922133072 1.00525145 1.21535868 1.080936e+00
#>  [10,] 0.96407187  1.730025921 1.41080607 1.76713425 1.706267e+00
#>  [11,] 0.04748024  0.021343795 0.03785445 0.02818398 5.157715e-02
#>  [12,] 0.45297132  0.531339495 0.42509856 0.38353487 8.496486e-01
#>  [13,] 0.43104100  2.391699109 1.58212678 0.53915009 2.639521e+00
#>  [14,] 0.38055366  0.189107400 0.17845993 0.14601477 1.428546e-01
#>  [15,] 0.85733889  0.376375149 0.42904851 0.35228854 4.549264e-01
#>  [16,] 0.12898108  0.059564310 0.11269061 0.16404650 9.052870e-02
#>  [17,] 0.60196374  3.041056402 1.13316773 1.27215813 4.245364e+00
#>  [18,] 0.89870682  0.820242478 0.52394750 1.03111005 3.170984e-01
#>  [19,] 0.03962856  0.039275996 0.04787973 0.03615701 2.464443e-02
#>  [20,] 0.40823544  3.678864732 0.59917061 1.24236086 7.034160e+00
#>  [21,] 0.42830773  1.270656580 1.16612033 1.40047609 2.299693e+00
#>  [22,] 0.30001176  0.932000587 0.32473417 0.37724890 7.544009e-01
#>  [23,] 0.20944173  0.731761121 0.56042215 0.55186811 3.345337e-01
#>  [24,] 0.86621476  3.904636419 2.19485980 1.17352899 9.104315e+00
#>  [25,] 0.57936358  2.017362653 0.81794439 1.04546755 1.911862e+00
#>  [26,] 0.06492521  0.164550543 0.17952146 0.13259292 7.916350e-02
#>  [27,] 0.30742699  0.110160465 0.47632261 1.24790513 7.122424e-01
#>  [28,] 0.60259831  0.918703751 0.92308334 1.22126722 5.774165e-01
#>  [29,] 0.91685412  1.053904256 2.49515224 3.13336552 1.413329e+00
#>  [30,] 0.20671252  1.606918636 0.40234778 0.45502920 1.452378e+00
#>  [31,] 0.23472399  1.431000511 0.59149288 0.61522921 3.502723e-01
#>  [32,] 0.89729723  0.206301677 1.01967082 0.92642109 3.918880e-01
#>  [33,] 0.30219302  0.324968074 0.27079566 0.10032827 1.112153e+00
#>  [34,] 0.37658111  1.189228066 0.50857366 1.69010746 1.506356e+00
#>  [35,] 0.04961565  0.144491264 0.06269722 0.10436966 1.495526e-01
#>  [36,] 0.97842994 25.339662697 3.03051191 6.22847427 2.838330e+02
#>  [37,] 0.80182566  2.739181323 1.35733218 1.84793300 1.927814e+00
#>  [38,] 0.47071633  2.296616644 1.82490565 0.10768776 3.753921e-01
#>  [39,] 0.33181935  9.482530344 0.53882369 1.67124754 2.795403e+01
#>  [40,] 0.45009044  0.426405969 0.46533369 3.46504984 6.471300e-01
#>  [41,] 0.82478919 15.398211864 2.39254893 1.42252463 1.134795e+01
#>  [42,] 0.39472770  1.293504814 0.66224036 0.88533318 6.936537e-01
#>  [43,] 0.73313990  0.792998519 0.46405893 0.63802438 8.293092e-01
#>  [44,] 0.52623679  0.729435382 0.48177254 1.08114357 7.987265e-01
#>  [45,] 0.45852510  4.417190922 1.32503559 1.92775423 5.436447e+00
#>  [46,] 0.80771423 10.321952342 1.12125772 3.46238539 7.592795e+00
#>  [47,] 0.84763122  3.068147531 2.41619245 3.39865294 1.041840e+01
#>  [48,] 0.13975716  0.321284766 0.35849451 0.17475267 8.345800e-02
#>  [49,] 0.19978242  0.076707752 0.12262412 0.10823470 1.641942e-01
#>  [50,] 0.54807078  0.321492487 0.31255581 0.51450228 5.579686e-01
#>  [51,] 0.09712431  0.390165276 0.08535259 0.08430774 2.329115e-01
#>  [52,] 0.02347394  0.006746202 0.02622071 0.02384899 1.547558e-02
#>  [53,] 0.56354040  4.645275265 1.26895960 0.23600718 1.487667e+00
#>  [54,] 0.11300795  0.059261935 0.09932344 0.20954273 1.182892e-01
#>  [55,] 0.09184337  0.132614162 0.13381345 0.08160880 1.262332e-01
#>  [56,] 0.69824975  0.066249331 0.16276681 0.09630583 4.590860e-02
#>  [57,] 0.64431011  1.327678082 1.05346045 1.44146847 2.234204e+00
#>  [58,] 0.76303210  5.303275764 2.49159997 1.38037136 2.282895e+00
#>  [59,] 0.20026442  2.598715316 0.27059464 1.32842906 1.375966e+01
#>  [60,] 0.94303788  1.023057944 1.46954861 1.14426890 1.208648e+00
#>  [61,] 0.76528838  0.846404748 1.64822348 0.90958392 7.699276e-01
#>  [62,] 0.75315318  3.926563954 2.01967145 0.40907634 1.889847e+00
#>  [63,] 0.37618152  0.879140182 0.56996238 0.43054185 9.200434e-01
#>  [64,] 0.75856126  0.161591992 0.39948680 0.28621780 7.841553e-02
#>  [65,] 0.27764748  0.240282071 0.20495307 0.24515546 3.029998e-01
#>  [66,] 0.85859741  0.956496021 0.68277121 1.75563917 4.164809e+00
#>  [67,] 0.69954323  2.350831426 1.57066027 1.48357997 1.038172e+00
#>  [68,] 0.60861314  0.768055505 2.27542217 0.59766860 5.896826e-01
#>  [69,] 0.47624197  0.218584874 0.15940865 0.20021185 1.794073e-01
#>  [70,] 0.49769580  0.042035001 0.10108720 0.07830028 3.306837e-02
#>  [71,] 0.24985026  0.392221639 0.21482558 0.17589895 2.336196e-01
#>  [72,] 0.83882583  4.443048308 1.47375617 4.19954366 1.055162e+01
#>  [73,] 0.32659785  0.302866403 0.23084839 0.16203994 7.108523e-01
#>  [74,] 0.87660334  2.561584837 0.83485269 1.64091854 1.111380e+01
#>  [75,] 0.29401441  0.194440362 0.40388997 0.50193681 2.675737e-01
#>  [76,] 0.66021973  0.746197404 0.47947214 0.27229050 8.912671e-01
#>  [77,] 0.38005487  0.678907930 0.42905685 0.25910139 1.998884e+00
#>  [78,] 0.76944940  1.369854684 1.12535321 0.92435478 1.950243e+00
#>  [79,] 0.81481930  8.662270842 2.49578058 2.76153692 1.979671e+01
#>  [80,] 0.79324960  1.039944082 0.77648544 1.48698230 4.209600e-01
#>  [81,] 0.35470916  0.489193462 0.38475512 0.37471742 8.788792e-01
#>  [82,] 0.79977941  0.988461131 1.01970651 1.47868145 9.197508e-01
#>  [83,] 0.32881561 25.767117474 0.61420597 3.82402252 4.966289e+01
#>  [84,] 0.21604119  0.479168224 0.34104326 0.40314461 7.191493e-01
#>  [85,] 0.36795759  0.961888469 0.57270216 0.67594318 1.299667e+00
#>  [86,] 0.90072684  1.081154878 0.57323246 1.02418100 6.669800e-01
#>  [87,] 0.49600277  7.907941442 0.73116158 0.82989087 5.510467e+00
#>  [88,] 0.17842930  2.711724620 0.27423265 1.29860142 2.885498e+01
#>  [89,] 0.68401214  0.530017305 0.71603448 0.66963118 1.278597e+00
#>  [90,] 0.59847118  1.020509935 1.28409192 0.65634363 1.932398e+00
#>  [91,] 0.97036301  1.736164798 5.10218404 1.20272372 5.045185e-01
#>  [92,] 0.48573399  0.681823768 1.93642509 1.56460188 9.435714e-01
#>  [93,] 0.18866067  0.225356642 0.19174906 0.27685012 4.713789e-01
#>  [94,] 0.16904557  1.273984569 0.31341030 0.53891780 8.747171e-01
#>  [95,] 0.93593815  5.664100734 1.70983329 1.73549529 1.401093e+01
#>  [96,] 0.04693083  0.012803508 0.01418444 0.01500060 8.739452e-03
#>  [97,] 0.78548284  0.537600299 0.45367474 0.55983755 3.656943e-01
#>  [98,] 0.90125117  4.479892125 1.10638094 2.49708772 2.733357e+00
#>  [99,] 0.57021122  2.331272512 0.81149892 2.16980880 5.713699e+00
#> [100,] 0.60404457  0.271685463 0.28347908 0.34892543 6.193273e-01