Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>               [,1]         [,2]        [,3]        [,4]         [,5]
#>   [1,] 0.499158328  0.699252867 0.662227767 0.611415791 7.736229e-01
#>   [2,] 0.837422538 21.989506126 1.949713218 9.344722537 1.102975e+02
#>   [3,] 0.063428011  0.783930646 0.139698416 0.383965169 1.851752e+00
#>   [4,] 0.428492947  0.709651276 1.108550878 0.602693909 4.749203e-01
#>   [5,] 0.353616879  0.277288714 0.226110448 0.313432741 3.519441e-01
#>   [6,] 0.875992570  0.854103106 2.640294151 0.688442502 6.410733e-01
#>   [7,] 0.044069652  0.022632891 0.044911126 0.026444943 3.941159e-02
#>   [8,] 0.801367100  0.721478611 0.622792733 0.468821674 4.724393e-01
#>   [9,] 0.324961889  6.264265614 0.713635592 0.499544552 1.337130e+00
#>  [10,] 0.295564924  4.522699165 0.605750797 1.822890390 1.800797e+01
#>  [11,] 0.736822030  0.447789964 0.326438502 0.629690196 3.165048e+00
#>  [12,] 0.994168078  0.718442859 0.445563727 1.265494517 4.765777e+00
#>  [13,] 0.592206023 25.783819785 1.332742818 4.802006801 4.772259e+01
#>  [14,] 0.388380052  2.216794855 0.633611935 1.783758146 2.258004e+00
#>  [15,] 0.605786626 13.156281304 1.306418893 8.059208454 1.171950e+01
#>  [16,] 0.140665826  0.830990244 0.405386020 0.589858777 2.050552e+00
#>  [17,] 0.127562243  1.122124669 0.429799599 0.534263387 9.108346e-01
#>  [18,] 0.419755765  0.194737356 0.258991446 0.205813950 9.702223e-02
#>  [19,] 0.121546727  0.333759610 0.219276579 0.391760243 2.958918e+00
#>  [20,] 0.027086795  0.010515534 0.170100860 0.529554183 1.427217e-02
#>  [21,] 0.819229225  0.307640237 0.292793982 0.209742480 1.543014e-01
#>  [22,] 0.512219629  1.058180022 0.669531009 0.652365382 6.991128e-01
#>  [23,] 0.823936427  0.820432942 0.749429950 1.532494212 1.101761e+00
#>  [24,] 0.339487058  1.499322364 0.356130534 1.663727517 5.106434e+00
#>  [25,] 0.203521968  0.804087563 0.481279959 0.298480408 2.519655e-01
#>  [26,] 0.164710105  2.253971880 0.338326950 1.505242047 6.211477e+00
#>  [27,] 0.247537115  0.061407083 0.185704065 0.176273515 6.327063e-02
#>  [28,] 0.997625460 14.429531516 1.431408127 9.541589868 5.439420e+01
#>  [29,] 0.887652668 18.701801562 1.876972708 2.719057450 1.221053e+01
#>  [30,] 0.268006101  0.435250336 0.927829497 2.369846628 6.490891e-01
#>  [31,] 0.423667646  0.163399760 0.393604662 0.572837441 3.056358e-01
#>  [32,] 0.152487972  0.426422857 0.191403858 0.269879911 9.854676e-01
#>  [33,] 0.478876200  0.526585541 0.315446708 0.262730263 4.720132e-01
#>  [34,] 0.674923165  0.413916397 0.706269233 0.452677467 4.535361e-01
#>  [35,] 0.974688980  1.866545576 1.027553762 0.685761033 1.834667e+00
#>  [36,] 0.855773131  1.720798413 1.269033297 3.998998680 3.032251e+00
#>  [37,] 0.141909444  0.266309319 0.164158818 0.239820718 6.257213e-01
#>  [38,] 0.121916387  0.336853731 0.119626568 0.269734589 3.835616e-01
#>  [39,] 0.164935485  0.039686195 0.119664606 0.113881322 1.599710e-02
#>  [40,] 0.286246652  0.760152558 0.293928807 0.546295613 1.693243e+00
#>  [41,] 0.618898841  1.213457707 0.941395638 2.967816105 6.468042e+00
#>  [42,] 0.541777238  2.632691806 1.232363486 0.819437659 1.223035e+00
#>  [43,] 0.558399257  3.178006270 1.186042626 0.773924312 1.094173e+00
#>  [44,] 0.477547672  0.818330947 0.656517886 0.722800497 4.854885e-01
#>  [45,] 0.223736894  0.653019269 0.361208442 0.513276161 1.492042e+00
#>  [46,] 0.747804847  0.404504740 2.344885098 1.976931507 1.013844e+00
#>  [47,] 0.876818303  0.218701733 0.373716696 0.443542484 5.056192e-01
#>  [48,] 0.080074187  0.709438436 0.242853274 0.971964424 1.241773e+00
#>  [49,] 0.857823861  7.758594302 3.273854380 0.924543880 7.372288e+00
#>  [50,] 0.605442491  0.993016470 0.703242983 0.519744827 5.649023e-01
#>  [51,] 0.319295888  1.016285596 0.514318655 0.673788804 1.922654e+00
#>  [52,] 0.459973943  1.893608548 0.868568847 1.402679720 2.114632e+00
#>  [53,] 0.069485659  0.010481185 0.012963541 0.007545526 1.855179e-02
#>  [54,] 0.938880373 15.150710341 2.437423298 5.020535774 2.466468e+01
#>  [55,] 0.392317608  0.195792500 0.252141018 0.261415477 1.348121e-01
#>  [56,] 0.495302668  4.192751145 1.057843834 2.363060973 2.190060e+00
#>  [57,] 0.189805514  1.450961881 0.235685254 0.480283317 2.437174e+00
#>  [58,] 0.960502137  1.274396481 1.593078827 0.830925929 9.627427e-01
#>  [59,] 0.397644263  5.028164297 0.980929548 2.744727809 1.362999e+01
#>  [60,] 0.560884495  0.215896796 0.312343713 0.361468957 4.483307e-01
#>  [61,] 0.025602011  0.781451804 0.066047072 0.296793273 2.648586e+00
#>  [62,] 0.353403670  2.394613824 0.961466254 0.900932700 1.264028e+01
#>  [63,] 0.799668815  1.578171922 0.671281265 0.499544476 4.379968e+00
#>  [64,] 0.454657472  0.359704847 0.391617564 0.317844677 4.555878e-01
#>  [65,] 0.255853304  0.254015440 0.195720628 0.148385234 1.221129e-01
#>  [66,] 0.428009200  0.615176902 0.463078374 0.416080874 7.585829e-01
#>  [67,] 0.210121456  0.315815447 0.178516816 0.134496012 2.808603e-01
#>  [68,] 0.873482689  9.367068154 1.565638895 3.666128585 1.920338e+01
#>  [69,] 0.152186692  0.084776834 0.159899445 0.114260135 1.008382e-01
#>  [70,] 0.727782419 13.261855056 0.995241722 6.949692433 3.661885e+01
#>  [71,] 0.853882083  1.131684943 1.360213051 1.380097310 1.088847e+00
#>  [72,] 0.491763485  0.323076567 1.236932416 1.815043509 3.528591e-01
#>  [73,] 0.544561251  1.340550039 0.591076224 0.824007955 4.561478e+00
#>  [74,] 0.640579579  1.945838574 1.038512157 0.621166715 8.240054e-01
#>  [75,] 0.256648297  7.935604204 0.452633994 1.056407326 7.305407e+00
#>  [76,] 0.065358213  0.322860033 0.095735003 0.241994238 3.719614e-01
#>  [77,] 0.110877090  0.055464572 0.141931090 0.264492166 1.279930e-01
#>  [78,] 0.068961691  0.748797453 0.177611085 0.028598529 4.451176e-01
#>  [79,] 0.419307675  3.143475195 0.608303322 0.840348338 2.648170e+00
#>  [80,] 0.211668814  0.968379237 0.344741693 0.147274877 2.524035e+00
#>  [81,] 0.896348764  0.696970280 0.647572521 0.882061004 1.904141e+00
#>  [82,] 0.247108327  2.200028270 0.565590396 0.691871035 4.466509e+00
#>  [83,] 0.722638619  0.904397369 0.777149384 0.405504216 6.416990e-01
#>  [84,] 0.267109134  0.964240159 0.392869507 0.799137835 2.951314e+00
#>  [85,] 0.147819770  0.082625213 0.142809137 0.116156288 6.957402e-02
#>  [86,] 0.230454250  0.854212181 0.320502255 0.420555742 2.974508e-01
#>  [87,] 0.051495944  0.186898302 0.140619590 0.089809348 1.097337e-01
#>  [88,] 0.152432469  0.197959439 0.101790516 0.103631050 2.601626e-01
#>  [89,] 0.063682337  0.435782856 0.124152047 0.391367673 1.159417e+00
#>  [90,] 0.513158314  4.208795635 0.776528473 0.362721664 1.054939e+01
#>  [91,] 0.345429814  8.187567268 0.909927560 1.100529588 2.438481e+00
#>  [92,] 0.339833120  1.239477351 0.473598959 0.299166221 9.450575e-01
#>  [93,] 0.912541804  7.023650751 1.916334644 2.468168100 1.055027e+01
#>  [94,] 0.940734856  0.924700184 1.023690255 1.281123687 9.656301e-01
#>  [95,] 0.997665707  0.321802140 0.902432800 1.133588003 7.598331e-01
#>  [96,] 0.654975892  0.077599404 0.196135077 0.232611406 2.313725e-01
#>  [97,] 0.040817734  0.179192255 0.080666639 0.113215369 1.649839e-01
#>  [98,] 0.007583677  0.001153776 0.001754742 0.001428918 7.054384e-04
#>  [99,] 0.308636958  1.625279618 0.809340521 1.274487677 7.812239e-01
#> [100,] 0.713103807  1.845974744 0.709164230 1.387946545 1.300042e+01