Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>               [,1]         [,2]         [,3]        [,4]         [,5]
#>   [1,] 0.403581480 6.218500e-01 0.5639816645 0.650402017 7.024313e-01
#>   [2,] 0.674327081 3.220519e+00 1.1406920493 0.953574938 1.196970e+00
#>   [3,] 0.708204100 1.238840e-01 0.1185180909 0.176482819 3.684799e-01
#>   [4,] 0.780311879 4.513755e-01 1.5024354128 1.229475569 1.059478e-01
#>   [5,] 0.844789159 2.865600e+00 2.3022040964 0.594968999 1.682706e+00
#>   [6,] 0.530972566 5.241075e-01 0.4956084814 0.738349589 1.774774e+00
#>   [7,] 0.847644185 1.701178e+00 1.1700632663 0.738106038 1.739376e+00
#>   [8,] 0.779375379 3.762410e+00 1.0752366337 4.120478951 6.922655e+00
#>   [9,] 0.250793001 3.264338e-01 0.6452693961 0.211142273 2.669330e-01
#>  [10,] 0.197125480 1.693979e-01 0.0791300861 0.084131936 1.062951e-01
#>  [11,] 0.251851775 4.388302e-01 0.2370720384 0.434393844 8.343096e-01
#>  [12,] 0.804212615 1.090665e+00 1.3394753077 0.872593533 5.160116e-01
#>  [13,] 0.273828113 1.164775e+00 0.4678922052 2.583259026 1.622542e+00
#>  [14,] 0.473012914 1.050289e+00 0.5743488901 1.735495786 1.432018e+00
#>  [15,] 0.896324910 6.390908e+00 2.3816020781 1.181485938 2.327251e+01
#>  [16,] 0.600711159 1.446150e+00 0.9867881877 1.654709062 1.312327e+00
#>  [17,] 0.023225233 7.126071e-03 0.0059542153 0.008336415 5.540073e-03
#>  [18,] 0.001750387 2.884120e-02 0.0040868518 0.002884684 4.133541e-02
#>  [19,] 0.549031760 2.286851e-01 0.2496326636 0.246895564 2.437365e-01
#>  [20,] 0.110248417 3.422770e-01 0.5270948436 0.064345863 2.476604e-01
#>  [21,] 0.462152240 7.176824e-01 0.3684011434 0.456611688 6.440741e-01
#>  [22,] 0.373336872 1.758507e+00 0.4814511820 0.825276382 2.695887e+00
#>  [23,] 0.305293542 5.977195e-01 0.3218338727 0.272835074 9.767019e-01
#>  [24,] 0.770698696 3.671480e+00 1.2848792837 1.521737693 1.028209e+01
#>  [25,] 0.622832637 9.127023e-01 0.6806699408 1.100143251 1.078409e+00
#>  [26,] 0.125283652 9.771868e-02 0.0873320018 0.123952551 1.310355e-01
#>  [27,] 0.854483052 7.093320e+00 1.6202622118 3.109985416 6.204422e+00
#>  [28,] 0.655942143 1.649980e+01 0.9554009237 3.972497434 2.284469e+01
#>  [29,] 0.609916360 6.551117e+00 1.0624718109 1.243036051 8.897794e+00
#>  [30,] 0.949806250 5.010736e+00 1.9219711638 2.806121821 3.132288e+00
#>  [31,] 0.554697499 6.273042e-02 0.0661463261 0.152964507 1.467115e-01
#>  [32,] 0.510502696 2.060318e+00 0.6913914792 1.189063654 1.965268e+01
#>  [33,] 0.049903525 1.735048e-01 0.0450244439 0.071375610 5.724567e-01
#>  [34,] 0.755000015 9.161181e-01 0.7276365328 0.735793966 9.772357e-01
#>  [35,] 0.923574931 4.621457e-01 1.4895810361 1.685540027 1.100463e+00
#>  [36,] 0.646113908 1.074671e+00 0.6460652764 0.211211576 3.580666e-01
#>  [37,] 0.465631132 2.138088e+00 1.0944439210 0.803240910 1.701477e+00
#>  [38,] 0.793732004 7.954850e+00 0.9709976165 4.678117207 7.568283e+00
#>  [39,] 0.542076449 2.987111e-01 0.4073212774 0.325648311 3.252988e-01
#>  [40,] 0.432501464 1.030645e+00 0.7600957901 0.357074449 1.224241e+00
#>  [41,] 0.391214445 5.774712e-01 0.3213581015 0.220455374 1.703174e-01
#>  [42,] 0.973033887 5.044187e+00 1.4186465830 4.670203564 2.592268e+00
#>  [43,] 0.472806100 1.011476e+00 1.6189594809 0.547005135 1.060413e+00
#>  [44,] 0.608624158 1.494233e-01 0.1750851198 0.266132433 2.830725e-01
#>  [45,] 0.107994655 3.158839e-01 0.0785823589 0.083036518 1.069037e-01
#>  [46,] 0.890851482 4.119806e-01 0.3440533150 0.291872981 6.426444e-01
#>  [47,] 0.815953593 3.267941e+00 1.7544918965 1.461107340 4.954274e+00
#>  [48,] 0.164223511 1.289420e+00 0.4101669111 0.128093783 1.223722e+00
#>  [49,] 0.671853540 1.064063e+00 1.0918502181 0.799571688 1.469113e+00
#>  [50,] 0.205235103 9.862245e-01 0.4198620296 0.234412935 2.455196e+00
#>  [51,] 0.692642533 2.397779e-01 0.2255457243 0.446636466 4.065573e-01
#>  [52,] 0.345399891 7.866279e+00 0.6225716905 1.694582458 5.134821e+00
#>  [53,] 0.760287892 4.585660e-01 0.2775389502 0.509196068 5.130934e-01
#>  [54,] 0.536809041 1.106047e+00 0.8406875156 0.760448882 5.159456e-01
#>  [55,] 0.925212087 1.007520e+01 3.3156666205 0.627306357 7.008038e+00
#>  [56,] 0.921575615 5.760763e-01 0.5519181220 0.769684957 6.442081e-01
#>  [57,] 0.458316273 4.626613e-01 0.5402254501 0.826840322 4.181805e-01
#>  [58,] 0.945556875 2.642557e+00 1.3033520587 1.951160169 6.127914e+00
#>  [59,] 0.425207177 5.888513e+00 1.0443129584 1.960136425 1.671263e+01
#>  [60,] 0.151988661 5.400301e-01 0.2124664161 0.224452991 3.977734e-01
#>  [61,] 0.982845955 1.549347e+00 0.9573967806 1.568033070 2.025924e+00
#>  [62,] 0.402879863 3.934802e+00 0.9524193400 2.038637134 2.198622e+01
#>  [63,] 0.940718177 1.792637e+00 0.6878600932 0.779949236 3.473561e+00
#>  [64,] 0.003218078 7.421575e-04 0.0008798702 0.001083649 3.905553e-04
#>  [65,] 0.011280238 7.474392e-02 0.0181764357 0.042720857 6.532127e-02
#>  [66,] 0.093576210 8.789674e-01 0.1777604136 0.588960536 2.295830e+00
#>  [67,] 0.886311974 2.296812e+00 1.2436114014 1.378194347 2.163586e+00
#>  [68,] 0.369199212 9.816645e-02 0.2265483208 0.198402283 5.200259e-02
#>  [69,] 0.129155861 8.639444e-01 0.3416270914 0.691202081 1.380479e+00
#>  [70,] 0.361812644 2.760292e-01 0.5389157968 0.377937292 6.421073e-01
#>  [71,] 0.985057678 8.132744e-01 1.2651917636 0.418273111 5.699198e-01
#>  [72,] 0.780235301 3.854361e-01 0.6839702283 0.419676201 5.377448e-01
#>  [73,] 0.569011199 2.596830e-01 0.6168326114 0.327830735 3.000200e-01
#>  [74,] 0.257987596 3.739824e-01 0.5254806549 0.471876655 4.314709e-01
#>  [75,] 0.903542893 3.948196e+00 1.8451141075 2.112858314 9.376551e+00
#>  [76,] 0.395594632 1.847854e-01 0.2795810065 0.257990484 5.779691e-01
#>  [77,] 0.693633539 1.860209e+00 1.1423617319 0.908606504 1.442517e+00
#>  [78,] 0.287862395 4.185701e-01 0.2194205828 0.207279697 4.552395e+00
#>  [79,] 0.204786850 8.995647e-02 0.0925142072 0.112286663 6.708551e-02
#>  [80,] 0.495463626 3.382284e+00 2.2075857380 0.259075795 1.737790e+00
#>  [81,] 0.944128870 2.964399e-01 0.9531039877 0.583172358 2.936575e-01
#>  [82,] 0.433571746 1.886171e+00 1.1368769765 1.641596982 3.324717e+00
#>  [83,] 0.928765815 1.186788e+00 1.0264739318 0.545480286 1.833212e+00
#>  [84,] 0.770884360 5.628079e-01 0.6354432684 0.359647821 6.027858e-01
#>  [85,] 0.358402491 1.327709e-01 0.4709028137 1.168079804 3.840954e-01
#>  [86,] 0.078226475 1.536257e-01 0.1535032553 0.076782398 2.226787e-01
#>  [87,] 0.884582168 8.195358e+00 1.2014773881 2.223990512 5.488538e+00
#>  [88,] 0.001572203 3.846283e-03 0.0022872825 0.002146800 2.250715e-02
#>  [89,] 0.094502677 9.899508e-02 0.0422922447 0.078186602 2.049133e-01
#>  [90,] 0.284567179 1.085965e+00 0.3616154136 1.028366103 3.323612e+00
#>  [91,] 0.440414882 3.515500e-01 0.2340069120 1.538079881 3.238332e-01
#>  [92,] 0.235820327 3.010249e-01 0.1839785108 0.526683395 8.406646e-01
#>  [93,] 0.357849628 5.757917e-01 0.5986832008 0.449842262 1.608998e+00
#>  [94,] 0.162667206 1.470939e+00 0.4009262897 0.435850039 1.143615e+00
#>  [95,] 0.126907689 4.082115e-02 0.1108328684 0.148348758 3.993893e-02
#>  [96,] 0.413317086 1.919144e+00 0.5591031537 0.980182197 1.565730e+00
#>  [97,] 0.223468062 7.593816e-01 0.2366051197 0.595820910 4.614682e+00
#>  [98,] 0.777252387 1.257957e+01 3.3567076589 9.463977815 4.762678e+01
#>  [99,] 0.413903629 3.721454e-01 2.0957074134 0.447473874 2.976279e-01
#> [100,] 0.389072971 2.253381e-01 0.1141195624 0.176633851 3.436379e-01