Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>                [,1]         [,2]         [,3]         [,4]         [,5]
#>   [1,] 0.8251973614  5.701230547 2.0273921261 2.054529e+00 9.829084e+00
#>   [2,] 0.1831985516  0.202857660 0.1718995154 1.921681e-01 4.269901e-01
#>   [3,] 0.2479723268  0.702366857 0.4212808624 3.821868e-01 1.368936e+00
#>   [4,] 0.5333320184  1.155214820 0.5601362561 6.867178e-01 6.092339e-01
#>   [5,] 0.5555443638  1.976219155 0.6060261060 7.697226e-01 3.405319e+00
#>   [6,] 0.0130984203  0.044491742 0.0199009885 4.712916e-02 1.445743e-01
#>   [7,] 0.5928794087  0.146294710 0.2304618786 2.746467e-01 4.349863e-01
#>   [8,] 0.4792220572  0.462898503 0.3752541462 3.931847e-01 4.073383e-01
#>   [9,] 0.8749005923  9.977575982 3.2989562635 7.531692e-01 1.614630e+00
#>  [10,] 0.7729851687  2.336575018 0.8868932027 1.302061e+00 6.950983e-01
#>  [11,] 0.4562597917  0.923619285 0.6178873083 2.687617e-01 6.281666e-01
#>  [12,] 0.3561238064  0.609241001 0.7578778220 5.042525e-01 4.672940e-01
#>  [13,] 0.7673630838  0.290164799 0.3349472085 4.173897e-01 8.125684e-01
#>  [14,] 0.2479150109  0.314981196 0.2679134770 2.730183e-01 1.264505e-01
#>  [15,] 0.3837462522  0.150571029 0.1170129449 1.208755e-01 3.133553e-01
#>  [16,] 0.2449821068  0.843469565 0.3992756594 6.848810e-01 4.312071e-01
#>  [17,] 0.5464961312  0.130274005 0.1512586939 1.994058e-01 4.031004e-01
#>  [18,] 0.9495587284  0.926662754 0.7035832528 6.743701e-01 5.800734e-01
#>  [19,] 0.0422202484  0.986768340 0.0814001354 1.389029e-01 6.615811e+00
#>  [20,] 0.9426022666  0.780551854 0.7833077814 1.586006e+00 4.465206e-01
#>  [21,] 0.6041278373  0.389816319 0.3546165904 3.998469e-01 8.762566e-01
#>  [22,] 0.0845119483  0.186896239 0.1118530402 1.260910e-01 1.214871e-01
#>  [23,] 0.4709187113  1.740554980 0.8993642900 9.949366e-01 4.266378e+00
#>  [24,] 0.9699496399  0.617410253 4.2634914566 1.534660e+01 4.268945e+00
#>  [25,] 0.4374618598  2.365791186 0.8148124588 8.108975e-01 1.425781e+00
#>  [26,] 0.0216506566  0.018316022 0.0198528753 3.210892e-02 1.703926e-02
#>  [27,] 0.6936571104  0.341995764 0.9828364745 6.250078e-01 2.893676e-01
#>  [28,] 0.0400833902  0.105776662 0.0609926073 4.270832e-02 1.404953e-01
#>  [29,] 0.0755803874  0.233241506 0.1296286290 9.005538e-02 6.129783e-01
#>  [30,] 0.3402500248  1.914916663 0.8659711070 1.265584e+00 1.282698e+00
#>  [31,] 0.6193219677  7.189074422 1.4872824053 5.685689e+00 1.554114e+01
#>  [32,] 0.7387142647  1.683837864 0.8208938907 1.917251e+00 3.065956e+00
#>  [33,] 0.6175611264  0.108068081 0.1031751900 2.251273e-01 1.815946e-01
#>  [34,] 0.7110740447  3.151711700 1.6465292183 6.145362e-01 3.894942e+00
#>  [35,] 0.7714683437  3.519478781 1.3263381812 2.733647e+00 1.911120e+00
#>  [36,] 0.9381326437  0.616328338 0.8533991110 2.746243e+00 1.730068e+00
#>  [37,] 0.8281670418  3.976710909 1.5335280079 2.244926e+00 6.688390e+00
#>  [38,] 0.6364102003  6.333672868 2.4520524506 8.675636e-01 2.079050e+00
#>  [39,] 0.0508555579  0.380564116 0.1670737068 9.595499e-03 1.795021e-01
#>  [40,] 0.0916645769  0.268357104 0.1343620058 1.142656e-01 9.376687e-02
#>  [41,] 0.5605330982  0.587022764 0.5385992055 5.501247e-01 2.059783e-01
#>  [42,] 0.5209168813  0.142060224 0.3563267649 4.534317e-01 8.195822e-01
#>  [43,] 0.2085704287  1.283724524 0.3617377627 2.355226e-01 9.953830e-01
#>  [44,] 0.9821219654  4.192255016 2.4834862755 2.055681e+00 2.824608e+00
#>  [45,] 0.0808335959  0.126665711 0.0843994167 1.228708e-01 4.196292e-01
#>  [46,] 0.7240786322  2.242489854 0.9369886789 4.835761e-01 3.604355e-01
#>  [47,] 0.7884045406  1.288314165 1.3147393933 8.068682e-01 9.245817e-01
#>  [48,] 0.0785880608  0.016614391 0.0226103529 2.356861e-02 1.178093e-02
#>  [49,] 0.7993670031  1.213144127 0.7093887658 7.795353e-01 1.021417e+00
#>  [50,] 0.8104135618  2.792852929 1.3111536356 2.624094e+00 1.151920e+01
#>  [51,] 0.7719829341  0.877539859 0.4482510762 6.728702e-01 3.425328e+00
#>  [52,] 0.7371879697  0.246838619 0.4905815250 9.468048e-01 9.457587e-01
#>  [53,] 0.1658312753  0.176259504 0.1014037647 1.137511e-01 2.727127e-01
#>  [54,] 0.5349086730  0.584923565 0.3834376803 4.213822e-01 2.201546e-01
#>  [55,] 0.2883314400  0.997217938 0.4763051456 1.648062e+00 1.597370e+00
#>  [56,] 0.8517398494  6.052483920 1.0717954691 4.445420e+00 4.049950e+00
#>  [57,] 0.7298811688 39.278349293 2.6834180686 9.984940e+00 2.066612e+01
#>  [58,] 0.5764032989  1.645581707 0.9133316009 6.210318e-01 2.110884e+00
#>  [59,] 0.8593858290  0.741664925 0.9124105815 5.140209e-01 4.636613e-01
#>  [60,] 0.9409642036  8.038161489 1.6202262939 1.513157e+00 7.084832e+00
#>  [61,] 0.2814040866  0.928636242 0.4509740833 2.483920e-01 2.538834e-01
#>  [62,] 0.9026621867  2.259131172 1.0186889008 3.450049e+00 1.185190e+00
#>  [63,] 0.5874629708  0.383843093 0.5064649393 5.469397e-01 4.175529e-01
#>  [64,] 0.6711851011  7.419705902 2.2584961832 1.034868e+00 4.243353e+00
#>  [65,] 0.9273637314  1.042242562 0.6698036441 6.760253e-01 1.174610e+00
#>  [66,] 0.9638327686 10.818349269 3.4351835512 2.145394e+00 1.168408e+01
#>  [67,] 0.1563317513  0.145776358 0.1175025439 5.760320e-02 8.585107e-02
#>  [68,] 0.6089058581  0.062185577 0.1624091872 9.333526e-02 4.813577e-02
#>  [69,] 0.0750305783  0.570284325 0.1596148878 3.045670e-01 7.689001e-01
#>  [70,] 0.4833912989 26.575091919 0.9354546389 6.573374e+00 2.043636e+01
#>  [71,] 0.3459933291  0.585518532 0.2664166617 4.220157e-01 1.479668e+00
#>  [72,] 0.7355276269  0.676837213 0.3694671910 6.585070e-01 7.945862e-01
#>  [73,] 0.4967223406  0.324598983 0.4787131449 4.668290e-01 2.770520e-01
#>  [74,] 0.3051882030  0.277897473 0.2302235619 2.518041e-01 3.700048e-01
#>  [75,] 0.6145753309  3.623995369 1.1754837250 1.736427e+00 4.721803e+00
#>  [76,] 0.2385425551  0.300631478 0.1839162622 2.276797e-01 6.573661e-01
#>  [77,] 0.3453033760  0.979855885 0.3778892601 4.783943e-01 3.505572e+00
#>  [78,] 0.2008757058  0.215131362 0.2344247033 1.153135e-01 1.312618e-01
#>  [79,] 0.9309475168  0.169010062 1.2659294914 4.842898e+00 1.166868e+00
#>  [80,] 0.8931232004  3.025087755 2.1647233141 3.904299e-01 1.206150e+00
#>  [81,] 0.9286019823  1.035945751 1.0913814592 5.335952e-01 1.506325e+00
#>  [82,] 0.6290394089  1.663404557 1.3002707977 1.408143e+00 1.902903e+00
#>  [83,] 0.9989267739  1.171251421 1.3522260976 1.436254e+00 1.146236e+00
#>  [84,] 0.7833577648  1.090417686 0.8027530965 8.989635e-01 6.110707e-01
#>  [85,] 0.7256847082  0.615252029 0.6776862504 4.793237e-01 9.462308e-01
#>  [86,] 0.2170962465  0.189725795 0.1077539330 2.231339e-01 4.391368e-01
#>  [87,] 0.7945411988  0.600716061 0.4340291569 3.255483e-01 3.724824e-01
#>  [88,] 0.9234889112  2.058755486 1.1483574708 4.053131e-01 6.161724e-01
#>  [89,] 0.0001930094  0.002034358 0.0006871232 1.034664e-04 7.816353e-04
#>  [90,] 0.5916262909 19.512791068 1.3319071799 5.930842e+00 6.750060e+01
#>  [91,] 0.2257523849  0.198537378 0.3170154827 4.139068e-01 2.628358e-01
#>  [92,] 0.1252847756  0.477916497 0.1536228696 1.611735e-01 3.229479e-01
#>  [93,] 0.2328872536  0.300589342 0.2564374253 1.662886e-01 2.042069e-01
#>  [94,] 0.6608922088  0.780922429 4.1281693925 4.799609e+00 1.170893e+00
#>  [95,] 0.0077195340  0.006456805 0.0072191831 9.688046e-03 1.011871e-02
#>  [96,] 0.8767679054  0.664917718 3.1543037730 2.739195e+00 1.977122e+00
#>  [97,] 0.7323890931  0.019193108 0.0430119526 5.659552e-02 1.913776e-02
#>  [98,] 0.5898859273  0.403932848 0.4977808397 8.298851e-01 9.882948e-01
#>  [99,] 0.6179059744  0.140187534 0.2406659432 2.247563e-01 1.599114e-01
#> [100,] 0.8221522942  3.868219069 2.0898157150 2.456449e+00 7.280483e+00