Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>              [,1]        [,2]       [,3]       [,4]         [,5]
#>   [1,] 0.67003612  0.69630094 0.89213360 1.02946582   1.01822580
#>   [2,] 0.09694424  0.08418202 0.08191437 0.05029663   0.05939301
#>   [3,] 0.78122139  3.05723697 1.05197513 0.92968303   4.42256380
#>   [4,] 0.59344747  2.82532649 0.92722580 1.08300625   1.74213804
#>   [5,] 0.31010766  0.03979634 0.06969313 0.13009851   0.01786956
#>   [6,] 0.38907914  1.14348794 0.72275722 1.86160916   0.91048837
#>   [7,] 0.32125996  0.40719092 0.78919068 0.55038469   0.28936386
#>   [8,] 0.41170594  0.72182178 0.42013441 0.37648421   0.34195183
#>   [9,] 0.76020409  2.62735381 0.81465573 1.64931031   4.75557876
#>  [10,] 0.88250233  0.42085466 0.54718859 0.38124166   0.34818063
#>  [11,] 0.14306648  0.76921157 0.16524651 0.33952378   1.56902267
#>  [12,] 0.11113899  0.31189430 0.18002487 0.26788893   0.50564808
#>  [13,] 0.80749890  0.32848385 0.73820814 0.54372746   0.38216086
#>  [14,] 0.67238802  0.56589419 0.48895063 0.51747732   0.32280194
#>  [15,] 0.68122921  0.80842253 0.71007954 0.66608986   1.53176291
#>  [16,] 0.80485925  7.23409184 1.76552409 2.28887261   2.29197459
#>  [17,] 0.36669648  1.40789963 0.48458174 0.31877443   3.57868156
#>  [18,] 0.65235384  1.70744135 0.97398005 0.70960816   1.44842511
#>  [19,] 0.20386966  0.38050358 0.44213363 0.30061209   0.40421421
#>  [20,] 0.14952198  0.11001624 0.13475080 0.14477049   0.08434790
#>  [21,] 0.78562959 13.05259227 1.17963994 5.81158260  23.67940761
#>  [22,] 0.69926882  0.49503413 0.28414833 0.59188830   1.50404514
#>  [23,] 0.36156364  0.48710252 0.51929923 0.77452734   0.37022669
#>  [24,] 0.05900058  0.13225215 0.13250927 0.05154999   0.07054459
#>  [25,] 0.18823601  0.04747665 0.11710983 0.15630468   0.18490589
#>  [26,] 0.95412179  2.43892729 1.14101006 0.88255715   0.61404470
#>  [27,] 0.25912907  0.74545459 0.47271567 0.18260161   0.22228808
#>  [28,] 0.16391606  1.11728239 0.32412734 0.61109716   4.34112327
#>  [29,] 0.40903828  0.58009649 0.24268211 0.53059776   0.53494363
#>  [30,] 0.17473245  0.27805021 0.30479273 0.10405141   0.21888043
#>  [31,] 0.31059169  0.64246484 0.45527967 0.55026131   0.75085873
#>  [32,] 0.94248990 52.27917806 1.53378980 6.18330090 227.64380115
#>  [33,] 0.56960853  2.19853127 1.17196384 1.01249044   1.63495779
#>  [34,] 0.86726616  1.53872475 0.79080390 0.58478686   1.02836094
#>  [35,] 0.10022334  1.03427781 0.23755410 0.27512101   0.85183321
#>  [36,] 0.22327356  2.19280563 0.44621787 0.45523087   4.14158455
#>  [37,] 0.77043336  0.66568422 0.54979056 0.66686629   1.24902445
#>  [38,] 0.68418172  1.09285288 0.67318628 1.03604617   1.49505537
#>  [39,] 0.66903646  1.30005277 1.62642812 2.70291201   2.46846891
#>  [40,] 0.26724986  0.99404664 0.40285661 0.81427140   2.33330310
#>  [41,] 0.04258684  0.35968912 0.07196260 0.11176037   0.49510492
#>  [42,] 0.16681311  0.26972794 0.19670063 0.46899211   0.15750132
#>  [43,] 0.54313175  0.62559756 1.25463060 0.48728964   0.33769489
#>  [44,] 0.34725450  0.63409669 0.28667656 0.29012245   0.36531606
#>  [45,] 0.52739210  0.07422435 0.24658771 0.14212962   0.07905109
#>  [46,] 0.06649395  0.05784271 0.05903006 0.03987379   0.03360248
#>  [47,] 0.42502593  6.67992480 0.86574714 1.53256017  16.42801816
#>  [48,] 0.73285938  3.65483670 1.52394704 3.41211661   2.63444157
#>  [49,] 0.26441586  0.52161062 0.26530737 0.29467121   0.62324799
#>  [50,] 0.12910348  0.66265975 0.31133513 0.32592709   0.75678220
#>  [51,] 0.29761423  2.27344170 0.30011745 1.64846490   9.83515672
#>  [52,] 0.39924529  2.04235787 0.77106940 1.10008778   1.08883254
#>  [53,] 0.18670907  2.70701965 0.25686633 3.24807561   3.45949225
#>  [54,] 0.84402530  6.08240370 1.25503437 2.15931235   3.74305639
#>  [55,] 0.73508660  0.67806065 0.99861613 0.82151856   0.95596412
#>  [56,] 0.94402707  4.41133560 1.49250480 3.07451501   6.33352398
#>  [57,] 0.19899886  2.31514161 0.41511100 1.06500453   0.82876764
#>  [58,] 0.29780914  0.63316680 0.45527945 0.32896967   0.59843281
#>  [59,] 0.62345270  0.38940451 0.98258701 1.05622601   0.60682520
#>  [60,] 0.97370138  1.96339692 0.94348071 0.95873957   3.14691804
#>  [61,] 0.24613969  0.52499872 0.36799735 0.16991795   0.16458261
#>  [62,] 0.17455212  0.86938654 0.30191804 0.43674263   1.81392858
#>  [63,] 0.78843396  2.89250421 5.50720612 1.53741653   0.80389973
#>  [64,] 0.27569940  0.74620492 0.38706903 0.29781503   1.01902428
#>  [65,] 0.09672563  0.51556541 0.14910250 0.34000033   1.07878666
#>  [66,] 0.71347517  2.78853382 1.03923179 5.35365887   2.93539073
#>  [67,] 0.10439175  0.26112920 0.13837791 0.19422272   0.35975400
#>  [68,] 0.03301211  0.14028171 0.05390987 0.03663922   0.06313857
#>  [69,] 0.99971437  3.03045238 1.65591597 4.35843600   9.83352825
#>  [70,] 0.54436415  0.25884798 1.65603900 1.71872426   0.27001273
#>  [71,] 0.60382679  0.72745870 1.38232521 1.72769378   0.20507882
#>  [72,] 0.49078627  2.04999473 0.75104042 0.49154050   3.20183096
#>  [73,] 0.30082090  0.04518570 0.16496036 0.13902721   0.03037023
#>  [74,] 0.84987221  1.01680005 0.71831506 0.62722027   1.96768315
#>  [75,] 0.62013892  1.43464281 0.63971206 2.33897798   2.76513681
#>  [76,] 0.26079553  0.38990552 0.20062970 0.27331553   0.32910260
#>  [77,] 0.42574368  0.23811343 0.17809276 0.32620756   0.13477378
#>  [78,] 0.57385593  2.82363944 1.49418523 0.27980660   0.81674760
#>  [79,] 0.64896791  1.13512363 0.90430741 0.91971870   5.14182195
#>  [80,] 0.52874328  0.52784307 1.05880312 1.17598894   0.58697249
#>  [81,] 0.73737413  3.61926213 1.73819046 2.23531980   6.30937035
#>  [82,] 0.50491723  5.82473281 1.30742944 1.21658392   4.24555470
#>  [83,] 0.49976518  2.17758244 0.76424966 0.52267442   3.10614955
#>  [84,] 0.36254115  0.15332145 0.10716280 0.10262974   0.33961719
#>  [85,] 0.16726793  0.24091997 0.23511638 0.19554682   0.26172433
#>  [86,] 0.52308128  2.85380728 1.04266467 1.12470031   1.85486497
#>  [87,] 0.20635277  1.58610007 0.70049438 0.21701600   6.42169649
#>  [88,] 0.97781797  2.81461096 1.26306925 1.52406223   1.47833770
#>  [89,] 0.10474600  0.01475910 0.06427788 0.10249938   0.02351122
#>  [90,] 0.24768797  0.74502457 0.29332397 0.23559267   0.76746032
#>  [91,] 0.24304906  0.02301281 0.02051694 0.02531571   0.04800986
#>  [92,] 0.32673058  0.59543764 0.29395380 0.69044728   0.53742609
#>  [93,] 0.58747313  0.31709582 0.46750045 0.56542011   0.25292974
#>  [94,] 0.04945889  0.07688025 0.07172474 0.10355327   0.10096824
#>  [95,] 0.87047863  5.92164018 1.60692459 1.63702261   7.30318591
#>  [96,] 0.11673007 10.59167903 0.20604859 2.11373348  22.08288516
#>  [97,] 0.65086200  1.28205088 1.21270873 0.52584316   4.99274141
#>  [98,] 0.97834920  0.56038066 1.02616866 0.92934325   0.80736116
#>  [99,] 0.49172941  0.22139130 0.29160630 0.19211522   0.43895964
#> [100,] 0.94675788  2.37376045 1.27815799 0.68646970   0.72752803