Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>               [,1]        [,2]       [,3]         [,4]         [,5]
#>   [1,] 0.991357757  0.52018891 1.88277429  0.960221837 4.467620e-01
#>   [2,] 0.404218980  3.62289067 0.37268283  0.797307355 5.067074e+00
#>   [3,] 0.435768943  0.45961456 0.25000971  0.325378835 3.371160e-01
#>   [4,] 0.912262177 27.74354446 2.10231610  0.734970070 1.052791e+01
#>   [5,] 0.591981146  4.17483893 1.04391516  0.861825546 1.800986e+01
#>   [6,] 0.260919572  0.48008413 0.55748034  0.312661285 4.222276e-01
#>   [7,] 0.800905644 12.18800671 2.65386065  2.411878511 4.631235e+01
#>   [8,] 0.985948450  2.20801896 1.20724456  1.846066418 1.307362e+00
#>   [9,] 0.646727619  0.18518761 0.25100340  0.522004524 9.653089e-01
#>  [10,] 0.676791475  2.37733612 0.74790353  2.440644368 7.728973e+00
#>  [11,] 0.295700322  2.02708585 0.48937858  1.100435122 2.487758e+00
#>  [12,] 0.453192703  1.89693220 0.72719467  0.236796974 4.311464e+00
#>  [13,] 0.763615274  0.73388878 1.08910491  0.562945783 1.729977e-01
#>  [14,] 0.138032861  1.03615820 0.25013104  0.181855866 5.386700e-01
#>  [15,] 0.139691699  0.29166734 0.17760848  0.107430435 7.355070e-01
#>  [16,] 0.450426145  0.21604031 0.48221275  0.332014134 8.412941e-02
#>  [17,] 0.292753443  0.05087246 0.10744347  0.094842106 6.532331e-02
#>  [18,] 0.430216552  3.12499986 0.87919159  0.472090469 2.236660e+00
#>  [19,] 0.357564114  1.00982868 0.66524129  0.464999696 4.080479e-01
#>  [20,] 0.817715026  3.28260242 2.14594012  0.566571051 1.798774e+00
#>  [21,] 0.135129754  1.39882052 0.35763825  0.800049075 1.025585e+00
#>  [22,] 0.270122744  0.02569959 0.05673118  0.075297662 3.455956e-02
#>  [23,] 0.638289015  4.15249928 1.75374385  1.025333769 6.218491e+00
#>  [24,] 0.480597424  0.19396359 0.33096650  0.465238331 4.663734e-02
#>  [25,] 0.241521813  0.16380559 0.17772142  0.116323682 1.290603e-01
#>  [26,] 0.486991492  1.68285198 0.71173887  0.272332070 5.589524e-01
#>  [27,] 0.860963601  0.37996206 0.99294263  0.711819246 3.746738e-01
#>  [28,] 0.912641123  2.53661360 1.23552444  0.877419898 3.976483e+00
#>  [29,] 0.233019641  1.82207721 0.40861336  1.519700157 4.588115e+00
#>  [30,] 0.603637154 13.97363205 0.93241425  3.361199821 2.541293e+01
#>  [31,] 0.145483730  0.12406385 0.13336351  0.138336072 1.471853e-01
#>  [32,] 0.901031288  1.15679678 1.33386519  1.643701357 4.771821e+00
#>  [33,] 0.322403980  0.16012209 0.24840178  0.227211724 1.355716e-01
#>  [34,] 0.477538716  0.25039917 0.37233234  0.248151746 3.256030e-01
#>  [35,] 0.764794372  6.86904862 1.21288894  3.486429680 2.153381e+00
#>  [36,] 0.102184477  0.23699003 0.13434091  0.143278923 1.622980e-01
#>  [37,] 0.868003509  1.64692370 1.21157004  2.102855600 2.662272e+00
#>  [38,] 0.907353539  8.67602495 2.57995419  5.023013854 3.551801e+01
#>  [39,] 0.416199307  3.28886872 2.29657074  0.214536456 1.228921e+00
#>  [40,] 0.523106387  1.73672715 1.25301612  0.521340956 1.189657e+00
#>  [41,] 0.801698499  0.39799079 0.44431164  1.753341119 6.700240e-01
#>  [42,] 0.729098884  0.71778915 0.70235775  0.729475719 6.294565e-01
#>  [43,] 0.785437728  0.43347077 0.97152173  1.067377591 5.046916e-01
#>  [44,] 0.515854872  0.86528919 0.60580339  0.506168829 6.216920e-01
#>  [45,] 0.317999614  0.85507331 0.28843880  0.512199204 4.742626e-01
#>  [46,] 0.822099595  3.93767243 1.17334622  2.441860057 4.077867e+00
#>  [47,] 0.937270607 10.82973380 1.22223155  2.156465827 1.180257e+01
#>  [48,] 0.341422558  0.49914640 0.29066600  0.504561787 9.076275e-01
#>  [49,] 0.557127632  0.18372639 0.22032192  0.265182574 1.418485e-01
#>  [50,] 0.751872406 29.03463427 2.36033446  3.525396964 2.639255e+02
#>  [51,] 0.394087466  1.32391236 0.90449685  0.538843115 2.176942e+00
#>  [52,] 0.821723466  6.36887687 1.70776894  2.756149492 8.414471e+01
#>  [53,] 0.100128635  0.76710240 0.18651418  0.152408092 2.060242e+00
#>  [54,] 0.896348200  5.25187607 2.10605884  1.401069180 9.198895e+00
#>  [55,] 0.910374253  1.30419718 1.26561886  2.239494342 1.648402e+00
#>  [56,] 0.006707768  0.03152391 0.01321412  0.038762562 4.176027e-02
#>  [57,] 0.375120186  0.04305078 0.11437418  0.163094100 8.394972e-02
#>  [58,] 0.355053339  1.16043868 0.52460282  1.307845580 2.956306e+00
#>  [59,] 0.433209302  8.54159113 1.11926013  2.586840454 3.880625e+00
#>  [60,] 0.616123877  0.85362207 0.53957739  0.232428162 6.078698e-01
#>  [61,] 0.643791933  0.25883620 0.28044103  0.481878131 6.179633e-01
#>  [62,] 0.958562313  0.91249671 1.06453783  0.843362573 2.204006e+00
#>  [63,] 0.359727825  2.09126115 0.55801728  2.257770724 3.690530e+00
#>  [64,] 0.243465892  0.64663802 0.20633882  0.092799088 5.494683e-01
#>  [65,] 0.530914159  1.40726764 0.61602646  0.951490781 1.120984e+00
#>  [66,] 0.073655595  0.35895341 0.10975961  0.252729554 6.799683e-01
#>  [67,] 0.014944240  0.00193497 0.00298050  0.002461031 2.786017e-03
#>  [68,] 0.402170561  0.60362779 0.20686354  0.430883391 7.698956e-01
#>  [69,] 0.831256044 32.59406162 1.74890017  5.344961090 3.241468e+01
#>  [70,] 0.355168222  3.36451493 0.73363014  0.943783343 2.596557e+00
#>  [71,] 0.163185156  0.04376537 0.20461927  0.274452399 1.581563e-02
#>  [72,] 0.228904446  0.20987407 0.17457443  0.068115671 2.558313e-01
#>  [73,] 0.986547596  4.37576801 1.01725767  0.491481969 3.181311e+01
#>  [74,] 0.800578637  2.80247190 1.00235481  1.242652733 1.261395e+00
#>  [75,] 0.661100683  0.39131840 0.32041520  0.591526044 5.487182e-01
#>  [76,] 0.401077433  2.95619393 0.96841615  1.004660576 4.639254e+00
#>  [77,] 0.885337080 22.30148333 3.73438244  1.298911521 2.756840e+01
#>  [78,] 0.622064088  0.82603052 0.46606455  1.001994097 2.385377e+00
#>  [79,] 0.049795004  0.07200268 0.04878560  0.036344441 9.560840e-02
#>  [80,] 0.378966497  2.60341520 0.51877604  0.292992343 2.548493e+00
#>  [81,] 0.965067769  0.88014979 1.54489112  1.245112446 7.107783e-01
#>  [82,] 0.147847658  3.20848194 0.28762163  1.322345871 6.160119e+00
#>  [83,] 0.530591529  0.67483000 1.14528688  0.538246647 8.058901e-01
#>  [84,] 0.939220640 31.10343158 2.80705465 11.958075990 6.104653e+01
#>  [85,] 0.310344902  0.60142036 0.53872672  0.547636682 6.196157e-01
#>  [86,] 0.805838387  1.17457200 0.85179717  1.124734567 7.481446e-01
#>  [87,] 0.301772856  0.33745610 0.24494334  0.091664180 3.342624e-01
#>  [88,] 0.568107054  7.10261040 1.28756034  2.729580245 9.205682e+00
#>  [89,] 0.602901546  3.20639346 1.32414492  1.891202584 2.635548e+00
#>  [90,] 0.155911173  0.02857818 0.07788481  0.054773239 1.154291e-02
#>  [91,] 0.566853764  1.04658205 0.81519214  1.122745551 9.466900e-01
#>  [92,] 0.222919479  0.57131327 0.50566775  0.761796254 5.447622e-01
#>  [93,] 0.589502939  0.35791313 0.66537905  0.345780833 1.849541e+00
#>  [94,] 0.695931095  2.15724716 0.91697086  1.055603842 1.447162e+00
#>  [95,] 0.303532762  0.23620834 0.37803477  0.749826388 4.071109e-01
#>  [96,] 0.376272675  0.74842177 0.52758027  0.115629928 2.592406e-01
#>  [97,] 0.731136147  0.81376780 1.19430754  0.940850966 1.198883e+00
#>  [98,] 0.309575671  2.50233146 0.54147401  1.645504989 2.135405e+01
#>  [99,] 0.345356748  2.49086661 0.63893862  2.401525641 8.465089e+00
#> [100,] 0.450726115  1.23218380 0.47770092  1.572360109 3.910933e+00