Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>               [,1]         [,2]        [,3]        [,4]         [,5]
#>   [1,] 0.061675056  0.224230819 0.071440269 0.109798856  0.319836369
#>   [2,] 0.973910463  4.743654179 1.892852903 1.782114995 10.321797598
#>   [3,] 0.760289110  0.159771095 0.243981264 0.502971855  0.462311429
#>   [4,] 0.383207151  0.286110896 1.303594781 0.672243473  0.235124453
#>   [5,] 0.195570561  0.180701211 0.117570690 0.139624021  0.198615497
#>   [6,] 0.673857159  0.128515664 0.541718878 0.420182310  0.261431145
#>   [7,] 0.794355768  4.064144854 1.874583014 1.422436355 10.176825501
#>   [8,] 0.190564488  0.179525114 0.356162972 0.141432729  0.172958446
#>   [9,] 0.138412434  0.106985621 0.088194181 0.094444551  0.141427670
#>  [10,] 0.734688848  3.117873502 1.312903020 3.028929322  2.099433802
#>  [11,] 0.371651107  1.549120196 0.740432243 1.085951669  1.537545936
#>  [12,] 0.193902286  0.763381695 0.234688401 0.612658783  1.153203463
#>  [13,] 0.883153850  7.478979195 1.733275507 3.648921849  5.148405399
#>  [14,] 0.830053301  2.760496932 0.965710405 0.668314651  5.318018896
#>  [15,] 0.301838952  0.333619762 0.465706106 0.361543137  0.277015888
#>  [16,] 0.101600034  0.127916334 0.346521127 0.313549121  0.142041399
#>  [17,] 0.770444216  5.302598104 1.078834698 1.356938828  5.015563148
#>  [18,] 0.876183975  4.418916428 1.911430951 3.666802116  2.516286726
#>  [19,] 0.402640982  0.242870448 0.367087382 0.253297182  0.421367223
#>  [20,] 0.678585736  2.014670411 0.718008129 1.378361080  4.534065604
#>  [21,] 0.089079760  0.159919671 0.099056985 0.114264422  0.101677240
#>  [22,] 0.804141597  2.623584009 1.101141080 2.223369607  4.098175616
#>  [23,] 0.698767555  2.881214459 1.745603511 0.556745820  1.949475092
#>  [24,] 0.365492268  5.104182700 0.629034877 1.069989316 12.400523895
#>  [25,] 0.021349241  0.061230983 0.049406298 0.007294428  0.044074786
#>  [26,] 0.902529626  2.949013979 1.505122711 2.180668887  2.658250198
#>  [27,] 0.242941897  0.021845626 0.315978482 0.335804148  0.013123122
#>  [28,] 0.392598826  0.355609778 0.787367901 0.158735750  0.919237297
#>  [29,] 0.200583862  0.074717194 0.091366782 0.103944621  0.037678277
#>  [30,] 0.843051692  9.588348574 2.429282534 0.612068550  1.777249309
#>  [31,] 0.690224078  0.415385802 0.808419530 1.129930385  0.295815180
#>  [32,] 0.433861701  0.678907820 0.330007369 0.444923523  0.315323173
#>  [33,] 0.201036808  0.506541348 0.272889454 0.307364289  0.485364022
#>  [34,] 0.888343953  3.121605800 1.481159004 5.240897763  4.885387133
#>  [35,] 0.800531863  0.554733718 0.975028765 1.990012485  1.105356500
#>  [36,] 0.434600959  1.593669096 0.571130390 1.096875509  1.323989925
#>  [37,] 0.389098237  0.099289075 0.142719928 0.104561796  0.184737995
#>  [38,] 0.301346454  0.756537349 0.628264106 0.755856345  0.931601161
#>  [39,] 0.120404837  0.411827028 0.238078800 0.172377883  0.322723703
#>  [40,] 0.351990010  0.162464774 0.125431940 0.175667104  0.239943102
#>  [41,] 0.326258662  2.070657342 0.475184242 0.726606635  4.815107342
#>  [42,] 0.387141898  3.156862341 0.836930372 1.700468394  6.899308893
#>  [43,] 0.466686504  0.988314444 0.593130462 0.325625698  2.933599863
#>  [44,] 0.552443645  0.957105684 0.675042363 0.570653505  2.074426294
#>  [45,] 0.395372466  0.074619278 0.266380635 0.278262269  0.053760259
#>  [46,] 0.266258674 22.325498214 0.658567414 1.437721716  3.105537147
#>  [47,] 0.004739542  0.004252493 0.006150061 0.005001025  0.005565753
#>  [48,] 0.799474480  4.737845852 1.556628639 3.175229967 15.899258482
#>  [49,] 0.520999929  0.957948959 0.791012400 0.747706733  0.632138564
#>  [50,] 0.764432046  1.063834381 0.550874889 0.623088782  2.655125719
#>  [51,] 0.051249851  0.079585756 0.068299222 0.042594584  0.047153357
#>  [52,] 0.933647374  3.132674878 1.743551515 1.190957338  0.700204690
#>  [53,] 0.249292969  0.188130897 0.170931372 0.223324908  0.365184494
#>  [54,] 0.067476086  0.837348444 0.284717991 0.062927424  0.535657460
#>  [55,] 0.887385175  0.131374806 1.618124591 0.361730082  0.053142456
#>  [56,] 0.901554824  8.313225688 1.141282465 3.437693106 22.905386552
#>  [57,] 0.947447927  0.921037442 0.882357610 0.427378962  0.388258543
#>  [58,] 0.586318457  0.605275132 0.548249501 0.637811752  1.496454774
#>  [59,] 0.051792088  0.032350145 0.053287975 0.049253230  0.031741289
#>  [60,] 0.710427318  7.497294577 1.455770477 5.778987730 21.122003182
#>  [61,] 0.588897671  1.650539263 1.044889234 0.995357663  2.785752022
#>  [62,] 0.454276308 11.623013074 1.205330821 1.335464268  4.145535838
#>  [63,] 0.654620668  1.421200316 0.659368340 0.757598903  4.791642543
#>  [64,] 0.463292295  0.148463453 0.432519644 0.590126224  0.135943579
#>  [65,] 0.013914367  0.283331923 0.036766657 0.086160056  0.243772683
#>  [66,] 0.695281362  3.344994444 1.023742705 1.225871638  5.372244788
#>  [67,] 0.680970686  0.504409179 0.457936188 0.760338476  0.600240242
#>  [68,] 0.553952067  1.748431603 1.508433042 0.357759153  1.544410285
#>  [69,] 0.611507542  1.520795166 0.761511727 0.904661775  1.390494702
#>  [70,] 0.396680596  0.196684081 0.270425920 0.158491558  0.182049996
#>  [71,] 0.850928390  2.491163515 0.684513296 1.989597374  7.615631603
#>  [72,] 0.858777136  2.371072379 1.324180126 1.531521679  0.783692354
#>  [73,] 0.924068696  1.075533962 0.516807472 1.254070513  0.610200604
#>  [74,] 0.015415540  0.007995151 0.017527205 0.071513786  0.022191840
#>  [75,] 0.342693782  3.611892919 0.749460162 2.110574933 10.403482862
#>  [76,] 0.755571619  1.830185009 1.479529423 0.747094501  0.381264685
#>  [77,] 0.475393753  0.392323395 0.275135241 0.406491049  0.605523642
#>  [78,] 0.025079663  0.041473020 0.036287355 0.034340344  0.049894821
#>  [79,] 0.187237736  0.109631454 0.165116115 0.161242520  0.131683624
#>  [80,] 0.706891337  0.295438206 0.363984306 0.462716540  0.334555744
#>  [81,] 0.075133131  0.048209374 0.046723808 0.024449737  0.061290604
#>  [82,] 0.978370367  1.390269367 1.880436332 0.645721313  0.922807081
#>  [83,] 0.537188561  0.351295553 0.375519122 0.526160240  0.846563141
#>  [84,] 0.047665276  0.153764483 0.126439098 0.145727417  0.265775416
#>  [85,] 0.886054994  0.063923425 0.322614837 0.209388467  0.157533770
#>  [86,] 0.531166163  0.887137624 1.540611793 2.508591688  0.866180832
#>  [87,] 0.759581176  0.669182223 1.395149600 0.889883266  0.495348910
#>  [88,] 0.071831204  0.320666160 0.152667546 0.126204967  0.232168811
#>  [89,] 0.761094006  7.342740405 1.279590904 0.587820458  7.826536544
#>  [90,] 0.698488922  4.542473557 1.250048438 1.540693862  3.952452307
#>  [91,] 0.659261779 34.686791334 1.590648490 7.597064647 24.498599441
#>  [92,] 0.983029161  0.972777576 0.736884338 1.049090356  1.642395846
#>  [93,] 0.036265538  0.035432414 0.026920787 0.052511884  0.101955828
#>  [94,] 0.288934063  1.736444259 0.782371025 2.210695992  9.424573016
#>  [95,] 0.959112237  2.610046887 1.667204813 2.501504824  3.705122297
#>  [96,] 0.240345604  0.320626484 0.263078625 0.059505259  0.106691570
#>  [97,] 0.994446137  7.614867576 1.673000680 6.238908494 23.672177402
#>  [98,] 0.079681585  0.086903648 0.087481284 0.067627948  0.028809674
#>  [99,] 0.834087128  9.609270547 1.265039855 8.625507067 26.298879660
#> [100,] 0.593907788  3.417011391 1.212710036 0.844776211  1.192542459