Simulation from X-Vine models

XVineSim() generates multivariate samples from X-Vine models given one of the variables is less than 1. The simulation algorithm selects the X-vine specification associated with the conditioned variable and uses recursive approaches to return an \(N \times d\) data matrix. For more details on simulation algorithms, refer to Kiriliouk, A., Lee, J., & Segers, J. (2023)

Usage

XVineSim(N, XVS, k)

Arguments

N

Integer; the number of samples to simulate.

XVS

A list of three matrix components for each conditioning variable:

reproduced structure matrices

family matrices

parameter matrices. To specify the argument XVS, see XVineSpec().

k

Integer; the \(k\)th, \(k\in {1,\ldots,d}\), conditioning index to condition on.

Value

An \(N\times d\) data matrix used to calculate extremal functions in ParetoSim().

References

Kiriliouk, A., Lee, J., & Segers, J. (2023). X-Vine Models for Multivariate Extremes. arXiv preprint arXiv:2312.15205.

Examples

##  A 5-dim X-vine model
StrMtx <- matrix(c(1, 1, 2, 2, 4,
                  0, 2, 1, 3, 2,
                  0, 0, 3, 1, 3,
                  0, 0, 0, 4, 1,
                  0, 0, 0, 0, 5),5,byrow = TRUE)
ParMtx <- matrix(c(0, 1.5, 2, 2.5, 2,
                   0, 0, 2, 2.5, 0.7,
                   0, 0, 0, 0.4, -0.3,
                   0, 0, 0, 0, 0.1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
FamMtx <- matrix(c(0, 1, 2, 3, 4,
                   0, 0, 3, 4, 1,
                   0, 0, 0, 3, 1,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0),5,byrow = TRUE)
##  X-Vine specification
XVS=XVineSpec(M = StrMtx, Mmod = FamMtx, Mpar = ParMtx)
##  Simulate from X-vine model when X_1 is less than 1.
XVineSim(N = 100, XVS = XVS, k = 1)
#>                [,1]         [,2]         [,3]         [,4]         [,5]
#>   [1,] 3.812795e-01 3.463828e+00 7.245930e-01 1.236626e+00 4.600108e+00
#>   [2,] 4.841271e-01 5.661966e+00 1.324210e+00 2.825996e+00 1.447557e+01
#>   [3,] 1.507756e-01 1.689485e+00 3.959396e-01 2.615363e-01 1.991233e+00
#>   [4,] 1.457990e-01 3.931829e-02 8.879508e-02 5.051165e-02 1.853415e-02
#>   [5,] 2.030867e-01 6.749885e+00 4.584993e-01 1.040717e+00 3.622614e+00
#>   [6,] 9.640652e-02 1.556948e-01 1.041895e-01 1.051846e-01 1.201132e-01
#>   [7,] 6.105853e-01 9.676404e+00 1.828047e+00 2.014300e+00 8.387636e+00
#>   [8,] 4.394550e-01 3.240305e+00 1.168491e+00 1.190056e+00 1.893608e+00
#>   [9,] 1.958126e-01 1.117370e-01 3.285735e-01 4.500074e-01 3.023840e-01
#>  [10,] 6.362581e-01 1.246823e+00 1.095887e+00 2.127027e+00 1.279339e+00
#>  [11,] 7.024254e-01 1.025728e+00 7.747789e-01 1.239441e+00 1.286231e+00
#>  [12,] 1.601753e-01 2.199739e+00 2.706545e-01 1.062811e+00 7.905504e+00
#>  [13,] 1.341145e-01 1.323764e+00 1.786702e-01 4.291945e-01 2.817141e+00
#>  [14,] 9.736768e-01 2.496505e+00 1.112656e+00 1.017795e+00 1.764464e+00
#>  [15,] 7.464616e-01 8.223944e-01 9.396934e-01 1.143910e+00 9.579540e-01
#>  [16,] 2.643491e-01 7.844720e-01 5.914461e-01 9.937703e-01 7.710502e-01
#>  [17,] 1.200943e-01 2.509252e-01 1.607031e-01 2.230337e-01 2.703364e-01
#>  [18,] 4.645512e-01 2.207273e-01 3.408672e-01 4.326773e-01 2.847413e-01
#>  [19,] 5.809018e-01 1.147199e+00 1.104275e+00 8.072741e-01 3.646179e-01
#>  [20,] 2.432190e-01 2.667678e-01 1.615687e-01 1.853409e-01 1.458760e-01
#>  [21,] 7.655927e-01 1.711430e+00 1.141874e+00 8.858552e-01 1.066214e+00
#>  [22,] 7.769781e-01 2.155598e+00 1.371820e+00 1.327467e+00 6.807417e-01
#>  [23,] 9.307347e-01 5.068581e-01 4.605381e-01 6.630144e-01 7.341453e-01
#>  [24,] 8.828291e-01 2.254414e+00 2.595902e+00 1.514965e+00 4.088239e+00
#>  [25,] 1.590578e-01 8.628374e-02 7.373753e-02 4.716869e-02 8.480539e-02
#>  [26,] 5.595851e-01 1.129088e+00 8.213218e-01 8.614734e-01 3.277107e+00
#>  [27,] 7.416368e-01 1.154468e+01 2.083357e+00 4.120520e-01 4.220332e+00
#>  [28,] 3.538569e-01 7.324287e-01 5.885410e-01 5.499558e-01 4.790253e-01
#>  [29,] 7.611129e-01 9.660052e-01 7.075016e-01 9.945798e-01 7.819312e+00
#>  [30,] 7.249422e-01 2.302803e-01 6.778524e+00 1.622397e+01 1.175805e+00
#>  [31,] 1.123310e-01 8.545161e-01 2.246622e-01 4.832887e-01 4.798539e-01
#>  [32,] 8.005051e-01 2.163144e+00 9.768173e-01 8.484171e-01 1.401933e+00
#>  [33,] 5.455400e-01 6.643034e-01 4.228672e-01 7.579568e-01 1.058527e+00
#>  [34,] 3.730622e-02 8.651941e-02 4.078991e-02 6.891714e-02 2.645176e-01
#>  [35,] 2.387145e-03 6.195657e-03 3.181048e-03 2.752926e-03 3.146512e-03
#>  [36,] 4.702233e-01 1.901136e-01 1.892694e-01 1.139419e-01 1.153088e-01
#>  [37,] 8.772867e-01 6.975180e-01 5.955770e-01 1.187113e+00 1.499920e+00
#>  [38,] 6.114469e-01 3.009416e+00 1.633743e+00 1.171785e+00 2.103373e+00
#>  [39,] 6.729626e-01 9.124595e-01 5.579844e-01 8.915893e-01 3.616278e-01
#>  [40,] 7.124934e-01 6.750094e-01 5.198001e-01 4.719587e-01 3.761430e+00
#>  [41,] 5.907805e-01 2.755703e+00 7.114179e-01 5.391647e-01 1.255305e+00
#>  [42,] 7.376564e-01 3.864045e+00 9.482551e-01 2.133165e+00 1.271361e+01
#>  [43,] 1.182511e-02 1.447960e-02 9.290825e-03 1.154980e-02 1.449923e-02
#>  [44,] 6.254409e-01 4.201630e+01 1.863732e+00 2.628995e+00 6.702504e+01
#>  [45,] 7.250368e-01 1.931124e+00 4.693417e-01 1.427298e+00 3.493052e+00
#>  [46,] 9.563482e-01 7.753288e+00 2.272151e+00 2.500022e+00 1.114299e+01
#>  [47,] 7.019033e-01 8.920184e-01 8.768972e-01 3.726833e+00 1.687729e+00
#>  [48,] 3.303150e-01 2.474305e+00 4.188942e-01 1.347830e+00 9.140827e+00
#>  [49,] 9.628367e-02 1.247845e-01 2.406149e-01 6.824659e-01 2.354196e-01
#>  [50,] 2.946856e-01 1.809336e+00 4.343932e-01 9.308515e-01 5.738720e+00
#>  [51,] 1.290298e-01 1.633829e-01 1.795643e-01 2.014472e-01 1.955912e-01
#>  [52,] 4.204740e-01 3.411793e-01 4.981228e-01 4.083216e-01 3.770695e-01
#>  [53,] 9.804396e-01 9.490378e-01 6.477839e-01 9.936106e-01 5.913878e-01
#>  [54,] 5.449985e-01 1.459285e-01 1.996607e-01 2.922525e-01 3.137908e-01
#>  [55,] 8.973099e-01 2.814354e+00 1.341022e+00 3.035015e+00 9.877321e+00
#>  [56,] 7.336473e-01 2.439146e+00 8.152245e-01 1.847508e+00 1.069642e+00
#>  [57,] 7.273740e-01 1.385827e+00 8.582463e-01 1.015944e+00 3.415557e-01
#>  [58,] 8.367754e-01 1.032272e+00 1.054852e+00 1.198777e+00 1.119622e+00
#>  [59,] 6.981086e-01 8.209032e+00 1.200446e+00 3.672967e+00 1.288954e+01
#>  [60,] 3.512984e-01 3.827600e-01 3.021127e-01 6.293116e-01 2.661403e-01
#>  [61,] 3.303900e-01 1.352265e+00 6.706874e-01 1.063207e+00 1.749665e+00
#>  [62,] 1.087319e-05 3.926006e-04 2.572793e-05 6.622784e-05 2.264063e-03
#>  [63,] 1.144538e-01 2.895095e-01 2.003286e-01 1.965440e-01 1.196602e+00
#>  [64,] 2.771464e-01 1.526730e+00 4.579813e-01 1.642689e+00 5.582117e-01
#>  [65,] 4.299043e-01 7.361917e-02 7.507332e-02 1.603435e-01 1.083507e-01
#>  [66,] 9.782891e-01 1.899606e+00 1.736318e+00 1.094448e+00 1.285100e+00
#>  [67,] 1.616360e-01 6.256796e-03 2.587783e-02 1.008073e-01 1.739143e-03
#>  [68,] 6.148587e-01 3.818539e+00 8.137386e-01 1.873511e+00 3.971179e+00
#>  [69,] 7.482703e-01 7.224086e+00 2.225647e+00 2.204689e+00 8.913937e+00
#>  [70,] 1.742793e-01 1.111494e-01 1.287304e-01 8.412939e-02 1.531196e-01
#>  [71,] 3.834225e-01 6.862518e-02 2.042341e-01 1.077222e-01 8.958672e-03
#>  [72,] 6.333837e-01 5.236176e-01 6.218749e-01 1.597524e-01 5.843397e-02
#>  [73,] 6.780073e-01 1.249825e+00 4.749413e-01 1.366832e+00 1.578131e+00
#>  [74,] 3.610049e-01 2.937000e+00 3.051553e-01 2.029285e+00 6.398488e+00
#>  [75,] 6.453225e-02 1.246633e-01 1.034624e-01 1.076435e-01 3.668857e-01
#>  [76,] 3.307222e-01 6.631948e+00 9.601850e-01 1.549957e+00 3.099511e+00
#>  [77,] 2.771883e-02 1.026403e-01 5.069140e-02 5.343802e-02 3.037544e-01
#>  [78,] 5.595222e-01 3.140472e-01 4.445395e-01 3.920632e-01 3.719874e-01
#>  [79,] 9.156636e-01 2.888134e+00 1.432699e+00 1.743758e+00 4.325033e+00
#>  [80,] 4.646573e-01 9.151668e-01 3.828651e-01 1.168072e+00 2.036780e+00
#>  [81,] 1.519615e-01 4.600915e-02 8.666679e-02 7.967529e-02 7.140875e-02
#>  [82,] 5.087515e-01 6.534000e-01 7.485068e-01 4.095745e-01 3.925055e+00
#>  [83,] 6.878062e-01 2.363418e+00 1.403120e+00 1.586217e+00 7.662843e+00
#>  [84,] 9.507496e-01 2.067873e+00 5.378563e+00 7.829989e-01 4.412196e+00
#>  [85,] 8.997425e-01 6.807735e+00 1.609009e+00 1.465651e+00 3.992150e+00
#>  [86,] 4.772772e-01 1.035808e+01 6.616727e-01 5.585669e+00 1.256270e+02
#>  [87,] 4.683716e-01 1.987361e-01 2.870204e-01 1.903144e-01 2.783056e-01
#>  [88,] 6.269295e-01 1.011881e+00 2.185723e-01 6.853201e-01 2.779682e+00
#>  [89,] 8.241438e-01 3.345187e+00 1.016902e+00 1.311852e+00 4.136092e+00
#>  [90,] 2.681968e-01 1.304037e+00 4.047504e-01 1.342113e+00 2.374557e+00
#>  [91,] 4.646048e-01 1.509680e+00 1.080317e+00 2.893702e-01 1.603906e-01
#>  [92,] 3.202654e-01 3.622168e-01 7.781119e-01 4.469964e-01 7.478822e-01
#>  [93,] 3.912385e-01 5.102267e-01 4.955986e-01 2.987344e-01 3.629455e-01
#>  [94,] 6.873063e-01 4.417374e-01 5.159208e-01 9.400593e-01 2.052978e-01
#>  [95,] 5.033228e-01 3.846994e-01 1.520432e+00 1.363153e+00 1.948037e-01
#>  [96,] 3.255916e-01 9.402826e-01 1.059400e+00 6.423385e-01 9.285965e-01
#>  [97,] 6.176547e-01 3.005255e+00 1.177592e+00 9.170180e-01 5.473195e+00
#>  [98,] 6.043822e-01 8.950080e-01 6.532677e-01 1.009703e+00 7.141152e-01
#>  [99,] 1.114656e-02 4.604022e-03 6.213047e-03 5.645136e-03 6.236923e-03
#> [100,] 4.657457e-01 8.484610e-01 5.523621e-01 4.741693e-01 1.250180e+00